% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansur2.R
\docType{data}
\name{ansur2.male}
\alias{ansur2.male}
\title{\code{ansur2.male}}
\source{
\url{https://www.openlab.psu.edu/ansur2/}
}
\description{
The Anthropometric Survey of US Army Personnel (ANSUR 2 or ANSUR II) data were published internally in 2012. 
They were made available publicly in 2017. They have replaced ANSUR I as the most comprehensive publicly available 
data set on body size and shape. They include 93 measures for over 6,000 adult US military personnel 
(4,082 men and 1,986 women). In contrast to the ANSUR I data, the new sample includes reservists. 
Despite the presence of reservists in the sample, it is still not an approximation of the US Civilian population. 
Consequently, while there is useful information here, designs and standards based on these data will not accommodate 
most user populations in the intended manner.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/ansur2}
}
\references{
C.C. Gordon, C.L. Blackwell, B. Bradtmiller, J.L. Parham, P. Barrientos, S.P. Paquette, B.D. Corner, J.M. Carson,
 J.C. Venezia, B.M. Rockwell, M. Mucher and S. Kristensen (Dec 2014), 
 \href{http://tools.openlab.psu.edu/publicData/ANSURII-TR15-007.pdf}{2012 ANTHROPOMETRIC SURVEY OF U.S. ARMY PERSONNEL: METHODS AND SUMMARY STATISTICS}, Technical Report NATICK/TR-15/007, 
 U.S. Army Natick Soldier Research, Development and Engineering Center Natick, Massachusetts 01760-2642.
}
