% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{concepts}
\alias{concepts}
\title{concepts}
\usage{
concepts(pkgs = "wwwdata")
}
\arguments{
\item{pkgs}{text: names of one or more packages (default: \code{"wwwdata"})}
}
\value{
text vector with sorted concepts and keywords
}
\description{
Extracts all concepts and keywords from a one or more package. As default only the concepts and keywords from the package \code{wwwdata} returned. 
If you use \code{concepts(NULL)} the concepts and keywords from all installed(!) packages are returned.
}
\examples{
concepts()
concepts('stats')
\dontrun{concepts(NULL)}
}
