% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.a_rod_2016}
\alias{dasl.a_rod_2016}
\title{A-Rod 2016}
\format{23 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/a-rod-2016/?sf_paged=1}{A-Rod 2016}
}
\description{
Alex Rodriguez (known to fans as A-Rod)was the youngest player ever to hit 500 home runs. The file holds the number of home runs hit by A-Rod during the 1994–2016 seasons. Describe the distribution, mentioning its shape and any unusual features.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.baseball-reference.com/players/r/rodrial01.shtml}
}
\concept{Data Display}
