% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.accidents}
\alias{dasl.accidents}
\title{Accidents}
\format{6 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/accidents/?sf_paged=1}{Accidents}
}
\description{
Progressive Insurance asked customers who had been involved in auto accidents how far they were from home when the accident happened.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Progressive Insurance Newsletter out of print
}
