% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.acela}
\alias{dasl.acela}
\title{Acela}
\format{16000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/acela/?sf_paged=1}{Acela}
}
\description{
Acela
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented -- based on PVA data
}
\concept{Analysis of Variance}
