% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.acid_rain}
\alias{dasl.acid_rain}
\title{Acid rain}
\format{26 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/acid-rain/?sf_paged=1}{Acid rain}
}
\description{
Froliger and Kane measured the pH (a scale on which a value of 7 is neutral and values below 7 are acidic) of water collected from precipitation events in Allegheny County, Pennsylvania between December 20, 1973 and May 23, 1974. Display the distribution of these values and describe with words and numbers what you see.  Which of the ways of displaying the distribution seems to work best? The data file includes the dates of the events.  Do these dates offer any useful additional information about the largest (least acidic) values? (Hint: you may need to consult a calendar for 1973-4.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Reference: (J.O.Froliger and R. Kane, "Precipitation: Its Acidic Nature," Science 189 (8 August 1975): 455-457).
}
\concept{Data Display}
