% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.activating_baking_yeast}
\alias{dasl.activating_baking_yeast}
\title{Activating baking yeast}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/activating-baking-yeast/?sf_paged=1}{Activating baking yeast}
}
\description{
To shorten the time it takes him to make his favorite pizza, a student designed an experiment to test the effect of sugar and milk on the activation times for baking yeast. Specifically, he tested four different recipes and measured how many seconds it took for the same amount of dough to rise to the top of a bowl. He randomized the order of the recipes and replicated each treatment 4 times.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
student project
}
\concept{Analysis of Variance}
