% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.adoptions}
\alias{dasl.adoptions}
\title{Adoptions}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/adoptions/?sf_paged=1}{Adoptions}
}
\description{
The United States Census Bureau keeps track of the number of adoptions in each State (and Washington D.C.). The data includes the population of each state as well. How should adoptions be summarized and displayed?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.census.gov
}
\concept{Data Display}
