% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.advertising}
\alias{dasl.advertising}
\title{Advertising}
\format{500 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/advertising/?sf_paged=1}{Advertising}
}
\description{
Advertising
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Sampled from ages in PVA data.
}
\concept{Confidence Intervals for Proportions}
