% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.age_and_party_2016}
\alias{dasl.age_and_party_2016}
\title{Age and party 2016}
\format{4 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/age-and-party/?sf_paged=1}{Age and party 2016}
}
\description{
The Pew Research Center conducted a representative telephone survey in October of 2016. Among the reported results was the following table concerning the preferred political party affiliation of respondents and their ages for white voters. Is there evidence of age-based differences in party affiliation in the United States for white voters?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.people-press.org/2016/09/13/1-the-changing-composition-of-the-political-parties/}
}
