% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.aig_daily}
\alias{dasl.aig_daily}
\title{AIG daily}
\format{2810 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/aig-daily/?sf_paged=1}{AIG daily}
}
\description{
The American International Group (AIG) was once the 18th  largest corporation in the world. By early 2007 AIG had assets of $1 trillion, $110 billion in revenues, 74 million customers and 116,000 employees in 130 countries and jurisdictions. Yet just 18 months later, AIG found itself on the brink of failure and in need of emergency government assistance.Between 2007 and 2009 AIG stock lost more than 99% of its value, hitting $0.35 in early March. Could the crash have been predicted?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
