% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.airline_bumping_2017}
\alias{dasl.airline_bumping_2017}
\title{Airline bumping 2017}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/airline-bumping-2017/?sf_paged=1}{Airline bumping 2017}
}
\description{
Commercial airlines overbook flights, selling more tickets than they have seats, because a sizeable number of reservation holders don’t show up in time for their flights. But sometimes, there are more passengers wishing to board than there are seats. Most airlines try to entice travelers to voluntarily give up their seats in return for free travel or other awards, but they do have to “bump” some travelers involuntarily. Of course, they don’t like to offend passengers by bumping, so they are constantly trying to improve their systems for predicting how many passengers will show up. Have the rates of “bumping” changed? Here are data on the number of passengers involuntarily denied boarding (“bumping” is not the approved term) per 10,000 passengers during the periods of January to September in 2016 and 2017 by airline.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.transportation.gov/sites/dot.gov/files/docs/resources/individuals/aviation-consumer-protection/300341/2017-december-atcr.pdf}
}
\concept{Comparing Two Groups}
