% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.airport_screening}
\alias{dasl.airport_screening}
\title{Airport screening}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/airport-screening/?sf_paged=1}{Airport screening}
}
\description{
According to the U.S. Dept of Transportation, screening consists of “the systematic examination of persons and property using weapons-detecting procedures or facilities (electronic or physical search) for the purpose of detecting weapons and dangerous articles and to prevent their unauthorized introduction into sterile areas or aboard aircraft.” The data report the results of airport screening un the U.S. from 1977 to 2000.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.albany.edu/sourcebook/csv/t3177.csv}  adapted from: U.S. Department of Transportation, Federal Aviation Administration, Semiannual Report to Congress on the Effectiveness of the Civil Aviation Security Program, July 1 to "December 31, 1978, Exhibit 10; July 1 to December 31, 1982, Exhibit 10; July 1 to December 31, 1984, Exhibit 7; July 1 to December 31, 1989, p. 11 (Washington, DC: U.S. Department of Transportation); U.S. Department of Transportation, Federal Aviation Administration, Annual Report to Congress on Civil Aviation Security, January 1, 1993-December 31, 1993, p. 9; January 1, 1995-December 31, 1995, p. 11 (Washington, DC: U.S. Department of Transportation); and data provided by the U.S.  Department of Transportation, Federal Aviation Administration and Bureau of Transportation Statistics [Online]. Available:  \url{http://www.bts.gov/publications/national_transportation_statistics/}  2003/html/table_02_16.html [May 24, 2004]. Table adapted by SOURCEBOOK staff.
}
\concept{Multiple Regression}
