% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.all_the_efficiency}
\alias{dasl.all_the_efficiency}
\title{All the efficiency}
\format{75 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/all-the-efficiency/?sf_paged=2}{All the efficiency}
}
\description{
A sample of model 2011 cars from an online information service colleted to see how fuel efficiency (as highway mpg) relates to the cost (MSRP)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
