% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.alternative_energy_2016}
\alias{dasl.alternative_energy_2016}
\title{Alternative energy 2016}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/alternative-energy-2016/?sf_paged=2}{Alternative energy 2016}
}
\description{
\url{https://en.wikipedia.org/wiki/List_of_U.S._states_by_electricity_production_from_renewable_sources}
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_U.S._states_by_electricity_production_from_renewable_sources}
}
\concept{Regression}
