% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.amazon_books}
\alias{dasl.amazon_books}
\title{Amazon books}
\format{325 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/amazon-books/?sf_paged=2}{Amazon books}
}
\description{
Amazon books
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data harvested from Amazon.com
}
\concept{Correlation}
\concept{Multiple Regression}
\concept{Regression}
\concept{Scatterplot}
