% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.analgesics}
\alias{dasl.analgesics}
\title{Analgesics}
\format{27 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/analgesics/?sf_paged=2}{Analgesics}
}
\description{
A pharmaceutical company tested three formulations of a pain relief medicine for migraine headache sufferers. For the experiment, 27 volunteers were selected and 9 were randomly assigned to one of three drug formulations. The subjects were instructed to take the drug during their next migraine headache episode and to report their pain on a scale of 1 = no pain to 10 = extreme pain 30 minutes after taking the drug.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Analysis of Variance}
