% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.apples_and_gas}
\alias{dasl.apples_and_gas}
\title{Apples and gas}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/apples-and-gas/?sf_paged=2}{Apples and gas}
}
\description{
The price of delicious apples and regular gas are components of the Consumer Price Index. The data give those prices monthly for the year 2006
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\concept{Time Series Analysis}
