% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.archery}
\alias{dasl.archery}
\title{Archery}
\format{64 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/archery/?sf_paged=2}{Archery}
}
\description{
In Olympic Archery both men and women start with a field of 64 qualifiers. Each archer shoots a round of 72 arrows (total possible score: 720) to establish a seeding position. Then they participate in a single-elimination contest. Thus, the seeding round is the only one that provides data for all archers (because some are eliminated at each step of the elimination rounds). The data are the seeding round data for the 2008 Olympics
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.espn.com/oly/summer08/results?eventId=258}  (men) \url{http://www.espn.com/oly/summer08/results?eventId=211}   (women)
}
