% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.attendance_2016}
\alias{dasl.attendance_2016}
\title{Attendance 2016}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/attendance-2016/?sf_paged=2}{Attendance 2016}
}
\description{
American League baseball games are played under the designated hitter rule, meaning that pitchers, often weak hitters, do not come to bat. Baseball owners believe that the designated hitter rule means more runs scored, which in turn means higher attendance. Is there evidence that more fans attend games if the teams score more runs? The data give attendance total, home, and away for each team in major league baseball for the 2016 season. Does attendance respond to winning or runs scored?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Attendance from \url{http://www.espn.com/mlb/attendance}   Won and Lost totals from \url{http://www.baseball-reference.com/leagues/MLB/2016.shtml}
}
\concept{Comparing Two Groups}
\concept{Correlation}
\concept{Multiple Regression}
\concept{Regression}
\concept{Scatterplot}
\concept{Test for Difference between Two Groups}
