% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.auto_noise_filters}
\alias{dasl.auto_noise_filters}
\title{Auto noise filters}
\format{35 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/auto-noise-filters/?sf_paged=2}{Auto noise filters}
}
\description{
In a statement to a Senate Public Works Committee, a senior executive of Texaco, Inc., cited a study on the effectiveness of auto filters on reducing noise. Because of concerns about performance, two types of filters were studied, a standard silencer and a new device developed by the Associated Octel Company. Noise is in decibels/10. Type 1 = standard, Type 2 = Octel.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Analysis of Variance}
