% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.babysamp_98}
\alias{dasl.babysamp_98}
\title{Babysamp 98}
\format{200 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/babysamp-98/?sf_paged=2}{Babysamp 98}
}
\description{
1998 Baby data from  \url{http://www.nber.org/natality/ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/DVS/natality/}
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.nber.org/natality/ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/DVS/natality/}  Sample of 200 using set.seed(1) in R
}
\concept{Confidence Intervals for Proportions}
