% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ballplayer_births}
\alias{dasl.ballplayer_births}
\title{Ballplayer births}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ballplayer-births/?sf_paged=2}{Ballplayer births}
}
\description{
It has been suggested that children born in the summer have an advantage over their peers when it comes to sports, perhaps because they can be outdoors when they are young. The data report the number of professional ballplayers born in each month of the year for one season of professional baseball.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
