% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.barbershop_music}
\alias{dasl.barbershop_music}
\title{Barbershop music}
\format{34 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/barbershop-music/?sf_paged=2}{Barbershop music}
}
\description{
At a barbershop music singing competition, choruses are judged on three scales: Music (quality of the arrangement, etc.), Performance, and Singing. The scales are supposed to be independent of each other, and each is scored by a different judge, but a friend claims that he can predict a chorus’s singing score from the other two scores. Are the scores really independent?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Results of a barbershop competition as reported by Velleman
}
