% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.baseball_2016}
\alias{dasl.baseball_2016}
\title{Baseball 2016}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/baseball-2016/?sf_paged=3}{Baseball 2016}
}
\description{
Summaries of the performance of each of the major league baseball teams in 2016. Variables include: Games won, Games Lost, Runs peer game, At bats, Runs, Hits, Doubles, Triples, Home runs, Runs batted in, Bases stolen, Time caught stealing, Bases on Balls, Strikeouts, Hits/At Bats, On Base Percentage, Slugging percentage, On base+Slugging, Total bases, Double plays grounded into, Times hit by pitch, Sacrifice hits, Sacrifice flies, Intentional base on balls, and Runners Left On Base
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.baseball-reference.com/leagues/MLB/2016.shtml}
}
\concept{Boxplots}
\concept{Test for Difference between Two Groups}
