% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.baseball_attendance}
\alias{dasl.baseball_attendance}
\title{Baseball attendance}
\format{80 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/baseball-attendance/?sf_paged=3}{Baseball attendance}
}
\description{
Baseball attendance
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.baseball-reference.com}
}
\concept{Multiple Regression}
