% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.baseball_circumferences}
\alias{dasl.baseball_circumferences}
\title{Baseball circumferences}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/baseball-circumferences/?sf_paged=3}{Baseball circumferences}
}
\description{
Baseball circumferences
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented based on manufacturing data -- private communication RDD
}
\concept{Quality Control}
