% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.baseball_salaries_2015}
\alias{dasl.baseball_salaries_2015}
\title{Baseball salaries 2015}
\format{132 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/baseball-salaries-2015/?sf_paged=3}{Baseball salaries 2015}
}
\description{
Ballplayers have been signing ever larger contracts. The highest salaries (in millions of dollars per season) for each year since 1874 are in the data file.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.baseball-reference.com/leaders/Salary_leagues.shtml}
}
\concept{Nonparametric Methods}
\concept{Regression}
