% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.baseball_salaries_2016}
\alias{dasl.baseball_salaries_2016}
\title{Baseball salaries 2016}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/baseball-salaries-2016/?sf_paged=3}{Baseball salaries 2016}
}
\description{
Baseball salaries 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.surepayroll.com/resources/blog/mlb-team-payrolls}  (salaries) \url{http://www.mlb.com/mlb/standings/#20160904}   (wins)
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
