% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.baseball_weights}
\alias{dasl.baseball_weights}
\title{Baseball weights}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/baseball-weights/?sf_paged=3}{Baseball weights}
}
\description{
Baseball weights
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented based on manufacturing data -- private communication RDD
}
\concept{Quality Control}
