% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.basketball_shots}
\alias{dasl.basketball_shots}
\title{Basketball shots}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/basketball-shots/?sf_paged=3}{Basketball shots}
}
\description{
Basketball shots
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Student experiment from De Veaux
}
\concept{Multifactor Analysis of Variance}
