% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.basketballs}
\alias{dasl.basketballs}
\title{Basketballs}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/basketballs/?sf_paged=3}{Basketballs}
}
\description{
A company that makes basketballs has the motto: “Our basketballs are ready to play.” Therefore, it is important to the company that the basketballs are inflated with the proper amount of air when shipped. Most basketballs are inflated to 7 to 9 pounds per square inch. Recently the company selected a random basketball from its production line at four different time periods over its normal production day. The results were recorded.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented data based on Basketball production
}
\concept{Quality Control}
