% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.batteries}
\alias{dasl.batteries}
\title{Batteries}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/batteries/?sf_paged=3}{Batteries}
}
\description{
A student experiment was run to test the performance of 4 brands of batteries under 2 different Environments (room temperature and cold). For each of the 8 treatments, 2 batteries of a particular brand were put into a flashlight. The flashlight was then turned on and allowed to run until the light went out. The number of minutes the flashlight stayed on was recorded. Each treatment condition was run twice.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
student experiment
}
\concept{Multifactor Analysis of Variance}
