% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.being_successful}
\alias{dasl.being_successful}
\title{Being successful}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/being-successful/?sf_paged=3}{Being successful}
}
\description{
In a random sample of U.S. adults surveyed in December 2011, Pew Research asked how important it is “to you personally” to be successful in a high-paying career or profession. Responses are recorded by sex and age.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.pewsocialtrends.org/files/2012/04 /Women-in-the-Workplace.pdf
}
\concept{Quality Control}
\concept{Tables}
