% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.belmont_stakes_2015}
\alias{dasl.belmont_stakes_2015}
\title{Belmont stakes 2015}
\format{147 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/belmont-stakes-2015/?sf_paged=3}{Belmont stakes 2015}
}
\description{
The Belmont Stakes is the last and longest of the three horse races that make up the Triple Crown. Curiously, in some of the Belmont races horses have run clockwise around the track, and in others they have run counterclockwise. Do the horses care? But note that the length of the race has also not been consistent. In fact, there have been five different lengths from 1.125 miles to 1.625 miles.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.belmontstakes.com/past-winners.aspx}
}
