% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.bike_safety_2015}
\alias{dasl.bike_safety_2015}
\title{Bike safety 2015}
\format{22 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/bike-safety-2015/?sf_paged=4}{Bike safety 2015}
}
\description{
The Bicycle Helmet Safety Institute website includes a report on the number of bicycle fatalities per year in the United States. The data gives the counts for the years 1994–2015.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.bhsi.org/stats.htm}
}
