% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.bird_species_2013}
\alias{dasl.bird_species_2013}
\title{Bird Species 2013}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/bird-species-2013/?sf_paged=4}{Bird Species 2013}
}
\description{
Bird Species 2013
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.birdsource.org
}
\concept{Data Display}
