% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.birth_order}
\alias{dasl.birth_order}
\title{Birth order}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/birth-order/?sf_paged=4}{Birth order}
}
\description{
A statistics professor at a large university polled his students to find out what their majors were and what position they held in the family birth order. The results are summarized in the table.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Class poll by Velleman
}
