% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.birthrates_2015}
\alias{dasl.birthrates_2015}
\title{Birthrates 2015}
\format{11 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/birthrates-2015/?sf_paged=4}{Birthrates 2015}
}
\description{
Births per 1000 population in the United States, starting in 1965. There has been concern that the birthrate may be declining. A good model for tends in birthrate may allow for some prediction.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
cdc.gov/nchs/
}
\concept{Regression}
