% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.blood_pressure}
\alias{dasl.blood_pressure}
\title{Blood Pressure}
\format{474 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/blood-pressure/?sf_paged=4}{Blood Pressure}
}
\description{
Does blood pressure, on average, change with age. The data here are two categorical variables: Blood pressure categorized as High, Normal, Low, and Age categorized as under 30, 30-49, and over 50.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Realistic data
}
\concept{Tables}
