% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.bls_output}
\alias{dasl.bls_output}
\title{BLS output}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/bls-output/?sf_paged=4}{BLS output}
}
\description{
BLS output
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Bureau of Labor Statistics
}
\concept{Time Series Analysis}
