% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.bodyfat}
\alias{dasl.bodyfat}
\title{Bodyfat}
\format{250 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/bodyfat/?sf_paged=4}{Bodyfat}
}
\description{
Measurements of 250 men of various ages. The percent of a man’s body that is fat is a matter of concern for health and fitness. But the %bodyfat is difficult and expensive to measure accurately. These data offer correct %bodyfat measurements along with a variety of easier to find measures. Can you build a model to predict the %bodyfat from other, more easily made, measurements?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
BYU Human Performance Research Center \url{http://www.byu.edu/chhp/intro.html#lrc}  Director: Mark Ricard  116A RB, (801) 378-8958
}
\concept{Confidence Intervals for Means}
\concept{Confidence Intervals for Proportions}
\concept{Data Display}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Multiple Regression}
\concept{Normal Probability Plots}
\concept{Regression}
\concept{Tables}
