% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.bond_funds}
\alias{dasl.bond_funds}
\title{Bond funds}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/bond-funds/?sf_paged=4}{Bond funds}
}
\description{
You have decided to invest in a bond fund and plan to limit your choice of funds to Morningstar “medalist” funds. But now you must choose between a taxable fund and a municipal bond fund that is at least partially tax-free. Which is better? Here are the % returns for the three-year period leading up to spring of 2013.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.morningstar.com
}
\concept{Comparing Two Groups}
