% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.bookstore_sales}
\alias{dasl.bookstore_sales}
\title{Bookstore sales}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/bookstore-sales/?sf_paged=4}{Bookstore sales}
}
\description{
Number of sales people working in a bookstore and sales (in $1000) that day. These are realistic but invented data.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented example
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
