% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.boomtowns_2015}
\alias{dasl.boomtowns_2015}
\title{Boomtowns 2015}
\format{70 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/boomtowns-2015/?sf_paged=4}{Boomtowns 2015}
}
\description{
In 2015, the website NewGeography.com listed its ranking of the best cities for job growth in the United States. Nonfarm employment is also provided.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.newgeography.com/content/004941-large-cities-rankings-2015-best-cities-job-growth}
}
\concept{Data Display}
