% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.boyle}
\alias{dasl.boyle}
\title{Boyle}
\format{25 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/boyle/?sf_paged=4}{Boyle}
}
\description{
Scientist Robert Boyle examined the relationship between the volume in which a gas is contained and the pressure in its container. He used a cylindrical container with a moveable top that could be raised or lowered to change the volume. He measured the Height in inches by counting equally spaced marks on the cylinder, and measured the Pressure in inches of mercury (as in a barometer). The relationship between volume (height) and pressure is now known as Boyle’s Law. Find the re-expression that makes the relationship straight, and you’ll re-discover Boyle’s Law.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://jap.physiology.org/content/87/4/1543}
}
\concept{Regression}
