% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.brain_waves}
\alias{dasl.brain_waves}
\title{Brain waves}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/brain-waves/?sf_paged=5}{Brain waves}
}
\description{
An experiment was performed to see whether sensory deprivation over an extended period of time has any effect on the alpha-wave patterns produced by the brain. To determine this, 20 subjects, inmates in a Canadian prison, were randomly split into two groups. Members of one group were placed in solitary confinement. Those in the other group were allowed to remain in their own cells. Seven days later, alpha-wave frequencies were measured for all subjects.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
P. Gendreau et al., “Changes in EEG Alpha Frequency and Evoked Response Latency During Solitary Confinement,” Journal of Abnormal Psychology 79 [1972]: 54–59)
}
\concept{Analysis of Variance}
\concept{Boxplots}
\concept{Comparing Two Groups}
