% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.brakes}
\alias{dasl.brakes}
\title{Brakes}
\format{15 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/brakes/?sf_paged=5}{Brakes}
}
\description{
Stopping distances in feet for a car tested 3 times at each of 5 speeds. We hope to create a model that predicts Stopping Distance from the Speed of the car.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
