% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.broadway_shows}
\alias{dasl.broadway_shows}
\title{Broadway shows}
\format{33 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/broadway-shows/?sf_paged=5}{Broadway shows}
}
\description{
In 2016 13.27 million people attended a Broadway show, paying an average of more than $100 per ticket. The Broadway League, Inc. (\url{https://www.broadwayleague.com/research/statistics-broadway-nyc/)}  provides some historical and current data. These variables are available for each year since the 1984-85 season: Season  (The initial year of the season, so the 1984-85 season is 1984.) Gross ($M) Attendance (M) Note: before 2009 this is Paid Attendance. Beginning 2009 it is Attendance.) Playing weeks (Total weeks during each show performed summed over all shows; the best measure of Broadway’s overall activity.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.broadwayleague.com/research/statistics-broadway-nyc/}
}
\concept{Multiple Regression}
