% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.burger_king_items}
\alias{dasl.burger_king_items}
\title{Burger King items}
\format{122 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/burger-king-items/?sf_paged=5}{Burger King items}
}
\description{
Burger King publishes full nutrition information on its menu. These data are for the foods on the menu recently. (Visit the site listed as the reference for the most current list.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.bk.com/pdfs/nutrition.pdf}
}
\concept{Multiple Regression}
\concept{Regression}
