% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.burgers}
\alias{dasl.burgers}
\title{Burgers}
\format{7 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/burgers/?sf_paged=5}{Burgers}
}
\description{
Fast food is often considered unhealthy because much of it is high in both fat and sodium. But are the two related? The data give the fat and sodium contents of several brands of burgers.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
