% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.buy_from_a_friend}
\alias{dasl.buy_from_a_friend}
\title{Buy from a friend}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/buy-from-a-friend/?sf_paged=5}{Buy from a friend}
}
\description{
A researcher at Cornell University wanted to know how friendship might affect simple sales such as this. She randomly divided subjects into two groups and gave each group descriptions of items they might want to buy. One group was told to imagine buying from a friend whom they expected to see again. The other group was told to imagine buying from a stranger. The data are the prices offered by the experiment participants.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
J. J. Halpern, “The Transaction Index: A Method for Standardizing Comparisons of Transaction  Characteristics Across Different Contexts,” Group Decision and Negotiation, 6: 557–572, 1997
}
\concept{Test for Difference between Two Groups}
