% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ca_house_prices}
\alias{dasl.ca_house_prices}
\title{CA House Prices}
\format{28 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ca-house-prices/?sf_paged=5}{CA House Prices}
}
\description{
Home prices in two neighborhoods near San Francisco. Palo Alto is an older neighborhood and Foster City, a newer one. How do prices compare?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Prices sampled from Zillow.com in August 2017.
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
