% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.camp_sites}
\alias{dasl.camp_sites}
\title{Camp sites}
\format{46 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/camp-sites/?sf_paged=5}{Camp sites}
}
\description{
The dataset is the number of camp sites at each of the public parks in Vermont
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data from Vermont tourist brochure
}
\concept{Boxplots}
