% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.candy_bars}
\alias{dasl.candy_bars}
\title{Candy bars}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/candy-bars/?sf_paged=5}{Candy bars}
}
\description{
The dataset holds facts about candy bars read from their nutrition labels. The data are a good example for multiple regression (e.g. what contributes to the calories of a candy bar?). For such an analysis, the indicator variable for nuts appears to work well. Note that 5 sugar-free candy bars are marked as NA in the sugar variable to remove them from analyses using sugar. The data have been standardized on a per serving basis to equalize for different serving sizes.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
data collected by a student reading nutrition labels
}
\concept{Multiple Regression}
