% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cape10_2017}
\alias{dasl.cape10_2017}
\title{CAPE10 2017}
\format{1641 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cape10-2017/?sf_paged=5}{CAPE10 2017}
}
\description{
CAPE10 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.quandl.com/data/MULTPL/SHILLER_PE_RATIO_MONTH-Shiller-PE-Ratio-by-Month}
}
\concept{Normal model}
