% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.car_discounts}
\alias{dasl.car_discounts}
\title{Car discounts}
\format{100 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/car-discounts/?sf_paged=5}{Car discounts}
}
\description{
Car discounts
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
