% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.car_origins}
\alias{dasl.car_origins}
\title{Car origins}
\format{359 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/car-origins/?sf_paged=5}{Car origins}
}
\description{
Car origins
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
