% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.car_rental}
\alias{dasl.car_rental}
\title{Car rental}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/car-rental/?sf_paged=6}{Car rental}
}
\description{
Car rental
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented
}
\concept{Nonparametric Methods}
