% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.car_speeds}
\alias{dasl.car_speeds}
\title{Car speeds}
\format{500 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/car-speeds/?sf_paged=6}{Car speeds}
}
\description{
Car speeds
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Website no longer active Collected by John Beale, engineering student at Stanford
}
\concept{Boxplots}
\concept{Test for Difference between Two Groups}
