% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.car_speeds_100}
\alias{dasl.car_speeds_100}
\title{Car speeds 100}
\format{100 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/car-speeds-100/?sf_paged=6}{Car speeds 100}
}
\description{
John Beale of Stanford, California, recorded the speeds of cars driving past his house, where the speed limit read 20 mph (He actually recorded every car for a two-month period. These are 100 representative readings.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Website no longer active Collected by John Beale, engineering student at Stanford
}
\concept{Normal Probability Plots}
