% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.carbon_footprint}
\alias{dasl.carbon_footprint}
\title{Carbon footprint}
\format{82 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/carbon-footprint/?sf_paged=6}{Carbon footprint}
}
\description{
Carbon footprint
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.fueleconomy.gov  (data are no longer organized in this way)
}
\concept{Nonparametric Methods}
