% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.carbon_footprint_2015}
\alias{dasl.carbon_footprint_2015}
\title{Carbon footprint 2015}
\format{69 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/carbon-footprint-2015/?sf_paged=6}{Carbon footprint 2015}
}
\description{
Carbon footprint 2015
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.fueleconomy.gov  (data are no longer organized in this way)
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
