% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cars}
\alias{dasl.cars}
\title{Cars}
\format{38 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cars/?sf_paged=6}{Cars}
}
\description{
Measurements on 38 1978-79 model automobiles. Gas mileage in miles per gallon as measured by Consumers’ Union on a test track.  Other values as reported by automobile manufacturer. Used to illustrate regression model building and diagnosis. Be sure to check the residuals when predicting MPG.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Henderson and Velleman. (1981). "Building Regression Models Interactively". Biometrics 37,  400. Data originally collected from Consumer Reports.
}
\concept{Analysis of Variance}
\concept{Experiment Design}
\concept{Regression}
