% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.case_shiller_by_city}
\alias{dasl.case_shiller_by_city}
\title{Case-Shiller by city}
\format{311 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/case-shiller-by-city/?sf_paged=6}{Case-Shiller by city}
}
\description{
The S&P/Case-Shiller Home Price Indices track changes in the value of residential real estate nationally and in 20 metropolitan regions. (Some of these indices are actually traded on the Chicago Mercantile Exchange.) The data set Case-Shiller by City gives the monthly index values for each of the 20 cities tracked by the Case-Shiller index and two national composite series.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
