% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.case_study_i_pv}
\alias{dasl.case_study_i_pv}
\title{Case study I-PV}
\format{7600 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/case-study-i-pv/?sf_paged=6}{Case study I-PV}
}
\description{
Case study I-PV
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
sample from \url{https://kdd.ics.uci.edu/databases/kddcup98/kddcup98.html}
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
