% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ceo_compensation_2014}
\alias{dasl.ceo_compensation_2014}
\title{CEO Compensation 2014}
\format{434 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ceo-compensation-2014/?sf_paged=6}{CEO Compensation 2014}
}
\description{
Beginning in 2017, public companies will be required to disclose the ratio of CEO pay to median worker pay. The Glassdoor Economic Research Blog has published the data for 2014. The data includes CEO identities, companies, CEO compensation, median worker compensation (compiled by Glassdoor), and the ratio of CEO to worker compensation.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.glassdoor.com/research/ceo-pay-ratio/}
}
\concept{Boxplots}
\concept{Confidence Intervals for Means}
\concept{Displaying Quantitative Data}
\concept{Normal Probability Plots}
\concept{Summarizing Quantitative Data}
