% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ceo_salary_2012}
\alias{dasl.ceo_salary_2012}
\title{CEO Salary 2012}
\format{500 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ceo-salary-2012/?sf_paged=6}{CEO Salary 2012}
}
\description{
CEO Salary 2012
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Forbes
}
