% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cereal_company}
\alias{dasl.cereal_company}
\title{Cereal company}
\format{27 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cereal-company/?sf_paged=6}{Cereal company}
}
\description{
Cereal company
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data collected in a super market
}
\concept{Comparing Two Groups}
\concept{Nonparametric Methods}
