% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cereals}
\alias{dasl.cereals}
\title{Cereals}
\format{77 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cereals/?sf_paged=6}{Cereals}
}
\description{
Nutritionists are concerned that people have a good breakfast. But what does that mean? students collected nutrition information from the nutrition labels of cereals in one supermarket.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data collected in a super market
}
\concept{Analysis of Variance}
\concept{Data Display}
\concept{Multiple Regression}
\concept{Regression}
