% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cholesterol_and_smoking}
\alias{dasl.cholesterol_and_smoking}
\title{Cholesterol and smoking}
\format{43 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cholesterol-and-smoking/?sf_paged=7}{Cholesterol and smoking}
}
\description{
A study examined the health risks of smoking measured the cholesterol levels of people who had smoked for at least 25 years and people of similar ages who had smoked for no more than 5 years and then stopped
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Boxplots}
