% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.chromatography}
\alias{dasl.chromatography}
\title{Chromatography}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/chromatography/?sf_paged=7}{Chromatography}
}
\description{
Chromatography
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Fabricated from consulting data
}
\concept{Multifactor Analysis of Variance}
