% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cigarettes}
\alias{dasl.cigarettes}
\title{Cigarettes}
\format{816 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cigarettes/?sf_paged=7}{Cigarettes}
}
\description{
Data on 816 brands of cigarettes. What relationships are there among the nicotine content, tars, and CO? Are any brands unusually high or low in nicotine? Can you account for that?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.econdataus.com/cigrs.html}
}
\concept{Regression}
