% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.city_climate}
\alias{dasl.city_climate}
\title{City climate}
\format{60 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/city-climate/?sf_paged=7}{City climate}
}
\description{
The data give the average January Temperature (in degrees Fahrenheit) and Latitude (in degrees north of the equator) for 59 U.S. cities. How are they related?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
