% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.city_temperatures}
\alias{dasl.city_temperatures}
\title{City temperatures}
\format{35 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/city-temperatures/?sf_paged=7}{City temperatures}
}
\description{
City temperatures
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Paired Data}
