% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.climate_change_2016}
\alias{dasl.climate_change_2016}
\title{Climate change 2016}
\format{58 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/climate-change-2016/?sf_paged=7}{Climate change 2016}
}
\description{
Scientists claim that changes in the mean global temperature are primarily due to changes in CO2 levels. Both trends are here from 1959 to 2016. For an alternative, the data includes the annual closing price of the Dow Jones Industrial Average. Can it predict global temperature?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Global temperature from \url{https://www.ncdc.noaa.gov/cag/data-info/global}  Global temperature anomaly data come from the Global Historical Climatology Network-Monthly (GHCN-M) data set and International Comprehensive Ocean-Atmosphere Data Set (ICOADS), which have data from 1880 to the present. These two datasets are blended into a single product to produce the combined global land and ocean temperature anomalies. The available timeseries of global-scale temperature anomalies are calculated with respect to the 20th century average, while the mapping tool displays global-scale temperature anomalies with respect to the 1981-2010 base period. For more information on these anomalies, please visit Global Surface Temperature Anomalies. CO2 from ftp://aftp.cmdl.noaa.gov/products/trends/co2/co2_annmean_mlo.txt DJIA from \url{https://www.measuringworth.com}
}
\concept{Multiple Regression}
\concept{Regression}
