% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cloud_seeding}
\alias{dasl.cloud_seeding}
\title{Cloud seeding}
\format{26 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cloud-seeding/?sf_paged=7}{Cloud seeding}
}
\description{
In an experiment to determine whether seeding clouds with silver iodide increases rainfall, 52 clouds were randomly assigned to be seeded or not. The amount of rain they generated was then measured (in acre-feet).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Chambers, Cleveland, Kleiner, and Tukey. (1983). Graphical Methods for Data Analysis. Wadsworth International Group, Belmont, CA, 351. Original Source: Simpson, Alsen, and Eden. (1975). A Bayesian analysis of a multiplicative treatment effect in weather modification. Technometrics 17, 161-166.
}
\concept{Blocking}
\concept{Boxplots}
\concept{Comparing Two Groups}
\concept{Paired Data}
