% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.coasters_2015}
\alias{dasl.coasters_2015}
\title{Coasters 2015}
\format{241 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/coasters-2015/?sf_paged=7}{Coasters 2015}
}
\description{
The website rcdb.com, the Roller Coaster Database, holds facts about every roller coaster in the world, current or past. (If you know of one that is missing, please let the site master know.) These data are for recently opened coasters, most of which are still in operation.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
rcdb.com
}
\concept{Boxplots}
\concept{Correlation}
\concept{Multiple Regression}
\concept{Regression}
\concept{Scatterplot}
