% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.coffee_prices_2017}
\alias{dasl.coffee_prices_2017}
\title{Coffee prices 2017}
\format{336 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/coffee-prices-2017/?sf_paged=7}{Coffee prices 2017}
}
\description{
Coffee is the world’s second largest legal export commodity (after oil) and is the second largest source of foreign exchange for developing nations. The United States consumes about one-fifth of the world’s coffee. The International Coffee Organization (ICO) computes a coffee price index using Colombian, Brazilian, and a mixture of other coffee data. Data are provided for the monthly average ICO price index (in $US) from Jan 2009 to December 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.ico.org/new_historical.asp?section=Statistics}
}
\concept{Time Series Analysis}
