% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.col_2016}
\alias{dasl.col_2016}
\title{COL 2016}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/col-2016/?sf_paged=7}{COL 2016}
}
\description{
COL 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.numbeo.com/cost-of-living}
}
\concept{Boxplots}
