% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.colall_2016}
\alias{dasl.colall_2016}
\title{COLall 2016}
\format{576 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/colall-2016/?sf_paged=8}{COLall 2016}
}
\description{
The cost of a variety of common items in 576 cities around the world in $, adjusted so that New York, U.S.A. is 100.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.numbeo.com/cost-of-living}
}
\concept{Boxplots}
