% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.commercial_properties}
\alias{dasl.commercial_properties}
\title{Commercial properties}
\format{286 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/commercial-properties/?sf_paged=8}{Commercial properties}
}
\description{
Commercial properties
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
private communication
}
\concept{Regression}
