% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.commute_times_sample100}
\alias{dasl.commute_times_sample100}
\title{Commute times sample100}
\format{100 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/commute-times-sample100/?sf_paged=8}{Commute times sample100}
}
\description{
The data are a random sample from the data in Population commute times.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Random sample of 100 taken from Population Commute Times
}
\concept{Confidence Intervals for Means}
\concept{Data Display}
