% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.commute_times_sample50}
\alias{dasl.commute_times_sample50}
\title{Commute times sample50}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/commute-times-sample50/?sf_paged=8}{Commute times sample50}
}
\description{
Commute times sample50
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Random sample of 50 taken from Population Commute Times
}
\concept{Confidence Intervals for Means}
