% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.companies}
\alias{dasl.companies}
\title{Companies}
\format{79 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/companies/?sf_paged=8}{Companies}
}
\description{
Facts about companies selected from the Forbes 500 list for 1986. This is a 1/10 systematic sample from the alphabetical list of companies. The Forbes 500 includes all companies in the top 500 on any of the criteria, and thus has almost 800 companies in the list.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Forbes magazine
}
\concept{Boxplots}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
