% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.companies_quickstart}
\alias{dasl.companies_quickstart}
\title{Companies Quickstart}
\format{262 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/companies-quickstart/?sf_paged=8}{Companies Quickstart}
}
\description{
Facts about companies selected from the Forbes 500 list for 2000.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Forbes magazine
}
