% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.computer_chip_manufacturing}
\alias{dasl.computer_chip_manufacturing}
\title{Computer chip manufacturing}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/computer-chip-manufacturing/?sf_paged=8}{Computer chip manufacturing}
}
\description{
Computer chip manufacturing
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Quality Control}
