% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.computer_lab_fees}
\alias{dasl.computer_lab_fees}
\title{Computer lab fees}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/computer-lab-fees/?sf_paged=8}{Computer lab fees}
}
\description{
The technology committee at a school has stated that the average time spent by students per lab visit has increased and the increase supports their argument that they need to increase lab fees. To substantiate this claim, the committee randomly sampled 12 student lab visits and noted the amount of time spent using the computer. The times in minutes are given:
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
