% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.congestion_cost}
\alias{dasl.congestion_cost}
\title{Congestion cost}
\format{65 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/congestion-cost/?sf_paged=8}{Congestion cost}
}
\description{
Congestion cost
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Texas Transportation Institute
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
