% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.consumer_spending}
\alias{dasl.consumer_spending}
\title{Consumer spending}
\format{150 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/consumer-spending/?sf_paged=8}{Consumer spending}
}
\description{
Consumer spending
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data adapted from real spending data from American Express
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
