% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.consumer_spending_post_holiday}
\alias{dasl.consumer_spending_post_holiday}
\title{Consumer spending post holiday}
\format{1200 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/consumer-spending-post-holiday/?sf_paged=8}{Consumer spending post holiday}
}
\description{
Consumer spending post holiday
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data adapted from real spending data from American Express
}
\concept{Comparing Two Groups}
