% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.containers_revisited}
\alias{dasl.containers_revisited}
\title{Containers revisited}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/containers-revisited/?sf_paged=8}{Containers revisited}
}
\description{
Containers revisited
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Student experiment from De Veaux
}
\concept{Multifactor Analysis of Variance}
