% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.contrast_baths}
\alias{dasl.contrast_baths}
\title{Contrast baths}
\format{76 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/contrast-baths/?sf_paged=9}{Contrast baths}
}
\description{
Contrast baths
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
R. G. Janssen,D. A. Schwartz, and P. F. Velleman, “A Randomized Controlled Study of Contrast Baths on Patients with Carpal Tunnel Syndrome,” Journal of Hand Therapy, 2009.
}
\concept{Analysis of Variance}
\concept{Experiment Design}
