% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cookies}
\alias{dasl.cookies}
\title{Cookies}
\format{99 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cookies/?sf_paged=9}{Cookies}
}
\description{
A student performed an experiment to compare chocolate chip cookie recipes. He baked batches of cookies with different amounts of Sugar: 0.5, 0.375, and 0.25 cups, with three different kinds of oil, and with three different kinds of Chips: milk, semisweet, and dark chocolate. Cookie quality was scored on a 0 to 10 scale by other students in the class, with the mean score being the result for each recipe. The experiment design was a Latin Square.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
student experiment
}
\concept{Multifactor Analysis of Variance}
