% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cost_of_living_2016}
\alias{dasl.cost_of_living_2016}
\title{Cost of living 2016}
\format{576 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cost-of-living-2016/?sf_paged=9}{Cost of living 2016}
}
\description{
Numbeo.com lists the cost of living (COL) for 576 cities around the world. They report the typical cost of a number of staples. The cost of living is made up of many components. These data report a variety of everyday costs. How are they related? Can an overall cost of living be constructed from them?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.numbeo.com/cost-of-living}
}
\concept{Boxplots}
\concept{Regression}
