% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cost_of_living_2017}
\alias{dasl.cost_of_living_2017}
\title{Cost of Living 2017}
\format{511 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cost-of-living-2017/?sf_paged=9}{Cost of Living 2017}
}
\description{
Cost of Living 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
numbeo.com
}
\concept{Correlation}
\concept{Multiple Regression}
\concept{Regression}
\concept{Scatterplot}
