% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cost_of_living_2018}
\alias{dasl.cost_of_living_2018}
\title{Cost of living 2018}
\format{540 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cost-of-living-2018/?sf_paged=9}{Cost of living 2018}
}
\description{
Cost of Living Index (Excl. Rent) is a relative indicator of consumer goods prices, including groceries, restaurants, transportation and utilities. Cost of Living Index doesn’t include accommodation expenses such as rent or mortgage. If a city has a Cost of Living Index of 120, it means Numbeo estimates it is 20% more expensive than New York (excluding rent). Rent Index is an estimation of prices of renting apartments in the city compared to New York City. If Rent index is 80, Numbeo estimates that price of rents in that city is on an average 20% less than the price in New York. Groceries Index is an estimation of grocery prices in the city compared to New York City. To calculate this section, Numbeo uses weights of items in the “Markets” section for each city. Restaurants Index is a comparison of prices of meals and drinks in restaurants and bars compared to NYC. Cost of Living Plus Rent Index is an estimation of consumer goods prices including rent comparing to New York City. Local Purchasing Power shows relative purchasing power in buying goods and services in a given city for the average wage in that city. If domestic purchasing power is 40, this means that the inhabitants of that city with the average salary can afford to buy on an average 60% less goods and services than New York City residents with an average salary..
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.numbeo.com/cost-of-living/rankings.jsp}
}
\concept{Data Display}
