% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.couples}
\alias{dasl.couples}
\title{Couples}
\format{201 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/couples/?sf_paged=9}{Couples}
}
\description{
Couples
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Paired Data}
