% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cpi_2017}
\alias{dasl.cpi_2017}
\title{CPI 2017}
\format{132 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cpi-2017/?sf_paged=9}{CPI 2017}
}
\description{
CPI 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.bls.gov/cgi-bin/surveymost}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
\concept{Time Series Analysis}
