% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cpi_worldwide_2016}
\alias{dasl.cpi_worldwide_2016}
\title{CPI Worldwide 2016}
\format{497 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cpi-worldwide-2016/?sf_paged=9}{CPI Worldwide 2016}
}
\description{
The Consumer Price Index (CPI) summarizes the cost of a representative market basket of goods that includes groceries, restaurants, transportation, utilities, and medical care. Global companies often use the CPI to determine living allowances and salaries for employees. Inflation is often measured by how much the CPI changes from year to year. Relative CPIs can be found for different cities. We have data giving CPI components relative to New York City. For New York City, each index is 100(%).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.numbeo.com/cost-of-living/rankings.jsp}
}
