% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cramming}
\alias{dasl.cramming}
\title{Cramming}
\format{25 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cramming/?sf_paged=9}{Cramming}
}
\description{
Students in two basic Spanish classes were required to learn 50 new vocabulary words. One group of 45 students received the list on Monday and studied the words all week. Statistics summarizing this group’s scores on Friday’s quiz are given. The other group of 25 students did not get the vocabulary list until Thursday. They also took the quiz on Friday, after “cramming” Thursday night. Then, when they returned to class the following Monday, they were retested—without advance warning. Both sets of test scores for these students are given.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
