% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cranberry_juice}
\alias{dasl.cranberry_juice}
\title{Cranberry juice}
\format{150 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cranberry-juice/?sf_paged=9}{Cranberry juice}
}
\description{
Cranberry juice
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
