% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.craters}
\alias{dasl.craters}
\title{Craters}
\format{168 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/craters/?sf_paged=9}{Craters}
}
\description{
Meteor Crater in Arizona was the first recognized impact crater and was identified as such only in the 1920s. With the help of satellite images, more and more craters have been identified; now more than 180 are known. These, of course, are only a small sample of all the impacts the earth has experienced: Only 29% of earth’s surface is land, and many craters have been covered or eroded away. Astronomers have recog-nized a roughly 35 million-year cycle in the frequency of cratering, although the cause of this cycle is not fully understood. The data hold information about craters. craters from the most recent 35Ma (million years) may be the more reliable data, and are suitable for analyses relating age and diameter.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Earth Impact Database \url{http://www.unb.ca/passc/ImpactDatabase/}
}
\concept{Regression}
