% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.crawling}
\alias{dasl.crawling}
\title{Crawling}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/crawling/?sf_paged=9}{Crawling}
}
\description{
Researchers at the University of Denver Infant Study Center wondered whether temperature might influence the age at which babies learn to crawl. Perhaps the extra clothing that babies wear in cold weather would restrict movement and delay the age at which they started crawling. Data were collected on 208 boys and 206 girls. Parents reported the month of the baby’s birth and the age (in weeks) at which their child first crawled. The table gives the average Temperature (°F) when the babies were 6 months old and average Crawling Age (in weeks) for each month of the year.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Janette B. Benson,"Season of birth and onset of locomotion: Theoretical and methodological implications,"Infant Behavior and Development 16:1,pp 69-81
}
\concept{Confidence Intervals for Means}
\concept{Regression}
