% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.credit_card_bank}
\alias{dasl.credit_card_bank}
\title{Credit card bank}
\format{1800 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/credit-card-bank/?sf_paged=10}{Credit card bank}
}
\description{
Credit card bank
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
modified from large data set from bank consulting (RDD)
}
\concept{Data Displays}
