% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.credit_card_bank_info}
\alias{dasl.credit_card_bank_info}
\title{Credit card bank info}
\format{22 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/credit-card-bank-info/?sf_paged=10}{Credit card bank info}
}
\description{
Credit card bank info
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
information for data set modified from large data set from bank consulting (RDD)
}
\concept{Data Displays}
