% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.credit_card_charges}
\alias{dasl.credit_card_charges}
\title{Credit card charges}
\format{500 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/credit-card-charges/?sf_paged=10}{Credit card charges}
}
\description{
A credit card company wants to see how much customers in a particular segment of their market use their credit card. They have provided data on the amount spent by 500 selected customers during a 3-month period and have asked you to summarize the expenditures. (Data are realistic, but disguised for confidentiality.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
modified from raw
}
\concept{Data Display}
