% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.crocodile_lengths}
\alias{dasl.crocodile_lengths}
\title{Crocodile lengths}
\format{32 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/crocodile-lengths/?sf_paged=10}{Crocodile lengths}
}
\description{
The ranges inhabited by the Indian gharial crocodile and the Australian saltwater crocodile overlap in Bangladesh. Suppose a very large crocodile skeleton is found there, and we wish to determine the species of the animal. Wildlife scientists have measured the lengths of the heads and the complete bodies of several crocs (in centimeters) of each species.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown SDM 4
}
