% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.crowdedness}
\alias{dasl.crowdedness}
\title{Crowdedness}
\format{62 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/crowdedness/?sf_paged=10}{Crowdedness}
}
\description{
In a Chance magazine article (Summer 2005), Danielle Vasilescu and Howard Wainer used data from the United Nations Center for Human Settlements to investigate aspects of living conditions for several countries. Among the variables they looked at were the country’s per capita gross domestic product (GDP, in $) and Crowdedness, defined as the average number of persons per room living in homes there. Vasilescu and Wainer re-express GDP to -10000/GDP. Doing that reveals an outlier that may be due to an error in the data.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Chance magazine article summer 2005 by Vasilescu and Wainer, citing data fromn the United Nations Center for Human Settlements.
}
\concept{Nonparametric Methods}
\concept{Regression}
