% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cuckoos_and_quality_control}
\alias{dasl.cuckoos_and_quality_control}
\title{Cuckoos and quality control}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cuckoos-and-quality-control/?sf_paged=10}{Cuckoos and quality control}
}
\description{
Cuckoos and quality control
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
The data are drawn from the work of O. M. Latter in 1902 and were used in a fundamental textbook on statistical quality control by L. H. C. Tippett (1902–1985), one of the pioneers in that field.
}
\concept{Nonparametric Methods}
