% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cups}
\alias{dasl.cups}
\title{Cups}
\format{32 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cups/?sf_paged=10}{Cups}
}
\description{
Cups
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
From Stat 231 De Veaux’s class project
}
\concept{Analysis of Variance}
\concept{Boxplots}
