% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.cyclists_2015}
\alias{dasl.cyclists_2015}
\title{Cyclists 2015}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/cyclists-2015/?sf_paged=10}{Cyclists 2015}
}
\description{
Cyclists 2015
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
National Highway Transportation Safety Agency, found at www-nrd.nhtsa.dot.gov
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
