% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.dalia}
\alias{dasl.dalia}
\title{Dalia}
\format{13356 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/dalia/?sf_paged=10}{Dalia}
}
\description{
Dalia collects data via smartphone from users worldwide. This survey asked (among many other questions) about access to cars and the use of ride-hailing apps
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
DaliaReasearch (private communication)
}
\concept{Data Displays}
