% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.darts}
\alias{dasl.darts}
\title{Darts}
\format{36 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/darts/?sf_paged=10}{Darts}
}
\description{
Darts
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Student experiment from De Veaux
}
\concept{Multifactor Analysis of Variance}
