% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.defect_monitoring}
\alias{dasl.defect_monitoring}
\title{Defect monitoring}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/defect-monitoring/?sf_paged=10}{Defect monitoring}
}
\description{
The data are from a production process that makes 250 units each hour. The data were collected over a normal 12-hour shift one day.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Quality Control}
