% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.demographics}
\alias{dasl.demographics}
\title{Demographics}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/demographics/?sf_paged=11}{Demographics}
}
\description{
Demographics
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
census.gov
}
\concept{Multiple Regression}
