% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.depression_and_the_internet}
\alias{dasl.depression_and_the_internet}
\title{Depression and the internet}
\format{169 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/depression-and-the-internet/?sf_paged=11}{Depression and the internet}
}
\description{
The September 1998 issue of the American T Psychologist published an article by Kraut et al. that reported on an experiment examining “the social and psychological impact of the Internet on 169 people in 73 households during their first 1 to 2 years online.” In the experiment, 73 households were offered free Internet access for 1 or 2 years in return for allowing their time and activity online to be tracked. The members of the households who participated in the study were also given a battery of tests at the beginning and again at the end of the study. The conclusion of the study made news headlines: Those who spent more time online tended to be more depressed at the end of the experiment. The news reports about this study clearly concluded that using the Internet causes depression. Is such a conclusion warranted?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
"Internet paradox. A social technology that reduces social involvement and psychological well-being?", Kraut R1, Patterson M, Lundmark V, Kiesler S, Mukopadhyay T, Scherlis W., Am Psychol. 1998 Sep;53(9):1017-31.
}
