% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.designated_hitter_2012}
\alias{dasl.designated_hitter_2012}
\title{Designated hitter 2012}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/designated-hitter-2012/?sf_paged=11}{Designated hitter 2012}
}
\description{
In the National league all players take a turn at bat. But in the American league, a “designated hitter” usually bats for the pitcher, who is likely not to be a strong batter. The theory is that a designated hitter will lead to more hits, more runs, and a higher-scoring game. The data give the average runs per game and total home runs for major league baseball teams during the 2012 season. Is there a discernible difference between the leagues?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
mlb.com
}
\concept{Nonparametric Methods}
