% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.dexterity}
\alias{dasl.dexterity}
\title{Dexterity}
\format{93 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/dexterity/?sf_paged=11}{Dexterity}
}
\description{
Dexterity
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
personal communication
}
\concept{Blocking}
\concept{Paired Data}
