% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.diamond_quality}
\alias{dasl.diamond_quality}
\title{Diamond quality}
\format{2690 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/diamond-quality/?sf_paged=11}{Diamond quality}
}
\description{
Diamond quality
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data scraped from the web by Lou Valente of JMP
}
\concept{Nonparametric Methods}
