% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.diamonds}
\alias{dasl.diamonds}
\title{Diamonds}
\format{2690 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/diamonds/?sf_paged=11}{Diamonds}
}
\description{
Data on raw diamonds from the internet. Price of a diamond depends on its Carat weight, color, clarity, and cut. The data are for 2690 diamonds of a variety of weights, colors, clarity, and cut. What predicts the price? Do the variables need to be reexpressed?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data scraped from the web by Lou Valente of JMP
}
\concept{Multiple Regression}
\concept{Regression}
