% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.direct_mail_follow_up}
\alias{dasl.direct_mail_follow_up}
\title{Direct mail follow up}
\format{30000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/direct-mail-follow-up/?sf_paged=11}{Direct mail follow up}
}
\description{
Direct mail follow up
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Based on consulting data (R. De Veaux)
}
\concept{Analysis of Variance}
\concept{Experiment Design}
