% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.dirt_bikes_2014}
\alias{dasl.dirt_bikes_2014}
\title{Dirt bikes 2014}
\format{114 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/dirt-bikes-2014/?sf_paged=11}{Dirt bikes 2014}
}
\description{
Motorcycles designed to run off-road, often known as dirt bikes, are specialized vehicles. The dataset holds data on 114 many attributes of dirt bikes. Some cost as little as $1399, while others are substantially more expensive. One interest is in building a model to predict the price of a dirt bike from attributes of the bikes.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Jiang Lu, Joseph B. Kadane, and Peter Boatwright, The Dirt on Bikes: An Illustration of CART Models for Brand Differentiation
}
\concept{Multiple Regression}
\concept{Regression}
