% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.disk_drives_2016}
\alias{dasl.disk_drives_2016}
\title{Disk drives 2016}
\format{9 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/disk-drives-2016/?sf_paged=11}{Disk drives 2016}
}
\description{
Disk drive capacity is often given in terabytes (TB), where 1 TB = 1000 gigabytes, or about a trillion bytes. A search of prices for external disk drives on Amazon.com in mid-2016 found the data on capacity and price.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.amazon.com/s/ref=nb_sb_noss?url=search-alias} %3Daps&field-keywords=32+TB+hard+drive&rh=i%3Aaps%2Ck%3A32+TB+hard+drive
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
