% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.djia_2017}
\alias{dasl.djia_2017}
\title{DJIA 2017}
\format{2142 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/djia-2017/?sf_paged=12}{DJIA 2017}
}
\description{
DJIA 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://finance.yahoo.com/quote/} %5EDJI/history/
}
\concept{Normal model}
\concept{Time Series Analysis}
