% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.doctors_and_life_expectancy}
\alias{dasl.doctors_and_life_expectancy}
\title{Doctors and life expectancy}
\format{40 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/doctors-and-life-expectancy/?sf_paged=12}{Doctors and life expectancy}
}
\description{
Life expectancy at birth, TV’s per capita, and doctor’s per capita for countries of the world. Doctors predict life expectancy, but is that causal? TVs also predict life expectancy.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
