% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.donor_profiles}
\alias{dasl.donor_profiles}
\title{Donor profiles}
\format{916 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/donor-profiles/?sf_paged=12}{Donor profiles}
}
\description{
Donor profiles
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
PVA data set sample
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
