% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.doritos}
\alias{dasl.doritos}
\title{Doritos}
\format{6 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/doritos/?sf_paged=12}{Doritos}
}
\description{
Some students checked 6 bags of Doritos marked with a net weight of 28.3 grams. They carefully weighed the contents of each bag and recorded the weights in grams.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
