% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.dow_jones_2015}
\alias{dasl.dow_jones_2015}
\title{Dow Jones 2015}
\format{44 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/dow-jones-2015/?sf_paged=12}{Dow Jones 2015}
}
\description{
The Dow Jones stock index measures the performance of the stocks of America’s largest companies. A regression of the Dow prices on years 1972–2015 appears to be successful, but the residuals raise some questions.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
finance.yahoo.com
}
