% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.dow_jones_2017}
\alias{dasl.dow_jones_2017}
\title{Dow Jones 2017}
\format{259 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/dow-jones-2017/?sf_paged=12}{Dow Jones 2017}
}
\description{
Dow Jones 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://fred.stlouisfed.org}
}
\concept{Time Series Analysis}
