% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.down_the_drain}
\alias{dasl.down_the_drain}
\title{Down the Drain}
\format{7 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/down-the-drain/?sf_paged=12}{Down the Drain}
}
\description{
Most water tanks have a drain plug so that the tank may be emptied when it’s to be moved or repaired. How long it takes a certain size of tank to drain depends on the size of the plug, as shown in the table.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Classical calculus exercise
}
