% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.downloading}
\alias{dasl.downloading}
\title{Downloading}
\format{48 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/downloading/?sf_paged=12}{Downloading}
}
\description{
To see how much of a difference time of day made on the speed at which he could download files, a college sophomore performed an experiment. He placed a file on a remote server and then proceeded to download it at three different time periods of the day. He downloaded the file 48 times in all, 16 times at each Time of Day, and recorded the Time in seconds that the download took.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
student project
}
\concept{Analysis of Variance}
