% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.downloads}
\alias{dasl.downloads}
\title{Downloads}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/downloads/?sf_paged=12}{Downloads}
}
\description{
Downloads
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Manufactured data
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
