% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.drivers_licenses_2014}
\alias{dasl.drivers_licenses_2014}
\title{Drivers Licenses 2014}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/drivers-licenses-2014/?sf_paged=12}{Drivers Licenses 2014}
}
\description{
Drivers Licenses 2014
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.fhwa.dot.gov/policyinformation/statistics.cfm}
}
\concept{Tables}
