% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.drug_use_2013}
\alias{dasl.drug_use_2013}
\title{Drug use 2013}
\format{33 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/drug-use-2013/?sf_paged=12}{Drug use 2013}
}
\description{
The 2013 World Drug Report investigated the prevalence of drug use as a percentage of the population aged 15 to 64. Data from 32 European countries are shown.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
World Drug Report, 2013. www.unodc .org/unodc/en/data-and-analysis/WDR-2013.html
}
\concept{Regression}
