% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.e_commerce}
\alias{dasl.e_commerce}
\title{E-commerce}
\format{34 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/e-commerce/?sf_paged=13}{E-commerce}
}
\description{
Quarterly e-commerce retail sales (in millions of dollars) in the United States from 1999 to 2008
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
U.S. Census Bureau; www.census.gov
}
\concept{Time Series Analysis}
