% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.e_mails}
\alias{dasl.e_mails}
\title{E-mails}
\format{155 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/e-mails/?sf_paged=13}{E-mails}
}
\description{
A university teacher saved every e-mail receive from students in a large introductory statistics class during one term. He then counted, for each student who had sent him at least one e-mail, how many e-mails each student had sent. What is the distribution of e-mail communications?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Velleman personal communication
}
\concept{Data Display}
