% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.earnings}
\alias{dasl.earnings}
\title{Earnings}
\format{706 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/earnings/?sf_paged=13}{Earnings}
}
\description{
Earnings
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://new.time.com/money/best-colleges/rankings/best-colleges/}
}
\concept{Regression}
