% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.education_and_mortality}
\alias{dasl.education_and_mortality}
\title{Education and mortality}
\format{58 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/education-and-mortality/?sf_paged=13}{Education and mortality}
}
\description{
Data give the mortality rate (deaths per 100,000 people) and the education level (average number of years in school) for 58 U.S. cities.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Regression}
