% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.education_by_age}
\alias{dasl.education_by_age}
\title{Education by age}
\format{1000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/education-by-age/?sf_paged=13}{Education by age}
}
\description{
Education by age
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
