% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.egyptians}
\alias{dasl.egyptians}
\title{Egyptians}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/egyptians/?sf_paged=13}{Egyptians}
}
\description{
Egyptians
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
A.Thomson and R.Randall-Maciver, Ancient Races of the Thebaid, Oxford: Oxford University Press, 1905
}
\concept{Test for Difference between Two Groups}
