% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.el_nino_2017}
\alias{dasl.el_nino_2017}
\title{El Nino 2017}
\format{58 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/el-nino-2017/?sf_paged=13}{El Nino 2017}
}
\description{
El Nino 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
personal communication
}
\concept{Blocking}
\concept{Correlation}
\concept{Paired Data}
\concept{Regression}
\concept{Scatterplot}
