% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.election_2000}
\alias{dasl.election_2000}
\title{Election 2000}
\format{67 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/election-2000/?sf_paged=13}{Election 2000}
}
\description{
The outcome of the 2000 U.S. presidential election was determined in Florida amid much controversy. Even years later, historians continue to debate who really received the most votes. The main race was between George W. Bush and Al Gore, but two minor candidates played a significant role. To the political right of the major party candidates was Pat Buchanan, while to the political left was Ralph Nader. Generally, Nader earned more votes than Buchanan throughout the state. We would expect counties with larger vote totals to give more votes to each candidate. The dataset gives Buchanan’s and Nader’s vote totals by county in the state of Florida. Plot to identify the outlier and consider what it means.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.american.edu/spa/ccps/Data-Sets.cfm}
}
\concept{Regression}
