% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.energy_use_and_recession}
\alias{dasl.energy_use_and_recession}
\title{Energy use and recession}
\format{33 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/energy-use-and-recession/?sf_paged=13}{Energy use and recession}
}
\description{
Energy use and recession
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Regression}
