% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.epa_mpg_2013}
\alias{dasl.epa_mpg_2013}
\title{EPA mpg 2013}
\format{76 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/epa-mpg-2013/?sf_paged=13}{EPA mpg 2013}
}
\description{
EPA mpg 2013
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.fueleconomy.gov/}  But no longer easily obtained for large numbers of das
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
