% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.erp_effectiveness}
\alias{dasl.erp_effectiveness}
\title{ERP Effectiveness}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/erp-effectiveness/?sf_paged=13}{ERP Effectiveness}
}
\description{
When implementing a packaged Enterprise Resource Planning (ERP) system, many companies report that the module they first install is Financial Accounting. Among the measures used to gauge the effectiveness of their ERP system implementation is acceleration of the financial close process. The data hold a sample of 8 companies that report their average time (in weeks) to financial close before and after the implementation of their ERP system.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Comparing Two Groups}
\concept{Nonparametric Methods}
