% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.eu_imports}
\alias{dasl.eu_imports}
\title{EU Imports}
\format{197 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/eu-imports/?sf_paged=14}{EU Imports}
}
\description{
EU Imports
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Download from the bottom of www .census.gov/foreign-trade/balance/c5700.htm
}
\concept{Time Series Analysis}
