% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.euro_dollar}
\alias{dasl.euro_dollar}
\title{Euro Dollar}
\format{52 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/euro-dollar/?sf_paged=14}{Euro Dollar}
}
\description{
Euro Dollar
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://fred.stlouisfed.org}
}
\concept{Time Series Analysis}
