% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.euro_us_gdp_growth_2017}
\alias{dasl.euro_us_gdp_growth_2017}
\title{Euro-US GDP growth 2017}
\format{56 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/euro-us-gdp-growth-2017/?sf_paged=14}{Euro-US GDP growth 2017}
}
\description{
Euro-US GDP growth 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.worldbank.org/indicator/NY.GDP.MKTP.KD.ZG}
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
