% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.exercise_equipment}
\alias{dasl.exercise_equipment}
\title{Exercise equipment}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/exercise-equipment/?sf_paged=14}{Exercise equipment}
}
\description{
A leading manufacturer of exercise equipment wanted to collect data on the effectiveness of their equipment. An August 2001 article in the journal Medicine and Science in Sports and Exercise compared how long it would take men and women to burn 200 calories during light or heavy workouts on various kinds of exercise equipment. The results summarized in the table are the average times for a group of physically active young men and women whose performances were measured on a representative sample of exercise equipment.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Medicine and Science in Sports and Exercise, 2001
}
\concept{Comparing Two Groups}
