% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.expensive_cities}
\alias{dasl.expensive_cities}
\title{Expensive cities}
\format{40 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/expensive-cities/?sf_paged=14}{Expensive cities}
}
\description{
Expensive cities
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
numbeo.com
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
