% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.eye_and_hair_color}
\alias{dasl.eye_and_hair_color}
\title{Eye and Hair color}
\format{1021 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/eye-and-hair-color/?sf_paged=14}{Eye and Hair color}
}
\description{
Students in a large statistics class were asked to report the eye color and hair color. Is there an association?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Student survey
}
\concept{Analysis of Variance}
