% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.f_stops}
\alias{dasl.f_stops}
\title{F-stops}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/f-stops/?sf_paged=14}{F-stops}
}
\description{
Some camera lenses have an adjustable aperture, the hole that lets the light in. The size of the aperture is expressed in a mysterious number called the f/stop. Each increase of one f/stop number corresponds to a halving of the light that is allowed to come through. When you halve the shutter speed, you cut down the light, so you have to open the aperture one notch. We could experiment to find the best f/stop value for each shutter speed. A table of recommended shutter speeds and f/stops for a camera lists the relationship like those given here:
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://garsteckphotography.blogspot.com/2011/04/using-manual-settings-must-know.html}
}
\concept{Correlation}
\concept{Scatterplot}
