% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.farmed_salmon}
\alias{dasl.farmed_salmon}
\title{Farmed salmon}
\format{153 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/farmed-salmon/?sf_paged=14}{Farmed salmon}
}
\description{
In 2004, a team of researchers published a study of contaminants in farmed salmon. Fish from many sources were analyzed for 14 organic contaminants. The study expressed concerns about the level of contaminants found. One of those was the insecticide mirex, which has been shown to be carcinogenic and is suspected to be toxic to the liver, kidneys, and endocrine system. The dataset holds 153 observed salmon samples and reports concentrations of a number of contaminant.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Ronald A. Hites, Jeffery A. Foran, David O. Carpenter, M. Coreen Hamilton, Barbara A. Knuth, and Steven J. Schwager, “Global Assessment of Organic Contaminants in Farmed Salmon,” Science 9 January 2004: Vol. 303, no. 5655, pp. 226–229
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Statistical Inference}
