% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.faulty_pixels}
\alias{dasl.faulty_pixels}
\title{Faulty pixels}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/faulty-pixels/?sf_paged=14}{Faulty pixels}
}
\description{
Faulty pixels
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented
}
\concept{Quality Control}
