% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.female_president}
\alias{dasl.female_president}
\title{Female president}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/female-president/?sf_paged=14}{Female president}
}
\description{
Female president
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Ronald A. Hites, Jeffery A. Foran, David O. Carpenter, M. Coreen Hamilton, Barbara A. Knuth, and Steven J. Schwager, “Global Assessment of Organic Contaminants in Farmed Salmon,” Science 9 January 2004: Vol. 303, no. 5655, pp. 226–229
}
\concept{Regression}
