% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fertility_and_life_expectancy_2014}
\alias{dasl.fertility_and_life_expectancy_2014}
\title{Fertility and life expectancy 2014}
\format{219 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fertility-and-life-expectancy-2014/?sf_paged=14}{Fertility and life expectancy 2014}
}
\description{
Fertility (births/woman) and Female life expectancy for 219 countries of the world. (Data is available on both variables for only 200). How is life expectancy related to fertility? Are there any outliers and, if so, what do they indicate?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://data.worldbank.org/indicator/SP.DYN.LE00.FE.IN}   for life expectancy at birth \url{http://data.worldbank.org/indicator/SP.DYN.TFRT.IN}  for fertility rates
}
\concept{Regression}
