% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fertilizers}
\alias{dasl.fertilizers}
\title{Fertilizers}
\format{120 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fertilizers/?sf_paged=14}{Fertilizers}
}
\description{
A biology student studied the effect of10 different fertilizers on the growth of mung bean sprouts. She sprouts 12 beans in each of 10 different petri dishes, and adds the same amount of fertilizer to each dish. After one week she measures the heights of the 120 sprouts in millimeters.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://pbil.univ-lyon1.fr/R/donnees/criminals1902.txt}  Garson, J.G. (1900) The metric system of identification of criminals, as used in in Great Britain and Ireland. The Journal of the Anthropological Institute of Great Britain and Ireland 30, 161–198. MacDonell, W.R. (1902) On criminal anthropometry and the identification of criminals. Biometrika 1, 2, 177–227. Student (1908) The probable error of a mean. Biometrika 6, 1–25.
}
\concept{Analysis of Variance}
