% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fingers_and_heights}
\alias{dasl.fingers_and_heights}
\title{Fingers and Heights}
\format{3000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fingers-and-heights/?sf_paged=15}{Fingers and Heights}
}
\description{
Gossett says in his seminal 1908 paper: “Before I had succeeded in solving my problem analytically, I had endeavoured to do so empirically. The material used was a correlation table containing the height and left middle finger measurements of 3000 criminals, from a paper by W. R. MacDonell (Biometrika, Vol. I., p. 219).” His method was to write the 3000 finger length values on cards, shuffle them thoroughly, and the deal out 750 hands of 4 cards. For each hand he then calculated (with a mechanical calculator) the mean and standard deviation. (Note; He divided by n (= 4) and not by n-1 (= 3).) He then found values of ybar – the population mean (which he knew because he had the population; it is 11.5474) and divided each by the standard deviation. The resulting values formed the distribution which he then correctly described. The finger measurements were originally given in mm and the heights in feet and inches. They have been converted to cm (at \url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/crimtab.html).}  The midpoint of intervals are used where MacDonnel gives a range of values.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://pbil.univ-lyon1.fr/R/donnees/criminals1902.txt}  Garson, J.G. (1900) The metric system of identification of criminals, as used in in Great Britain and Ireland. The Journal of the Anthropological Institute of Great Britain and Ireland 30, 161–198. MacDonell, W.R. (1902) On criminal anthropometry and the identification of criminals. Biometrika 1, 2, 177–227. Student (1908) The probable error of a mean. Biometrika 6, 1–25.
}
\concept{Confidence Intervals for Means}
