% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fire_sale}
\alias{dasl.fire_sale}
\title{Fire sale}
\format{60 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fire-sale/?sf_paged=15}{Fire sale}
}
\description{
Fire sale
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
