% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fish_diet}
\alias{dasl.fish_diet}
\title{Fish diet}
\format{6272 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fish-diet/?sf_paged=15}{Fish diet}
}
\description{
Medical researchers followed 6272 Swedish men for 30 years to see whether there was any association between the amount of fish in their diet and prostate cancer. The original study actually used pairs of twins, which enabled the researchers to discern that the risk of cancer for those who never ate fish actually was substantially greater.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Lancet, June 2001, “Fatty Fish Consumption and Risk of Prostate Cancer,”
}
\concept{Chi Square Tests}
\concept{Tables}
