% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fishers_irises}
\alias{dasl.fishers_irises}
\title{Fisher&#8217;s Irises}
\format{150 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fishers-irises/?sf_paged=15}{Fisher&#8217;s Irises}
}
\description{
In 1936 Sir Ronald Fisher presented data on irises as the example in a famous statistics paper. Ever since, “Fisher’s Iris data” have been a feature of statistics texts. Fisher presents 4 measurements of Iris flowers of three species. Can we differentiate the species? If so, how best to do that?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Fisher, R. A. (1936). The Use of Multiple Measurements in  Axonomic Problems. Annals of Eugenics  7, 179-188.
}
