% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fitness_data}
\alias{dasl.fitness_data}
\title{Fitness data}
\format{618 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fitness-data/?sf_paged=15}{Fitness data}
}
\description{
Fitness data
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
De Veaux personal communication
}
\concept{Correlation}
\concept{Scatterplot}
