% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.flights_2016}
\alias{dasl.flights_2016}
\title{Flights 2016}
\format{17 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/flights-2016/?sf_paged=15}{Flights 2016}
}
\description{
The data give the number of domestic U.S. flights flown in each year from 2000 to 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.transtats.bts.gov/Data_Elements.aspx?Data=2}
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
