% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.flights_on_time_2016}
\alias{dasl.flights_on_time_2016}
\title{Flights on time 2016}
\format{270 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/flights-on-time-2016/?sf_paged=15}{Flights on time 2016}
}
\description{
The Bureau of Transportation Statistics of the U.S. Department of Transportation publishes information about airline performance. The data report the percentage of flights departing on time each month from January 1994 through June 2016.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.transtats.bts.gov/HomeDrillChart.asp?URL_SelectYear=2016&URL_SelectMonth=6&URL_Time=1&URL_Selection=1}
}
\concept{Boxplots}
\concept{Confidence Intervals for Means}
