% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.floods_2015}
\alias{dasl.floods_2015}
\title{Floods 2015}
\format{21 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/floods-2015/?sf_paged=15}{Floods 2015}
}
\description{
The annual number of deaths from floods in the United states from 1995 through 2015. Years are not provided, but the data values are in time order.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
NOAA
}
\concept{Data Display}
