% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.food_consumption}
\alias{dasl.food_consumption}
\title{Food consumption}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/food-consumption/?sf_paged=15}{Food consumption}
}
\description{
Food consumption
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
FAOSTAT, the Food and Agriculture Organization of the United Nations
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
