% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.food_sales}
\alias{dasl.food_sales}
\title{Food sales}
\format{18 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/food-sales/?sf_paged=15}{Food sales}
}
\description{
Sales (in $) for one week were collected for 18 stores in a food store chain in the northeastern United States. The stores and the towns they are located in vary in size.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
