% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.football_salaries_2017}
\alias{dasl.football_salaries_2017}
\title{Football salaries 2017}
\format{32 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/football-salaries-2017/?sf_paged=15}{Football salaries 2017}
}
\description{
\url{http://www.spotrac.com/nfl/cap/}  (salaries) \url{http://www.espn.com/nfl/standings/_/season/2016}    (wins)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Football owners are constantly in competition for good players. The more wins, the more likely that the team will provide good business returns for the owners. The resources that each of the 32 teams has in the National Football League (NFL) vary, but the draft system is designed to counteract the advantages that wealthier teams may have.
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
