% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.foreclosure_rates_2008}
\alias{dasl.foreclosure_rates_2008}
\title{Foreclosure rates 2008}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/foreclosure-rates-2008/?sf_paged=15}{Foreclosure rates 2008}
}
\description{
Foreclosure rates 2008
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
realestate.msn.com; MSN Real Estate, April 2008
}
\concept{Nonparametric Methods}
