% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.foreclosures_2016}
\alias{dasl.foreclosures_2016}
\title{Foreclosures 2016}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/foreclosures-2016/?sf_paged=16}{Foreclosures 2016}
}
\description{
Foreclosures 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.statisticbrain.com/home-foreclosure-statistics/}
}
\concept{Data Displays}
