% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.framingham}
\alias{dasl.framingham}
\title{Framingham}
\format{1406 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/framingham/?sf_paged=16}{Framingham}
}
\description{
The Framingham Heart Study is one of the longest running health studies. It has followed original subjects, their children, and their grand children, looking for factors that affect cardiac health. These data only include subjects whose cholesterol was measured in the first exam. Source:  “Statistical Methods in Epidemiology” by H.A.Kahn and C.T.Sempos SBP: Systolic blood pressure at first exam DBP: Diastolic blood pressure at first exam CHOL: Serum choloesterol at first exam FRW :  Framingham relative weight; a standardized measure of weight adjusted for sex and height CIG: Number of cigarettes smoked/day at first exam DEATH: First biannual exam missed due to death; 0=”alive at tenth biannual exam.” (This exam wasgiven in the 18th year of the study.) CAUSE: 0=aliv e at exam 10, 1=Coronary Heart Disease (sudden), 2=CHD (not sudden), 3=Stroke,4=Other cardiovascular disease, 5=cancer, 6=other
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
"Statistical Methods in Epidemiology" by H.A.Kahn and C.T.Sempos
}
\concept{Boxplots}
\concept{Confidence Intervals for Means}
\concept{Correlation}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Regression}
\concept{Scatterplot}
\concept{Test for Difference between Two Groups}
