% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.freshman_15}
\alias{dasl.freshman_15}
\title{Freshman 15}
\format{68 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/freshman-15/?sf_paged=16}{Freshman 15}
}
\description{
Is it true that students tend to gain weight during their first year in college? Cornell Professor of Nutrition David Levitsky recruited students from two large sections of an introductory health course. Although they were volunteers, they appeared to match the rest of the freshman class in terms of demographic variables such as sex and ethnicity. The students were weighed during the first week of the semester, then again 12 weeks later. Based on Professor Levitsky’s data, estimate the mean weight gain in first-semester freshmen and comment on the “freshman 15.” (Weights are in pounds.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Prof David Levitsky, Cornell University
}
\concept{Blocking}
\concept{Nonparametric Methods}
\concept{Paired Data}
