% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.friday_the_13th_accidents}
\alias{dasl.friday_the_13th_accidents}
\title{Friday the 13th Accidents}
\format{6 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/friday-the-13th-accidents/?sf_paged=16}{Friday the 13th Accidents}
}
\description{
Many people fear Friday the 13th as an unlucky day. Researchers looked into this to see whether there were differences in traffic or in admissions to hospitals for road accidents on Friday 13th when compared with the adjacent Friday 6th.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
British Medical Journal Is Friday the 13th bad for your health? BMJ 1993;307:1584
}
\concept{Blocking}
\concept{Nonparametric Methods}
\concept{Paired Data}
