% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.frisbee_throws}
\alias{dasl.frisbee_throws}
\title{Frisbee throws}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/frisbee-throws/?sf_paged=16}{Frisbee throws}
}
\description{
A student performed an experiment with three different grips to see what effect it might have on the distance of a backhanded Frisbee throw. She tried it with her normal grip, with one finger out, and with the Frisbee inverted. She measured in paces how far her throws went.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
From student experiment (De Veaux)
}
\concept{Analysis of Variance}
