% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fudge}
\alias{dasl.fudge}
\title{Fudge}
\format{25 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fudge/?sf_paged=16}{Fudge}
}
\description{
Fudge
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Manufactured
}
\concept{Quality Control}
