% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fuel_economy}
\alias{dasl.fuel_economy}
\title{Fuel economy}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fuel-economy/?sf_paged=16}{Fuel economy}
}
\description{
Fuel economy
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.fueleconomy.gov/feg/download.shtml}
}
\concept{Correlation}
\concept{Displaying Quantitative Data}
\concept{Regression}
\concept{Scatterplot}
\concept{Summarizing Quantitative Data}
