% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fuel_economy_2016}
\alias{dasl.fuel_economy_2016}
\title{Fuel economy 2016}
\format{1211 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fuel-economy-2016/?sf_paged=16}{Fuel economy 2016}
}
\description{
The U.S. government provides fuel economy (in miles per gallon) and other information about late model cars sold in the US. How would you model the relationship between fuel economy and engine displacement (in liters)? Are there any cars that don’t fit the model? Can you explain why?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.fueleconomy.gov/feg/download.shtml}
}
\concept{Boxplots}
\concept{Confidence Intervals for Means}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
