% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fuel_economy_and_cylinders}
\alias{dasl.fuel_economy_and_cylinders}
\title{Fuel economy and cylinders}
\format{38 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fuel-economy-and-cylinders/?sf_paged=16}{Fuel economy and cylinders}
}
\description{
Fuel economy (mpg) and the number of cylinders in a sample of cars. Data extracted from a larger cars dataset.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Boxplots}
