% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fuel_economy_sample_2016}
\alias{dasl.fuel_economy_sample_2016}
\title{Fuel economy sample 2016}
\format{35 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fuel-economy-sample-2016/?sf_paged=16}{Fuel economy sample 2016}
}
\description{
The U.S. government provides fuel economy (in miles per gallon) and other information about late model cars sold in the US. How would you model the relationship between fuel economy and engine displacement (in liters)? Are there any cars that don’t fit the model? Can you explain why? This is a sample of 35 cars from the Fuel economy 2016 dataset.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.fueleconomy.gov/feg/download.shtml}
}
\concept{Correlation}
\concept{Scatterplot}
