% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.fuel_efficiency}
\alias{dasl.fuel_efficiency}
\title{Fuel efficiency}
\format{84 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/fuel-efficiency/?sf_paged=17}{Fuel efficiency}
}
\description{
We know from common sense and from Physics that heavier cars need more fuel, but exactly how does a car’s weight affect its fuel efficiency? The data set continues data on 38 cars including their fuel efficiency in miles per gallon measured on a track.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Consumer Reports
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
