% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.full_moon}
\alias{dasl.full_moon}
\title{Full moon}
\format{129 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/full-moon/?sf_paged=17}{Full moon}
}
\description{
Full moon
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
