% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gas_additives}
\alias{dasl.gas_additives}
\title{Gas additives}
\format{60 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gas-additives/?sf_paged=17}{Gas additives}
}
\description{
An experiment to test a new gasoline additive, Gasplus, was performed on three different cars: a sports car, a minivan, and a hybrid. Each car was tested with both Gasplus and regular gas on 10 different occasions and their gas mileage was recorded.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Multifactor Analysis of Variance}
