% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gas_prices_2016}
\alias{dasl.gas_prices_2016}
\title{Gas prices 2016}
\format{1287 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gas-prices-2016/?sf_paged=17}{Gas prices 2016}
}
\description{
Weekly gas prices for regular gas in the United States as reported by the U.S. Energy Information Administration for 2009 through August 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.eia.gov/dnav/pet/pet_pri_gnd_dcus_nus_w.htm}
}
\concept{Boxplots}
