% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gas_prices_2017}
\alias{dasl.gas_prices_2017}
\title{Gas Prices 2017}
\format{44 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gas-prices-2017/?sf_paged=17}{Gas Prices 2017}
}
\description{
Gas Prices 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.losangelesgasprices.com/GasPriceSearch.aspx}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
\concept{Time Series Analysis}
