% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gas_prices_2018}
\alias{dasl.gas_prices_2018}
\title{Gas prices 2018}
\format{301 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gas-prices-2018/?sf_paged=17}{Gas prices 2018}
}
\description{
Monthly gas prices for all grades and all formulations ($/gallon) in the United States as reported by the U.S. Energy Information Administration for 1993 through August 2018. Prices are available at the cite for all weeks. Data here are for the final week of each month.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.eia.gov/dnav/pet/hist/LeafHandler.ashx?n=PET&s=EMM_EPM0_PTE_NUS_DPG&f=W}
}
\concept{Boxplots}
