% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gators}
\alias{dasl.gators}
\title{Gators}
\format{25 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gators/?sf_paged=17}{Gators}
}
\description{
Wildlife researchers monitor many wildlife populations by taking aerial photographs. Can they estimate the weights of alligators accurately from the air? Here are data on the Weight of alligators (in pounds) and their Length (in inches).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
