% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gdp_and_djia_2017}
\alias{dasl.gdp_and_djia_2017}
\title{GDP and DJIA 2017}
\format{57 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gdp-and-djia-2017/?sf_paged=17}{GDP and DJIA 2017}
}
\description{
GDP and DJIA 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.worldbank.org/indicator/NY.GDP.MKTP.KD.ZG}  www.measuringworth.com
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
