% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gdp_by_state}
\alias{dasl.gdp_by_state}
\title{GDP by state}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gdp-by-state/?sf_paged=17}{GDP by state}
}
\description{
GDP by state
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_U.S._states_and_territories_by_GDP}  and census.gov
}
\concept{Regression}
