% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gemstones}
\alias{dasl.gemstones}
\title{Gemstones}
\format{115 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gemstones/?sf_paged=18}{Gemstones}
}
\description{
Gemstones
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
minerals.usgs.gov/ds/2005/140/ USGS
}
\concept{Regression}
