% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gestation}
\alias{dasl.gestation}
\title{Gestation}
\format{18 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gestation/?sf_paged=18}{Gestation}
}
\description{
For humans, pregnancy lasts about 280 days. In other species of animals, the length of time from conception to birth varies. Is there any evidence that the gestation period is related to the animal’s life span? The data give Gestation Period (in days) and Life Expectancy (in years) for 18 species of mammals.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
