% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.global}
\alias{dasl.global}
\title{Global}
\format{4000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/global/?sf_paged=18}{Global}
}
\description{
Data for 800 respondents in each of five countries. The variables provide demographic information (sex, age, education, marital status) and responses to questions of interest to marketers on personal finance and purchasing.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
