% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.global500_2014}
\alias{dasl.global500_2014}
\title{Global500 2014}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/global500-2014/?sf_paged=18}{Global500 2014}
}
\description{
The dataset gives profits (in $M) for 30 of the 500 largest global corporations (as measured by revenue).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.gfmag.com/global-data/economic-data/largest-companies?page=3}
}
