% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.golf_courses}
\alias{dasl.golf_courses}
\title{Golf courses}
\format{45 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/golf-courses/?sf_paged=18}{Golf courses}
}
\description{
Golf courses
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
