% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.golf_drives_2015}
\alias{dasl.golf_drives_2015}
\title{Golf drives 2015}
\format{199 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/golf-drives-2015/?sf_paged=18}{Golf drives 2015}
}
\description{
The average drive distance (in yards) for 199 professional golfers during a week on the men’s PGA tour in 2015.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
PGA.com
}
\concept{Confidence Intervals for Means}
\concept{Data Display}
