% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.golfers_2017}
\alias{dasl.golfers_2017}
\title{Golfers 2017}
\format{190 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/golfers-2017/?sf_paged=18}{Golfers 2017}
}
\description{
Golfers 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.espn.com/golf/statistics/_/year/2017/}
}
\concept{Multiple Regression}
