% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.google_stock_prices}
\alias{dasl.google_stock_prices}
\title{Google stock prices}
\format{2226 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/google-stock-prices/?sf_paged=18}{Google stock prices}
}
\description{
Daily opening and closing stock prices (adjusted for splits and dividends) for Google, Inc. from Aug 19, 2004 through June 21, 2013
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Yahoo! Finance
}
\concept{Time Series Analysis}
