% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.grades}
\alias{dasl.grades}
\title{Grades}
\format{73 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/grades/?sf_paged=18}{Grades}
}
\description{
Grades
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data collected in an Introductory Statistics class.
}
\concept{Multiple Regression}
