% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.graduate_earnings}
\alias{dasl.graduate_earnings}
\title{Graduate Earnings}
\format{706 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/graduate-earnings/?sf_paged=18}{Graduate Earnings}
}
\description{
Is college worth the expense? Which colleges have graduates who earn the most? And what is the best predictor of earnings 5-years out? The data provide several possible predictors and background information suitable for building regression models.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://new.time.com/money/best-colleges/rankings/best-colleges/}
}
\concept{Regression}
