% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.graphite_production}
\alias{dasl.graphite_production}
\title{Graphite production}
\format{25 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/graphite-production/?sf_paged=18}{Graphite production}
}
\description{
A graphite manufacturer makes long rolls of flexible graphite to be used to seal components in combustion engines. The specifications state that the mean strength should be 21.2 ounces per square yard with a standard deviation of 0.29. Further specifications state that no roll should have strength less than 20.2 or more than 22.2 ounces per square yard. If there is a defect in terms of the strength of the graphite rolls, the seal will not hold. After the roll is created, a beta scanner takes readings of the basis weight in ounces per square yard. The data is separated into 10 lanes with 20 scans in each lane. A sample consists of one roll from each lane. The results from 20 samples follow are in the data.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented
}
\concept{Quality Control}
