% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.grocery_prices}
\alias{dasl.grocery_prices}
\title{Grocery prices}
\format{29 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/grocery-prices/?sf_paged=18}{Grocery prices}
}
\description{
WinCo Foods, a large discount grocery retailer in the western United States, promotes itself as the lowest priced grocery retailer. In newspaper ads WinCo Foods published a price comparison for products between WinCo and several competing grocery retailers. One of the retailers compared against WinCo was Walmart, also known as a low price competitor. WinCo selected a variety of products, listed the price of the product charges at each retailer, and showed the sales receipt to prove the prices at WinCo were the lowest in the area. A sample of the products and their price comparison at both WinCo and Walmart are given.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Comparing Two Groups}
