% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.gross_domestic_product}
\alias{dasl.gross_domestic_product}
\title{Gross domestic product}
\format{96 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/gross-domestic-product/?sf_paged=19}{Gross domestic product}
}
\description{
Gross domestic product
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Multiple Regression}
