% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hand_dexterity}
\alias{dasl.hand_dexterity}
\title{Hand dexterity}
\format{175 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hand-dexterity/?sf_paged=19}{Hand dexterity}
}
\description{
Physical therapists measure a patient’s manual dexterity with a simple task. The patient picks up small cylinders from a 4 * 4 frame with one hand, flips them over (still with one hand), and replaces them in the frame. The task is timed for all 16 cylinders. The tool was originally normed for adults. In a follow-up study, researchers used this tool to study how dexterity improves with age in children and establish norms against which to compare a patient’s dexterity.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Hand Dexterity in Children:  Administration and Normative Values of the Functional Dexterity Test (FDT),  Gloria R Gogola, MD, Paul F Velleman, PhD,  Shuai Xu, BS, MS, Adrianne M Morse, BA. Barbara Lacy BS, Dorit Aaron, MA OTR CHT FAOTA, J Hand Surg Am. 2013 Dec;38(12):2426-31. doi: 10.1016/jhsa.2013.08.123. Epub 2013 Nov 1
}
\concept{Multiple Regression}
\concept{Regression}
