% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hard_water}
\alias{dasl.hard_water}
\title{Hard water}
\format{61 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hard-water/?sf_paged=19}{Hard water}
}
\description{
In an investigation of environmental causes of disease, data were collected on the annual mortality rate (deaths per 100,000) for males in 61 large towns in England and Wales. In addition, the water hardness was recorded as the calcium concentration (parts per million, ppm) in the drinking water.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Comparing Two Groups}
\concept{Regression}
\concept{Test for Difference between Two Groups}
