% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.harry_potter_revenue}
\alias{dasl.harry_potter_revenue}
\title{Harry Potter revenue}
\format{17 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/harry-potter-revenue/?sf_paged=19}{Harry Potter revenue}
}
\description{
The movie Harry Potter and the Sorcerer’s Stone opened as a great success. But every movie sees declining revenue over time. The dataset gives the daily revenues for the movie during its first 17 days.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Time Series Analysis}
