% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hawaii_tourism}
\alias{dasl.hawaii_tourism}
\title{Hawaii tourism}
\format{60 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hawaii-tourism/?sf_paged=19}{Hawaii tourism}
}
\description{
Much of the public and private industry in Hawaii depends on tourism. The following time series plot shows the number of domestic visitors to Hawaii by air from the rest of the United States per month from January 2002 through December 2006 before the financial crisis of 2008.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Time Series Analysis}
