% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hdi_2015}
\alias{dasl.hdi_2015}
\title{HDI 2015}
\format{188 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hdi-2015/?sf_paged=19}{HDI 2015}
}
\description{
The United Nations Development Programme (UNDP) uses the Human Development Index (HDI) in an attempt to summarize in one number the progress in health, education, and economics of a country. In 2015, the HDI was as high as 0.94 for Norway and as low as 0.35 for Niger.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://hdr.undp.org/en/data#}
}
\concept{Regression}
