% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hdi_2016}
\alias{dasl.hdi_2016}
\title{HDI 2016}
\format{188 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hdi-2016/?sf_paged=19}{HDI 2016}
}
\description{
The United Nations Development Programme (UNDP) uses the Human Development Index (HDI) in an attempt to summarize in one number the progress in health, education, and economics of a country.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://hdr.undp.org/en/data#}  \url{http://hdr.undp.org/en/composite/HDI}
}
\concept{Multiple Regression}
\concept{Regression}
