% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.health_expenditures}
\alias{dasl.health_expenditures}
\title{Health expenditures}
\format{96 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/health-expenditures/?sf_paged=19}{Health expenditures}
}
\description{
Health expenditures
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Multiple Regression}
