% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.heart_attack_charges}
\alias{dasl.heart_attack_charges}
\title{Heart attack charges}
\format{12844 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/heart-attack-charges/?sf_paged=19}{Heart attack charges}
}
\description{
Number of days spent in hospital by patients admitted to hospitals in New York during one year with a primary diagnosis of acute myocardial infarction (heart attack). Data are from public medicare records. Consider the distribution of stays. The data also include the age and sex of the patient and the diagnostic (DRG) code.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data originally from Medicare, not citable
}
\concept{Data Display}
\concept{Time Series Analysis}
