% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.heart_attack_stays}
\alias{dasl.heart_attack_stays}
\title{Heart attack stays}
\format{5065 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/heart-attack-stays/?sf_paged=20}{Heart attack stays}
}
\description{
Number of days spent in hospital by female patients admitted to hospitals in New York during one year with a primary diagnosis of acute myocardial infarction (heart attack). Data are from public medicare records. Consider the distribution of stays. The data also include the age of the patient
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data originally from Medicare, not citable
}
