% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.heights_and_weights}
\alias{dasl.heights_and_weights}
\title{Heights and weights}
\format{80 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/heights-and-weights/?sf_paged=20}{Heights and weights}
}
\description{
The heights and weights of students in a statistics class were recorded.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Dave Bock personal communication
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
