% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.historical_oil_prices_2016}
\alias{dasl.historical_oil_prices_2016}
\title{Historical Oil Prices 2016}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/historical-oil-prices-2016/?sf_paged=20}{Historical Oil Prices 2016}
}
\description{
The price (per barrel) of oil has fluctuated over time. Various attempts to model it are generally not successful.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.eia.gov/dnav/pet/pet_pri_spt_s1_a.htm}
}
\concept{Regression}
