% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.holiday_shopping}
\alias{dasl.holiday_shopping}
\title{Holiday shopping}
\format{500 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/holiday-shopping/?sf_paged=20}{Holiday shopping}
}
\description{
Holiday shopping
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data adapted from real spending data from American Express
}
\concept{Confidence Intervals for Proportions}
