% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.holiday_spending}
\alias{dasl.holiday_spending}
\title{Holiday spending}
\format{750 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/holiday-spending/?sf_paged=20}{Holiday spending}
}
\description{
Holiday spending
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Data adapted from real spending data from American Express amex.reasonable=Amex.full[Amex.full$Dec_2004
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
