% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.home_depot_sales}
\alias{dasl.home_depot_sales}
\title{Home depot sales}
\format{40 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/home-depot-sales/?sf_paged=20}{Home depot sales}
}
\description{
Home depot sales
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
corporate reports
}
\concept{Nonparametric Methods}
\concept{Regression}
