% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.home_price_index_2017}
\alias{dasl.home_price_index_2017}
\title{Home Price Index 2017}
\format{366 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/home-price-index-2017/?sf_paged=20}{Home Price Index 2017}
}
\description{
Home Price Index 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://fred.stlouisfed.org}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
