% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hopkins_forest}
\alias{dasl.hopkins_forest}
\title{Hopkins Forest}
\format{365 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hopkins-forest/?sf_paged=20}{Hopkins Forest}
}
\description{
The Hopkins Memorial Forest is a 2500-acre reserve in Massachusetts, New York, and Vermont managed by the Williams College Center for Environmental Studies (CES). As part of its mission, the CES monitors forest resources and conditions over the long term.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
hmf.williams.edu/researchacademics/data/
}
\concept{Boxplots}
\concept{Correlation}
\concept{Multiple Regression}
\concept{Regression}
\concept{Scatterplot}
