% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hormones}
\alias{dasl.hormones}
\title{Hormones}
\format{17 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hormones/?sf_paged=20}{Hormones}
}
\description{
Canadian researcher John Coates took saliva samples in the morning, twice a day for eight days, from 17 men working on a London mid-size trading f loor (trading a wide range of assets, with largest exposure to German interest rate futures), in June 2005, and classified each trader according to whether his testosterone level was high or low on that day (compared with the trader’s median over the period). High testosterone days differed from trader to trader, and high days differed from low days on average by 25% in testosterone level. He also recorded the profits or losses (P&L) in pounds sterling of each trader during 11 am–4 pm daily.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Coates JM, Herbert J. PNAS 2008;105:6167–6172. Copyright © National Academy of Sciences. Reproduced by the permission of National Academy of Sciences.)
}
\concept{Nonparametric Methods}
