% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hot_dogs}
\alias{dasl.hot_dogs}
\title{Hot dogs}
\format{13 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hot-dogs/?sf_paged=20}{Hot dogs}
}
\description{
Facts about hot dogs as reported by Consumer Reports. How do they relate to each other?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Consumer Reports
}
\concept{Regression}
