% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hotel_maids}
\alias{dasl.hotel_maids}
\title{Hotel maids}
\format{75 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hotel-maids/?sf_paged=20}{Hotel maids}
}
\description{
A Harvard psychologist recruited 75 female hotel maids to participate in a study. She randomly selected 41 of them, whom she informed (truthfully) that the work they do satisfies the Surgeon General’s recommendations for an active lifestyle, providing examples to show that their work is good exercise. The other 34 were told nothing. Various characteristics, such as weight, body fat, body mass index and blood pressure were recorded at the start of the study and again after four weeks. The researcher was interested in whether the information she provided would result in measurable physical changes. If there is a difference, it might challenge our understanding of the placebo effect because being informed could make a difference.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Crum, Alia J., and Ellen J. Langer. 2007. Mind-set matters: Exercise and the placebo effect. Psychological Science 18, no. 2: 165-171. Published Version doi:10.1111/j.1467-9280.2007.01867.x Citable Link \url{http://nrs.harvard.edu/urn-3:HUL.InstRepos:3196007}
}
