% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.housing_prices}
\alias{dasl.housing_prices}
\title{Housing prices}
\format{1057 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/housing-prices/?sf_paged=20}{Housing prices}
}
\description{
House prices and properties in New York. What properties of a house can predict its price? Can we use such a model to identify houses that are extraordinarily expensive or inexpensive?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
random sample of 1057 houses taken from full Saratoga Housing Data (De Veaux)
}
\concept{Correlation}
\concept{Multiple Regression}
\concept{Regression}
\concept{Scatterplot}
