% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.housing_prices_ge19}
\alias{dasl.housing_prices_ge19}
\title{Housing prices GE19}
\format{1734 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/housing-prices-ge19/?sf_paged=21}{Housing prices GE19}
}
\description{
House prices and properties in New York. What properties of a house can predict its price? Can we use such a model to identify houses that are extraordinarily expensive or inexpensive?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
random sample of 1734 houses taken from full Saratoga Housing Data (De Veaux)
}
\concept{Multiple Regression}
