% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.how_old_is_that_tree}
\alias{dasl.how_old_is_that_tree}
\title{How old is that Tree}
\format{27 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/how-old-is-that-tree/?sf_paged=21}{How old is that Tree}
}
\description{
One can determine how old a tree is by counting its rings, but that requires either cutting the tree down or extracting a sample from the tree’s core. Can we estimate the tree’s age simply from its  diameter?A forester measured 27 trees of the same species that had been cut down, and counted the rings to determine the ages of the trees.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
