% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hurricane_frequencies}
\alias{dasl.hurricane_frequencies}
\title{Hurricane frequencies}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hurricane-frequencies/?sf_paged=21}{Hurricane frequencies}
}
\description{
Hurricane frequencies
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.aoml.noaa.gov/hrd/hurdat/All_U.S._Hurricanes.html}
}
\concept{Chi Square Tests}
