% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hurricane_history}
\alias{dasl.hurricane_history}
\title{Hurricane history}
\format{168 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hurricane-history/?sf_paged=21}{Hurricane history}
}
\description{
Hurricane history
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/Atlantic_hurricane_season#2000s_2}
}
\concept{Comparing Two Groups}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
