% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.hurricanes_2015}
\alias{dasl.hurricanes_2015}
\title{Hurricanes 2015}
\format{226 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/hurricanes-2015/?sf_paged=21}{Hurricanes 2015}
}
\description{
The barometric pressure at the center of a hurricane is often used to measure the strength of the hurricane because it can predict the maximum wind speed of the storm. How well is the wind speed predicted by the barometric pressure?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.aoml.noaa.gov/hrd/hurdat/All_U.S._Hurricanes.html}
}
\concept{Multiple Regression}
\concept{Regression}
