% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.igf}
\alias{dasl.igf}
\title{Igf}
\format{1334 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/igf/?sf_paged=21}{Igf}
}
\description{
The data hold measurements on people of various ages. The main variable of interest is the level of insulin-like growth factor (igƒ) (J. Clin. Endocrinol. Metab. 78(3): 744–752, March 1994). Each row in the data set corresponds to one individual. See also Igf13, which concentrates on children.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
J. Clin. Endocrinol. Metab. 78(3) 744-752, March, 1994
}
\concept{Multiple Regression}
