% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.igf13}
\alias{dasl.igf13}
\title{Igf13}
\format{710 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/igf13/?sf_paged=21}{Igf13}
}
\description{
Measurements on children under 13 years of age. Most of the data was collected from physical examinations in schools. The main variable of interest is the level of insulin-like growth factor (igƒ) (J. Clin. Endocrinol. Metab. 78(3): 744–752, March 1994). Each row in the data set corresponds to one individual. See also the dataset Igf, which includes adults.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
J. Clin. Endocrinol. Metab. 78(3) 744-752, March, 1994
}
\concept{Multiple Regression}
