% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.illiad_injuries}
\alias{dasl.illiad_injuries}
\title{Illiad Injuries}
\format{184 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/illiad-injuries/?sf_paged=21}{Illiad Injuries}
}
\description{
Homer’s Iliad is an epic poem, compiled around 800 BCE, that describes several weeks of the last year of the 10-year siege of Troy (Ilion) by the Achaeans. The story centers on the rage of the great warrior Achilles. But it includes many details of injuries and outcomes, and is thus the oldest record of Greek medicine. The data report 146 recorded injuries for which both injury site and outcome are provided in the Illiad. Are some kinds of injuries more lethal than others?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
J Hand Surg Am. 2013 Sep;38(9):1790-3.  10.1016/j.jhsa.2013.06.022. Epub 2013
}
\concept{Chi Square Tests}
