% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.illiad_weapons}
\alias{dasl.illiad_weapons}
\title{Illiad weapons}
\format{184 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/illiad-weapons/?sf_paged=21}{Illiad weapons}
}
\description{
Homer’s Iliad is an epic poem, compiled around 800 BCE, that describes several weeks of the last year of the 10-year siege of Troy (Ilion) by the Achaeans. The story centers on the rage of the great warrior Achilles. But it includes many details of injuries and outcomes, and is thus the oldest record of Greek medicine. The data report 152 recorded injuries for which both injury site and weapon are provided in the Illiad.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
J Hand Surg Am. 2013 Sep;38(9):1790-3.  10.1016/j.jhsa.2013.06.022. Epub 2013
}
\concept{Chi Square Tests}
