% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.income_and_housing}
\alias{dasl.income_and_housing}
\title{Income and housing}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/income-and-housing/?sf_paged=21}{Income and housing}
}
\description{
How are housing costs related to median family income?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Office of Federal Housing Enterprise Oversight
}
\concept{Correlation}
\concept{Nonparametric Methods}
\concept{Regression}
\concept{Scatterplot}
