% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.income_and_internet}
\alias{dasl.income_and_internet}
\title{Income and Internet}
\format{5 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/income-and-internet/?sf_paged=21}{Income and Internet}
}
\description{
Income and Internet
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
pewresearch.org
}
\concept{Data Displays}
