% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.income_and_pets}
\alias{dasl.income_and_pets}
\title{Income and pets}
\format{5 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/income-and-pets/?sf_paged=21}{Income and pets}
}
\description{
Income and pets
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Source: American Veterinary Medical Association, Schaumburg, IL, 	 U.S. Pet Ownership and Demographics Sourcebook, 1993 and 1997	 \url{http://www.avma.org}
}
\concept{Data Displays}
