% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.income_vs_hours_2013}
\alias{dasl.income_vs_hours_2013}
\title{Income vs Hours 2013}
\format{13 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/income-vs-hours-2013/?sf_paged=22}{Income vs Hours 2013}
}
\description{
Income vs Hours 2013
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://dqydj.com/work-harder-earn-more-2013-incomes-by-hours-worked/}
}
\concept{Regression}
