% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.industrial_experiment}
\alias{dasl.industrial_experiment}
\title{Industrial experiment}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/industrial-experiment/?sf_paged=22}{Industrial experiment}
}
\description{
Industrial experiment
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Made up
}
\concept{Boxplots}
