% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.indy_500_2016}
\alias{dasl.indy_500_2016}
\title{Indy 500 2016}
\format{100 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/indy-500-2016/?sf_paged=22}{Indy 500 2016}
}
\description{
Indy 500 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Indianapolis_500_winners}
}
