% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.indy_500_2017}
\alias{dasl.indy_500_2017}
\title{Indy 500 2017}
\format{101 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/indy-500-2017/?sf_paged=22}{Indy 500 2017}
}
\description{
The 2.5-mile Indianapolis Motor Speedway has been the home to a race on Memorial Day nearly every year since 1911. Even during the first race, there were controversies. Ralph Mulford was given the checkered flag first but took three extra laps just to make sure he’d completed 500 miles. When he finished, another driver, Ray Harroun, was being presented with the winner’s trophy, and Mulford’s protests were ignored. Harroun averaged 74.6 mph for the 500 miles. In 2013, the winner, Tony Kanaan, averaged over 187 mph, beating the previous record by over 17 mph!
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Indianapolis_500_winners}
}
