% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.inflation_2016}
\alias{dasl.inflation_2016}
\title{Inflation 2016}
\format{21 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/inflation-2016/?sf_paged=22}{Inflation 2016}
}
\description{
The U.S. Consumer Price Index and year, every 5 years since 1916. These are the values for January of each year. What is the trend? Can we model it with a linear regression?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://inflationdata.com/Inflation/Consumer_Price_Index/HistoricalCPI.aspx?reloaded=true#Table}
}
\concept{Regression}
