% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.insurance_life_expectancy}
\alias{dasl.insurance_life_expectancy}
\title{Insurance life expectancy}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/insurance-life-expectancy/?sf_paged=22}{Insurance life expectancy}
}
\description{
Insurance life expectancy
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Multiple Regression}
