% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.insurance_policies}
\alias{dasl.insurance_policies}
\title{Insurance policies}
\format{200 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/insurance-policies/?sf_paged=22}{Insurance policies}
}
\description{
Insurance policies
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
