% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.insurance_profits}
\alias{dasl.insurance_profits}
\title{Insurance profits}
\format{98 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/insurance-profits/?sf_paged=22}{Insurance profits}
}
\description{
Insurance profits
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
