% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.insurance_profits_sales_rep}
\alias{dasl.insurance_profits_sales_rep}
\title{Insurance profits sales rep}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/insurance-profits-sales-rep/?sf_paged=22}{Insurance profits sales rep}
}
\description{
Insurance profits sales rep
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
random sample 30 insurance policy profits
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
