% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.interest_rates_2009}
\alias{dasl.interest_rates_2009}
\title{Interest rates 2009}
\format{62 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/interest-rates-2009/?sf_paged=22}{Interest rates 2009}
}
\description{
Average annual interest rates (banks prime lending) in the United States from 1966 through 2009
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Time Series Analysis}
