% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.interest_rates_and_mortgages_2015}
\alias{dasl.interest_rates_and_mortgages_2015}
\title{Interest rates and mortgages 2015}
\format{31 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/interest-rates-and-mortgages-2015/?sf_paged=23}{Interest rates and mortgages 2015}
}
\description{
The amount charged for mortgages may be related to the total value of mortgage loans in the US. Can that relationship be modeled? Does it depend as well on the year? Consider a rotating plot of interest rate, mortgage total, and year.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.freddiemac.com/pmms/pmms30.htm and www.federalreserve.gov/econresdata/releases/mortoutstand/current.htm \url{https://www.federalreserve.gov/econresdata/releases/mortoutstand/current.htm}
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
