% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.internet_users_2014}
\alias{dasl.internet_users_2014}
\title{Internet users 2014}
\format{70 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/internet-users-2014/?sf_paged=23}{Internet users 2014}
}
\description{
Internet users 2014
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://mdgs.un.org/unsd/mdg/Data.aspx}
}
\concept{Regression}
