% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ipod_failures}
\alias{dasl.ipod_failures}
\title{iPod failures}
\format{17 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ipod-failures/?sf_paged=23}{iPod failures}
}
\description{
iPod failures
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.macintouch.com/reliability/ipodfailures.html
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
