% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.iq_brain}
\alias{dasl.iq_brain}
\title{IQ Brain}
\format{40 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/iq-brain/?sf_paged=23}{IQ Brain}
}
\description{
A study examined brain size (measured as pixels counted in a digitized magnetic resonance image [MRI] of a cross section of the brain) and IQ (4 performance scales of the Wechsler IQ test) for college students. The data give the Performance IQ scores and Brain Size.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Correlation}
\concept{Scatterplot}
