% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.jcrew}
\alias{dasl.jcrew}
\title{JCrew}
\format{41 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/jcrew/?sf_paged=23}{JCrew}
}
\description{
J. Crew is a clothing company known for its preppy fashions, including jeans, khakis, and other basic items sold to young professionals through its catalogs, websites, and some 260 retail and outlet stores in the United States. (Michelle Obama shops there.) The data reports their quarterly revenue from Q1 2003 through the first quarter of 2013
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Company quarterly reports
}
\concept{Time Series Analysis}
