% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.job_growth}
\alias{dasl.job_growth}
\title{Job Growth}
\format{69 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/job-growth/?sf_paged=23}{Job Growth}
}
\description{
Job Growth
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown… broken link
}
\concept{Regression}
