% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.job_satisfaction}
\alias{dasl.job_satisfaction}
\title{Job satisfaction}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/job-satisfaction/?sf_paged=23}{Job satisfaction}
}
\description{
Job satisfaction
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Comparing Two Groups}
\concept{Nonparametric Methods}
\concept{Paired Data}
\concept{Test for Difference between Two Groups}
