% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.kentucky_derby_2016}
\alias{dasl.kentucky_derby_2016}
\title{Kentucky Derby 2016}
\format{143 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/kentucky-derby-2016/?sf_paged=23}{Kentucky Derby 2016}
}
\description{
Kentucky Derby 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.kentuckyderby.ag/kentuckyderby-results.php
}
\concept{Boxplots}
\concept{Correlation}
\concept{Scatterplot}
