% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.kentucky_derby_2017}
\alias{dasl.kentucky_derby_2017}
\title{Kentucky Derby 2017}
\format{143 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/kentucky-derby-2017/?sf_paged=23}{Kentucky Derby 2017}
}
\description{
The Kentucky Derby is a horse race that has been run every year since 1875 at Churchill Downs in Louisville, Kentucky. The race started as a 1.5-mile race, but in 1896, it was shortened to 1.25 miles because experts felt that 3-year-old horses shouldn’t run such a long race that early in the season. (It has been run in May every year but one—1901—when it took place on April 29.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.kentuckyderby.ag/kentuckyderby-results.php
}
\concept{Boxplots}
\concept{Displaying Quantitative Data}
\concept{Scatterplot}
\concept{Summarizing Quantitative Data}
