% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.komtek_technologies}
\alias{dasl.komtek_technologies}
\title{Komtek Technologies}
\format{49 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/komtek-technologies/?sf_paged=23}{Komtek Technologies}
}
\description{
Komtek Technologies
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
This example is based on 1998 case study written by J. Hunt, E. Landry, and J. Rao as part of the Babson College case series. The data and setting used in this example are based on the actual case study, but the data have been modified and the conclusions are fictitious.
}
\concept{Quality Control}
