% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.la_rainfall}
\alias{dasl.la_rainfall}
\title{LA rainfall}
\format{28 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/la-rainfall/?sf_paged=23}{LA rainfall}
}
\description{
The Los Angeles Almanac reports a number of variables about the weather in LA. Among them is the annual rainfall, reported here for 1991-2018. It is worthwhile to look up any outliers.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.laalmanac.com/weather/we13.php}
}
\concept{Regression}
\concept{Time Series Analysis}
