% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.late_arrivals_2016}
\alias{dasl.late_arrivals_2016}
\title{Late arrivals 2016}
\format{263 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/late-arrivals-2016/?sf_paged=24}{Late arrivals 2016}
}
\description{
The data report the percentage of flights that were late and the percentage that departed on time for each month from 1995 through early 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
U.S. Bureau of Transportation Statistics
}
\concept{Confidence Intervals for Means}
