% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.legionnaires_disease}
\alias{dasl.legionnaires_disease}
\title{Legionnaires disease}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/legionnaires-disease/?sf_paged=24}{Legionnaires disease}
}
\description{
In 1974, the Bellevue-Stratford Hotel in Philadelphia was the scene of an outbreak of what later became known as legionnaires’ disease. The cause of the disease was finally discovered to be bacteria that thrived in the air-conditioning units of the hotel. Owners of the Rip Van Winkle Motel, hearing about the Bellevue-Stratford, replace their air-conditioning system. The data are the bacteria counts in the air of eight rooms, before and after a new air-conditioning system was installed (measured in colonies per cubic foot of air). Has the new system has succeeded in lowering the bacterial count?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
