% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.libertys_nose}
\alias{dasl.libertys_nose}
\title{Libertys nose}
\format{18 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/libertys-nose/?sf_paged=24}{Libertys nose}
}
\description{
Is the Statue of Liberty’s nose too long? Her nose measures 4′6″, but she is a large statue, after all. Her arm is 42 feet long. That means her arm is 42/4.5 = 9.3 times as long as her nose. Is that a reasonable ratio? The data give arm and nose lengths of 18 girls in a statistics class, and the ratio of arm-to-nose length for each.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
