% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.life_expectancy}
\alias{dasl.life_expectancy}
\title{Life Expectancy}
\format{195 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/life-expectancy/?sf_paged=24}{Life Expectancy}
}
\description{
Life expectancy at birth in 195 countries.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
CIA Factbook
}
\concept{Data Display}
