% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.life_expectancy_us}
\alias{dasl.life_expectancy_us}
\title{Life expectancy US}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/life-expectancy-us/?sf_paged=24}{Life expectancy US}
}
\description{
Here is a table from the National Vital Statistics Report that gives the Life Expectancy for white males in the United States every decade during the 20th century (1 = 1900 to 1910, 2 = 1911 to 1920, etc.). Does a linear model relating life expectancy to decade fit? Would re-expressing either variable help?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
National Vital Statistics Report
}
\concept{Regression}
