% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.little_league_product_testing}
\alias{dasl.little_league_product_testing}
\title{Little league product testing}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/little-league-product-testing/?sf_paged=24}{Little league product testing}
}
\description{
Little league product testing
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
