% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.load_factors_2016}
\alias{dasl.load_factors_2016}
\title{Load factors 2016}
\format{197 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/load-factors-2016/?sf_paged=24}{Load factors 2016}
}
\description{
Load factors 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.transtats.bts.gov/Data_Elements.aspx?Data=2}
}
\concept{Boxplots}
