% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.load_factors_2017}
\alias{dasl.load_factors_2017}
\title{Load factors 2017}
\format{183 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/load-factors-2017/?sf_paged=24}{Load factors 2017}
}
\description{
The Research and Innovative Technology Administration of the Bureau of Transportation Statistics reports load factors (passenger-miles as a percentage of available seat miles) for commercial airlines for every month from October 2002 through 2017 for both domestic and international flights.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.transtats.bts.gov/Data_Elements.aspx?Data=2}
}
\concept{Boxplots}
