% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.lobsters_2016}
\alias{dasl.lobsters_2016}
\title{Lobsters 2016}
\format{67 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/lobsters-2016/?sf_paged=24}{Lobsters 2016}
}
\description{
The Maine lobster fishing industry is carefully controlled and licensed, and facts about it have been recorded for more than a century, so it is an important industry that we can examine in detail. The dataset holds annual data
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.maine.gov/dmr/commercial-fishing/landings/documents/lobster.table.pdf}
}
\concept{Multiple Regression}
\concept{Regression}
