% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.lottery_numbers}
\alias{dasl.lottery_numbers}
\title{Lottery numbers}
\format{49 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/lottery-numbers/?sf_paged=24}{Lottery numbers}
}
\description{
Lottery numbers
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.nationallottery.co.za
}
\concept{Chi Square Tests}
