% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.loyalty_program}
\alias{dasl.loyalty_program}
\title{Loyalty program}
\format{10000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/loyalty-program/?sf_paged=24}{Loyalty program}
}
\description{
Loyalty program
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Chi Square Tests}
