% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.magnet_schools}
\alias{dasl.magnet_schools}
\title{Magnet schools}
\format{1755 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/magnet-schools/?sf_paged=25}{Magnet schools}
}
\description{
Magnet schools
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Chance magazine Winter 2003
}
\concept{Tables}
