% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mail}
\alias{dasl.mail}
\title{Mail}
\format{36 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mail/?sf_paged=25}{Mail}
}
\description{
The dataset gives the number of pieces of mail received at a school office for 36 days.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
