% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.major_hurricanes_2013}
\alias{dasl.major_hurricanes_2013}
\title{Major hurricanes 2013}
\format{70 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/major-hurricanes-2013/?sf_paged=25}{Major hurricanes 2013}
}
\description{
Is global climate change leading to an increase in the number of major hurricanes? The data gives the number of hurricanes classified as major hurricanes in the Atlantic Ocean each year from 1944 through 2013, as reported by NOAA:
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.nhc.noaa.gov/data/
}
\concept{Data Display}
