% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.male_unemployment_2017}
\alias{dasl.male_unemployment_2017}
\title{Male unemployment 2017}
\format{222 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/male-unemployment-2017/?sf_paged=25}{Male unemployment 2017}
}
\description{
Male unemployment 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.worldbank.org/indicator/SL.UEM.TOTL.MA.ZS}
}
\concept{Regression}
