% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.manatees_2015}
\alias{dasl.manatees_2015}
\title{Manatees 2015}
\format{33 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/manatees-2015/?sf_paged=25}{Manatees 2015}
}
\description{
Manatees are gentle mammals that live in the waters off the coast of Florida and a few other places. Unfortunately, many are killed each year in collisions with powerboats. Marine biologists warn that the growing number of powerboats registered in Florida threatens the existence of manatees. The data here are the number of manatees killed each year since 1982 and the number of powerboats registered in Florida (in thousands) for those years. Is there a relationship?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Florida Fish and Wildlife Conservation Commission (myfwc.com/research/manatee/) and  the U.S. Coast Guard Office of Auxiliary and Boating Safety (\url{https://www.uscgboating.org/library/accident-statistics/Recreational-Boating-Statistics-2015.pdf)}
}
