% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.marijuana_2015}
\alias{dasl.marijuana_2015}
\title{Marijuana 2015}
\format{56 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/marijuana-2015/?sf_paged=25}{Marijuana 2015}
}
\description{
In 2015 the Council of Europe published a report entitled The European School Survey Project on Alcohol and Other Drugs (www.espad.org). Among other issues, the survey investigated the percent-ages of 16-year-olds who had used marijuana. The data are the results for 38 European countries.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
The European School Survey project on Alcohol and Other Drugs (www.espad.org)
}
\concept{Data Display}
