% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.market_value}
\alias{dasl.market_value}
\title{Market value}
\format{45 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/market-value/?sf_paged=25}{Market value}
}
\description{
Real estate agents want to set correctly the price of a house that’s about to go on the real estate market. They must choose a price that strikes a balance between one that is so high that the house takes too long to sell and one that’s so low that not enough value will go to the homeowner. One appraisal method is the “Comparative Market Analysis” approach by which the market value of a house is based on recent sales of similar homes in the neighborhood. Because no two houses are exactly the same, appraisers have to adjust comparable homes for such features as extra square footage, bedrooms, fireplaces, upgrading, parking facilities, swimming pool, lot size, location, and so on. The appraised market values and the selling prices of 45 homes from the same region are given.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Comparing Two Groups}
