% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.marketing_managers_salaries}
\alias{dasl.marketing_managers_salaries}
\title{Marketing managers salaries}
\format{25 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/marketing-managers-salaries/?sf_paged=25}{Marketing managers salaries}
}
\description{
Marketing managers salaries
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
data.bls.gov/oes
}
\concept{Analysis of Variance}
