% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.marriage_age_2016}
\alias{dasl.marriage_age_2016}
\title{Marriage age 2016}
\format{76 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/marriage-age-2016/?sf_paged=25}{Marriage age 2016}
}
\description{
Age at first marriage has changed over the course of the past century. In addition, the difference in the age of the husband and of the wife at first marriage has changed. Both the ages and the difference in ages can be interesting to analyze.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://factfinder.census.gov/faces/tableservices/jsf/pages/productview.xhtml?pid=ACS_16_1YR_B12007&prodType=table}  \url{http://www.census.gov/population/socdemo/hh-fam/tabMS-2.pdf}  Notes: 2005 to 2015 revised in top link
}
\concept{Boxplots}
