% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.marriage_age_2017}
\alias{dasl.marriage_age_2017}
\title{Marriage age 2017}
\format{78 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/marriage-age-2017/?sf_paged=25}{Marriage age 2017}
}
\description{
The estimated median age at fist marriage by sex from 1890 to 2017 is provided by the U.S. Census bureau. Since 1960, marriage ages have been increasing steadily. Has the difference between men’s and women’s first marriage age changed?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.census.gov/data/tables/time-series/demo/families/marital.html}
}
\concept{Regression}
