% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.math_scores_2013}
\alias{dasl.math_scores_2013}
\title{Math scores 2013}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/math-scores-2013/?sf_paged=25}{Math scores 2013}
}
\description{
The National Center for Education Statistic reports average mathematics achievement scores for eighth  graders in all 50 states.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Nces.ed.gov/nationsreportcard/
}
\concept{Data Display}
