% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mazes_and_smells}
\alias{dasl.mazes_and_smells}
\title{Mazes and smells}
\format{21 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mazes-and-smells/?sf_paged=26}{Mazes and smells}
}
\description{
Can pleasant smells improve learning? Researchers timed 21 subjects as they tried to complete paper-and-pencil mazes. Each subject attempted a maze both with and without the presence of a floral aroma. Subjects were randomized with respect to whether they did the scented trial first or second. Is there any  evidence that the floral scent improved the subjects’ ability to complete the mazes?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
A. R. Hirsch and L. H. Johnston, “Odors and Learning.” Chicago: Smell and Taste Treatment and Research Foundation
}
