% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.memory}
\alias{dasl.memory}
\title{Memory}
\format{104 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/memory/?sf_paged=26}{Memory}
}
\description{
In an experiment to test ginkgo bloba, subjects were assigned randomly to take ginkgo biloba supplements or a placebo. Their memory was tested to see whether it improved.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Solomon,P.R., Adams,F.,  Silver, A., Zimmer, J. and De Veaux, R.D.  Journal of the American Medical Assocation Vol. 288 pp.835-840 (2002), “Ginkgo for Memory Enhancement A Randomized Controlled Trial”
}
\concept{Comparing Two Groups}
\concept{Test for Difference between Two Groups}
