% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mens_weekly_earnings_2013}
\alias{dasl.mens_weekly_earnings_2013}
\title{Men&#8217;s weekly earnings 2013}
\format{41 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mens-weekly-earnings-2013/?sf_paged=26}{Men&#8217;s weekly earnings 2013}
}
\description{
Quarterly median weekly earnings from the first quarter of 2003 through the first quarter of 2013 for men, 25 years of age or older, in the United States
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
U.S. Bureau of Labor Statistics; www.bls.gov
}
\concept{Time Series Analysis}
