% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mens_weights}
\alias{dasl.mens_weights}
\title{Mens Weights}
\format{80 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mens-weights/?sf_paged=26}{Mens Weights}
}
\description{
Since the 1960s, the Centers for Disease Control and Prevention’s National Center for Health Statistics has been collecting health and nutritional information on people of all ages and backgrounds. The National Health and Nutrition Examination Survey (NHANES) of 2001–2002, measured a wide variety of variables, including body measurements, cardiovascular fitness, blood chemistry, and demographic information on more than 11,000 individuals. The file holds data on the weights of 80 men between 19 and 24 years old of average height (between 5′8″ and 5′10″ tall).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.cdc.gov/nchs/nhanes.htm
}
\concept{Normal Probability Plots}
