% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.michelson}
\alias{dasl.michelson}
\title{Michelson}
\format{100 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/michelson/?sf_paged=26}{Michelson}
}
\description{
In 1879, A. A. Michelson made 100 determinations of the velocity of light in air using a modification of a method proposed by the French physicist Foucault.  The data are given here as reported by Stigler. The measurements are derived from sets of often widely disparate numbers of observations.  The numbers are in km/sec, and have had 299,000 subtracted from them.  The currently accepted “true” velocity of light in vacuum is 299,792.5 km/sec.  Stigler has applied the corrections used by Michelson and reports that the “true” value appropriate for comparison to these measurements is 734.5. Each trial may be a summary of several experimental observations.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Stigler, Stephen M. (1977).  Do Robust Estimators Work with Real Data? The Annals of Statistics  5:4, 1075.  Used by permission.
}
\concept{Confidence Intervals for Means}
