% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.midterms}
\alias{dasl.midterms}
\title{Midterms}
\format{64 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/midterms/?sf_paged=26}{Midterms}
}
\description{
Midterms
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
personal communication
}
\concept{Regression}
