% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.minimum_wage_workers}
\alias{dasl.minimum_wage_workers}
\title{Minimum Wage Workers}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/minimum-wage-workers/?sf_paged=26}{Minimum Wage Workers}
}
\description{
Minimum Wage Workers
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.bls.com
}
\concept{Data Displays}
\concept{Tables}
