% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.montana}
\alias{dasl.montana}
\title{Montana}
\format{202 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/montana/?sf_paged=26}{Montana}
}
\description{
Montana
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
