% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.montana_revisited}
\alias{dasl.montana_revisited}
\title{Montana revisited}
\format{202 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/montana-revisited/?sf_paged=26}{Montana revisited}
}
\description{
Montana revisited
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
