% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mothers_and_fathers_1965_2011}
\alias{dasl.mothers_and_fathers_1965_2011}
\title{Mothers and fathers 1965-2011}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mothers-and-fathers-1965-2011/?sf_paged=26}{Mothers and fathers 1965-2011}
}
\description{
Mothers and fathers 1965-2011
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.pewsocialtrends.org/2013/03/14/modern-parenthood-roles-of-moms-and-dads-converge-as-they-balance-work-and-family/}
}
\concept{Tables}
