% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mothers_and_fathers_aspirations}
\alias{dasl.mothers_and_fathers_aspirations}
\title{Mothers and fathers aspirations}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mothers-and-fathers-aspirations/?sf_paged=26}{Mothers and fathers aspirations}
}
\description{
The New York Times combined survey data (economix.blogs.nytimes.com/2013/ 07/10/working-parents-wanting-fewer-hours/) with data from the U.S. Bureau of Labor Statistics (BLS) (www.bls.gov/news .release/archives/famee_04262013.htm) comparing how mothers and fathers would like to allocate their time compared with what they actually do. They asked a sample of parents with children 18 or under: “If money were no object, and you were free to do whatever you wanted, would you stay at home, would you work full time, or would you work part time?” Percent of respondents to this question choosing each alternative are reported in the “Desire” columns of the table. Data in the “Actual” column are from the BLS. (Note: “Unemployed” = unemployed and actively seeking work.) The table reports column percents (which may not add to 100% due to rounding)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.pewsocialtrends.org/2013/03/14/modern-parenthood-roles-of-moms-and-dads-converge-as-they-balance-work-and-family/}  and \url{http://www.bls.gov/news.release/archives/}  famee_04262013.htm
}
\concept{Tables}
