% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mouth_volume}
\alias{dasl.mouth_volume}
\title{Mouth volume}
\format{61 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mouth-volume/?sf_paged=27}{Mouth volume}
}
\description{
Researchers in Food Science studied how big people’s mouths tend to be. They measured mouth volume by pouring water into the mouths of subjects who lay on their backs. Unless this is your idea of a good time, it would be helpful to have a model to estimate mouth volume more simply. Fortunately, mouth volume is related to height. (Mouth volume is measured in cubic centimeters and height in meters.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Private communication
}
\concept{Regression}
