% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.movie_admissions_2016}
\alias{dasl.movie_admissions_2016}
\title{Movie Admissions 2016}
\format{5 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/movie-admissions-2016/?sf_paged=27}{Movie Admissions 2016}
}
\description{
Movie Admissions 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.mpaa.org/wp-content/uploads/2017/03/MPAA-Theatrical-Market-Statistics-2016_Final-1.pdf}
}
