% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.movie_budgets}
\alias{dasl.movie_budgets}
\title{Movie budgets}
\format{120 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/movie-budgets/?sf_paged=27}{Movie budgets}
}
\description{
Movie budgets
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Private communication
}
\concept{Regression}
