% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.movie_lengths_2010}
\alias{dasl.movie_lengths_2010}
\title{Movie lengths 2010}
\format{150 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/movie-lengths-2010/?sf_paged=27}{Movie lengths 2010}
}
\description{
Movie lengths 2010
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Picked up from previous editions
}
\concept{Data Display}
