% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.movie_profits}
\alias{dasl.movie_profits}
\title{Movie profits}
\format{609 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/movie-profits/?sf_paged=27}{Movie profits}
}
\description{
Does money purchase a good movie? Is the US Gross revenue related to either the budge or the Rotten Tomatoes score? The dataset holds data on 609 recent releases that includes the USGross (in $M), the Budget ($M), the Run Time (minutes), and the score given by the critics on the Rotten Tomatoes website.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
combined from IMDB.com  Boxofficemojo.com Rottentomatoes.com
}
\concept{Multiple Regression}
