% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mozart}
\alias{dasl.mozart}
\title{Mozart}
\format{125 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mozart/?sf_paged=27}{Mozart}
}
\description{
In a study published in the journal Psychological Science, Rauscher, Shaw, and Ky reported that when students were given a spatial reasoning section of a standard IQ test, those who listened to Mozart for 10 minutes improved their scores more than those who simply sat quietly.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Rauscher FH, Shaw GL, Ky KN. Listening to Mozart enhances spatial-temporal reasoning: toward a neurophysiological basis. Neurosci Lett 1995;185: 44-7
}
\concept{Confidence Intervals for Means}
\concept{Quality Control}
