% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mpg_2016}
\alias{dasl.mpg_2016}
\title{MPG 2016}
\format{119 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mpg-2016/?sf_paged=27}{MPG 2016}
}
\description{
MPG 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.fueleconomy.gov/feg/download.shtml}
}
\concept{Boxplots}
