% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.music_library}
\alias{dasl.music_library}
\title{Music library}
\format{192 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/music-library/?sf_paged=27}{Music library}
}
\description{
Students in an introductory statistics course were asked how many songs they had in their digital music library.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Class survey De Veaux Stat 201
}
\concept{Boxplots}
