% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mutual_fund_flows}
\alias{dasl.mutual_fund_flows}
\title{Mutual fund flows}
\format{154 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mutual-fund-flows/?sf_paged=27}{Mutual fund flows}
}
\description{
Mutual fund flows
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
