% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mutual_fund_returns_2017}
\alias{dasl.mutual_fund_returns_2017}
\title{Mutual fund returns 2017}
\format{98 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mutual-fund-returns-2017/?sf_paged=27}{Mutual fund returns 2017}
}
\description{
Mutual fund returns for a sample of mutual funds sampled in 2017. These are a good source of data for paired methods.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://news.morningstar.com/fund-category-returns/}
}
\concept{Comparing Two Groups}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
