% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.mutual_funds_2016}
\alias{dasl.mutual_funds_2016}
\title{Mutual funds 2016}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/mutual-funds-2016/?sf_paged=27}{Mutual funds 2016}
}
\description{
On December 30, 2016, the Standard and Poor’s (S&P) 500 index hit an all-time high.  During 2016, the S&P returned 12.25%.  Here is a histogram of the 2016 net returns (total return – annual expenses) for Money Magazine’s 50 Best Mutual Funds and ETFs. The net returns are computed from the data given by Money Magazine.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://time.com/money/4616747/best-mutual-funds-etfs-money-50/}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
