% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nail_polish}
\alias{dasl.nail_polish}
\title{Nail polish}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nail-polish/?sf_paged=27}{Nail polish}
}
\description{
A student, preparing for a triathlon, suspected that the 45 minutes each day she spent training in a chlorinated pool was damaging her nail polish. She wished to investigate whether the color of the nail polish might make a difference. She mounted acrylic nails on sticks and polished them with two different color nail polishes. She soaked them together in a chlorine solution equivalent to a swimming pool’s chlorination and then tapped them 100 times on a computer keyboard to simulate daily stress. The response is the % of nail chipped off as measured by scanning images of the nails and using an image processing program.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
personal communication
}
\concept{Experiment Design}
