% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nambe}
\alias{dasl.nambe}
\title{Nambe}
\format{59 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nambe/?sf_paged=28}{Nambe}
}
\description{
Nambe
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
