% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nambe_2018}
\alias{dasl.nambe_2018}
\title{Nambe 2018}
\format{59 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nambe-2018/?sf_paged=28}{Nambe 2018}
}
\description{
Nambe 2018
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Velleman Consulting (adjusted for inflation)
}
\concept{Regression}
