% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nashville}
\alias{dasl.nashville}
\title{Nashville}
\format{70 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nashville/?sf_paged=28}{Nashville}
}
\description{
A hospital in Nashville is considering changes to the prenatal care they offer. They collected the gestation times of 70 pregnancies that ended in live births. The established human gestation time is 266 days.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Statistical Inference}
