% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.neck_size}
\alias{dasl.neck_size}
\title{Neck size}
\format{250 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/neck-size/?sf_paged=28}{Neck size}
}
\description{
Neck size
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
