% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nenana_2017}
\alias{dasl.nenana_2017}
\title{Nenana 2017}
\format{101 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nenana-2017/?sf_paged=28}{Nenana 2017}
}
\description{
Nenana 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Nenana Ice Classic 2011. Nenana Ice Classic: Tanana River Ice Annual Breakup Dates, Version 1. [Indicate subset used]. Boulder, Colorado USA. NSIDC: National Snow and Ice Data Center. doi: \url{http://dx.doi.org/10.5067/DURFYP131STS.}  [Date Accessed]. Note: Our dates are 1 day lower than these dates. We count days after Jan 1.
}
\concept{Regression}
