% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nest_egg_index}
\alias{dasl.nest_egg_index}
\title{Nest Egg Index}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nest-egg-index/?sf_paged=28}{Nest Egg Index}
}
\description{
Nest Egg Index
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.statemaster.com/graph/eco_nes_egg_ind-economy-nest-egg-index}
}
\concept{Boxplots}
