% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.new_york_bridges_2016}
\alias{dasl.new_york_bridges_2016}
\title{New York bridges 2016}
\format{17504 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/new-york-bridges-2016/?sf_paged=28}{New York bridges 2016}
}
\description{
New York State inspectors assess all bridges in the state every two years including a bridge’s individual parts. Bridges are analyzed for their capacity to carry vehicular loads. Inspectors are required to evaluate, assign a condition score, and document the condition of up to 47 structural elements, including rating 25 components of each span of a bridge, in addition to general components common to all bridges. The NYSDOT condition rating scale ranges from 1 to 7, with 7 being in new condition and a rating of 5 or greater considered as good conditionBridges that cannot safely carry heavy vehicles, such as some tractor trailers, are posted with weight limits. Based upon inspection and load capacity analysis, any bridge deemed unsafe gets closed. How does the condition of the bridge relate to its age? Are there any outliers? Can you account for them by identifying them?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.ny.gov/Transportation/Bridge-Conditions-NYS-Department-of-Transportation/wpyb-cjy8}
}
\concept{Regression}
