% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.newspapers}
\alias{dasl.newspapers}
\title{Newspapers}
\format{14 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/newspapers/?sf_paged=28}{Newspapers}
}
\description{
Eurostat, an agency of the European Union (EU), conducts surveys on several aspects of daily life in EU countries. Recently, the agency asked samples of 1000 respondents in each of 14 European countries whether they read the newspaper on a daily basis.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Eurostat survey
}
