% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nhanes}
\alias{dasl.nhanes}
\title{NHANES}
\format{2623 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nhanes/?sf_paged=28}{NHANES}
}
\description{
The National Health and Nutrition Examination Survey (NHANES) is a program of studies designed to assess the health and nutritional status of adults and children in the United States. The survey is unique in that it combines interviews and physical examinations.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.cdc.gov/nchs/nhanes/}
}
\concept{Confidence Intervals for Means}
