% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nightmares}
\alias{dasl.nightmares}
\title{Nightmares}
\format{63 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nightmares/?sf_paged=28}{Nightmares}
}
\description{
Researchers  interviewed participants to find some who reliably fell asleep and awoke on one side and who could remember their dreams. They found 63 participants, of whom 41 were right-side sleepers and 22 slept on their left side. Then they interviewed them about their dreams. Of the 41 right-side sleepers, only 6 reported often having nightmares. But of the 22 left-side sleepers 9 reported nightmares. Is the difference significant?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.sleepandhypnosis.org/pdf/323031765.pdf}  “Sleeping Position, Dream Emotions, and Subjective Sleep Quality”, Mehmet Yucel Agargun, M.D., Murat Boysan, M.A., Lutfu Hanoglu, M.D.
}
\concept{Tables}
