% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nissan}
\alias{dasl.nissan}
\title{Nissan}
\format{100 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nissan/?sf_paged=28}{Nissan}
}
\description{
Richard DeVeaux owned a Nissan Maxima for 8 years. He recorded the car’s fuel efficiency (in mpg) each time he filled the tank. He wanted to know what fuel efficiency to expect as “ordinary” for his car. Knowing this, he was able to predict when he’d need to fill the tank again and to notice if the fuel efficiency suddenly got worse, which could be a sign of trouble.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Deveaux Data
}
\concept{Data Display}
\concept{Normal Probability Plots}
\concept{Normal model}
