% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.normal_temperature}
\alias{dasl.normal_temperature}
\title{Normal temperature}
\format{52 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/normal-temperature/?sf_paged=28}{Normal temperature}
}
\description{
Body temperatures of a random sample of 52 healthy adults, reported in degrees Fahrenheit.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.oaklandairport.com/news/statistics/passenger-history/}  \url{http://www.oaklandairport.com/wp-content/uploads/2016/08/enplanements_deplanements_FY10-1.xls}
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
