% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nuclear_plants}
\alias{dasl.nuclear_plants}
\title{Nuclear plants}
\format{32 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nuclear-plants/?sf_paged=29}{Nuclear plants}
}
\description{
Costs of construction for 32 light water nuclear plants.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Multiple Regression}
