% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ny_marathon_2016}
\alias{dasl.ny_marathon_2016}
\title{NY Marathon 2016}
\format{39 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ny-marathon-2016/?sf_paged=29}{NY Marathon 2016}
}
\description{
NY Marathon 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.nycmarathon.org
}
\concept{Blocking}
\concept{Paired Data}
