% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.nypd}
\alias{dasl.nypd}
\title{NYPD}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/nypd/?sf_paged=29}{NYPD}
}
\description{
NYPD
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.nbcnewyork.com/news/local/NYPD-Faces-Suit-For-Race-Sex-Discrimination-65164502.html}  (data source unknown)
}
\concept{Chi Square Tests}
