% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oakland_passengers_2016}
\alias{dasl.oakland_passengers_2016}
\title{Oakland passengers 2016}
\format{318 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oakland-passengers-2016/?sf_paged=29}{Oakland passengers 2016}
}
\description{
The data give the number of passengers at Oakland (CA) airport month by month since 1997.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.oaklandairport.com/news/statistics/passenger-history/}
}
\concept{Regression}
\concept{Time Series Analysis}
