% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.obesity_and_exercise}
\alias{dasl.obesity_and_exercise}
\title{Obesity and exercise}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/obesity-and-exercise/?sf_paged=29}{Obesity and exercise}
}
\description{
Obesity and exercise
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Boxplots}
\concept{Tables}
