% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oecd_economic_regulations}
\alias{dasl.oecd_economic_regulations}
\title{OECD economic regulations}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oecd-economic-regulations/?sf_paged=29}{OECD economic regulations}
}
\description{
A study by the U.S. Small Business Administration used historical data to model the GDP per capita of 24 of the countries in the Organization for Economic Cooperation and Development(OECD). The researchers hoped to show that more regulation leads to lower GDP/Capita. The multiple regression with all terms does have a significant P-value for Economic Regulation Index. However, Primary Education is not a significant predictor. If it is removed from the model, then OECD Regulation is no longer significant at .05. Was it added to the model just to judge the P-value of OECD regulation down to permit a publication that claimed an effect? Check to see whether you think there is such an effect.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Crain, M. W., The Impact of Regulatory Costs on Small Firms, available at www.sba.gov/advocacy/7540/49291
}
\concept{Multiple Regression}
