% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oecd_gdp}
\alias{dasl.oecd_gdp}
\title{OECD GDP}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oecd-gdp/?sf_paged=29}{OECD GDP}
}
\description{
OECD GDP
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.oecd.org/gdp/real-gdp-forecast.htm}
}
\concept{Regression}
