% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oecd_gdp_growth}
\alias{dasl.oecd_gdp_growth}
\title{OECD GDP Growth}
\format{35 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oecd-gdp-growth/?sf_paged=29}{OECD GDP Growth}
}
\description{
OECD GDP Growth
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.sba.gov/advo/research/rs264tot.pdf
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
