% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oecd_unemployment}
\alias{dasl.oecd_unemployment}
\title{OECD Unemployment}
\format{38 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oecd-unemployment/?sf_paged=29}{OECD Unemployment}
}
\description{
OECD Unemployment
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.oecd.org/unemp/unemployment-rate.htm}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
