% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oil_prices_2016}
\alias{dasl.oil_prices_2016}
\title{Oil prices 2016}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oil-prices-2016/?sf_paged=29}{Oil prices 2016}
}
\description{
The price (per barrel) of oil has fluctuated over time. Various attempts to model it are generally not successful. The data include both the inflation-adjusted prices of a barrel of oil from 1968 to 2016 and two prediction models.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.eia.gov/dnav/pet/pet_pri_spt_s1_a.htm}
}
\concept{Regression}
