% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oil_prices_monthly}
\alias{dasl.oil_prices_monthly}
\title{Oil prices monthly}
\format{75 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oil-prices-monthly/?sf_paged=29}{Oil prices monthly}
}
\description{
The price (per barrel) of oil has fluctuated over time. Various attempts to model it are generally not successful. The data include the monthly crude oil price in $/barrel from January 2001 to March 2007.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.eia.gov/dnav/pet/pet_pri_spt_s1_a.htm}
}
\concept{Time Series Analysis}
