% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.okcupid_catsdogs}
\alias{dasl.okcupid_catsdogs}
\title{OKCupid CatsDogs}
\format{59943 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/okcupid-catsdogs/?sf_paged=29}{OKCupid CatsDogs}
}
\description{
OKCupid CatsDogs
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
OKCupid.com
}
\concept{Tables}
