% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.old_faithful}
\alias{dasl.old_faithful}
\title{Old Faithful}
\format{222 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/old-faithful/?sf_paged=30}{Old Faithful}
}
\description{
It is a common belief that Yellowstone’s most famous geyser erupts once an hour at very predictable intervals. But, in fact, the intervals between eruptions can vary greatly. Can we predict the interval from, for example, the duration of the previous eruption? Are there other patterns in the data worth noting?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
