% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.old_pitchers}
\alias{dasl.old_pitchers}
\title{Old pitchers}
\format{14 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/old-pitchers/?sf_paged=30}{Old pitchers}
}
\description{
Old pitchers
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
