% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.olympic_jumps_2016}
\alias{dasl.olympic_jumps_2016}
\title{Olympic jumps 2016}
\format{27 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/olympic-jumps-2016/?sf_paged=30}{Olympic jumps 2016}
}
\description{
How are Olympic performances in various events related? The data gives winning long-jump and high-jump distances in meters, for the Summer Olympics from 1912 through 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
various
}
