% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.online_insurance}
\alias{dasl.online_insurance}
\title{Online insurance}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/online-insurance/?sf_paged=30}{Online insurance}
}
\description{
Online insurance
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Comparing Two Groups}
\concept{Nonparametric Methods}
\concept{Paired Data}
