% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.online_shopping}
\alias{dasl.online_shopping}
\title{Online Shopping}
\format{15 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/online-shopping/?sf_paged=30}{Online Shopping}
}
\description{
Online Shopping
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Comparing Two Groups}
\concept{Paired Data}
\concept{Regression}
