% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.oranges}
\alias{dasl.oranges}
\title{Oranges}
\format{14 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/oranges/?sf_paged=30}{Oranges}
}
\description{
As the number of oranges on a tree increases, the fruit tends to get smaller. The dataset gives numbers of oranges/tree and average weight/orange (in pounds).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
