% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ozone}
\alias{dasl.ozone}
\title{Ozone}
\format{546 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ozone/?sf_paged=30}{Ozone}
}
\description{
Ozone levels (in parts per billion, ppb) were recorded at sites in New Jersey monthly between 1926 and 1971. Here are boxplots of the data for each month (over the 46 years), lined up in order (January = 1):
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Cleveland W. Kleiner B. Mc Rae J. Warner J. "The Analysis of Ground-level ozone data from New Jersey, New York, Connecticut, and Massachusetts: Data quality Assessment and Temporal and Geographical Properties"
}
\concept{Boxplots}
\concept{Displaying Quantitative Data}
\concept{Regression}
\concept{Summarizing Quantitative Data}
