% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.paper_airplanes}
\alias{dasl.paper_airplanes}
\title{Paper airplanes}
\format{64 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/paper-airplanes/?sf_paged=30}{Paper airplanes}
}
\description{
Paper airplanes
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Student experiment from De Veaux
}
\concept{Multifactor Analysis of Variance}
