% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.paralyzed_veterans}
\alias{dasl.paralyzed_veterans}
\title{Paralyzed veterans}
\format{7600 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/paralyzed-veterans/?sf_paged=30}{Paralyzed veterans}
}
\description{
Paralyzed veterans
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Hypothesis Tests}
