% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.passive_smoke}
\alias{dasl.passive_smoke}
\title{Passive smoke}
\format{150 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/passive-smoke/?sf_paged=30}{Passive smoke}
}
\description{
Researchers measured the concentration (nanograms per milliliter) of cotinine in the blood of three groups of people: nonsmokers who have not been exposed to smoke, nonsmokers who have been Exposed To Smoke (ETS), and smokers. Cotinine is left in the blood when the body metabolizes nicotine, so its value is a direct measurement of the effect of passive smoke exposure.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Boxplots}
