% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.penguins}
\alias{dasl.penguins}
\title{Penguins}
\format{125 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/penguins/?sf_paged=30}{Penguins}
}
\description{
Emperor penguins are the most accomplished divers among birds, making routine dives of 5–12 minutes, with the longest recorded dive over 27 minutes. These birds can also dive to depths of over 500 meters! Since air-breathing animals like penguins must hold their breath while submerged, the duration of any given dive depends on how much oxygen is in the bird’s body at the beginning of the dive, how quickly that oxygen gets used, and the lowest level of oxygen the bird can tolerate. The rate of oxygen depletion is primarily determined by the penguin’s heart rate. Consequently, studies of heart rates during dives can help us understand how these animals regulate their oxygen consumption in order to make such impressive dives.The researchers equipped emperor penguins with devices that record their heart rates during dives. The dataset reports Dive Heart Rate (beats per minute), the Duration (minutes) of dives, and other related variables.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Jessica Meir personal communication
}
\concept{Regression}
