% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.personal_appearance}
\alias{dasl.personal_appearance}
\title{Personal appearance}
\format{7 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/personal-appearance/?sf_paged=31}{Personal appearance}
}
\description{
The GfK Roper Reports® Worldwide Survey asked 30,000 consumers in 23 countries about their attitudes on health, beauty, and other personal values. One question participants were asked was how important their personal appearance is to them. The data are a contingency table of responses to this question by age decade.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
GfK Roper Reports‹› Worldwide Survey
}
\concept{Chi Square Tests}
