% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.pi}
\alias{dasl.pi}
\title{Pi}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/pi/?sf_paged=31}{Pi}
}
\description{
Pi
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
pi
}
\concept{Chi Square Tests}
