% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.pima_indians}
\alias{dasl.pima_indians}
\title{Pima indians}
\format{768 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/pima-indians/?sf_paged=31}{Pima indians}
}
\description{
The Pima Indians of southern Arizona are a unique community. Their ancestors were among the first people to cross over into the Americas some 30,000 years ago. For at least two millennia, they have lived in the Sonoran Desert near the Gila River. Known throughout history as a generous people, they have given of themselves for the past 30 years helping researchers at the National Institutes of Health study certain diseases like diabetes and obe-sity. Young Pima Indians often marry other Pimas, making them an ideal group for genetic researchers to study. Pimas also have an extremely high incidence of diabetes. Researchers investigating factors for increased risk of diabetes examined data on  768 adult women of Pima Indian heritage. One possible predictor is the body mass index, BMI, calculated as weight/height2, where weight is measured in kilograms and height in meters. We are interested in the relationship between BMI and the incidence of diabetes.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
UCI Machine Learning Repository \url{http://archive.ics.uci.edu/ml/datasets}  /Pima+Indians+Diabetes.
}
\concept{Regression}
