% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.pizza_prices}
\alias{dasl.pizza_prices}
\title{Pizza prices}
\format{156 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/pizza-prices/?sf_paged=31}{Pizza prices}
}
\description{
Sales volume and price of a slice of plain pizza ($) in Baltimore, Dallas, Chicago, and Denver for 156 weeks. How are prices and sales volumes related? Are patterns the same across cities?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
From Norean Sharpe personal communication
}
\concept{Boxplots}
\concept{Correlation}
\concept{Data Display}
\concept{Displaying Quantitative Data}
\concept{Regression}
\concept{Scatterplot}
\concept{Summarizing Quantitative Data}
