% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.pizza_ratings}
\alias{dasl.pizza_ratings}
\title{Pizza ratings}
\format{29 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/pizza-ratings/?sf_paged=31}{Pizza ratings}
}
\description{
Pizza ratings
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Consumers' Union
}
\concept{Multiple Regression}
