% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.planets}
\alias{dasl.planets}
\title{Planets}
\format{9 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/planets/?sf_paged=31}{Planets}
}
\description{
On August 24, 2006, the International Astronomical Union voted that Pluto is not a planet. Some members of the public have been reluctant to accept that decision. The data show the average distance of each of the traditional nine planets from the sun. Is there a pattern? Does Pluto fit with the other “official” planets?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.enchantedlearning.com/subjects/astronomy/planets/}
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
