% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.planets_more_data}
\alias{dasl.planets_more_data}
\title{Planets more data}
\format{9 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/planets-more-data/?sf_paged=31}{Planets more data}
}
\description{
On August 24, 2006, the International Astronomical Union voted that Pluto is not a planet. Some members of the public have been reluctant to accept that decision. The data show a variety of facts about the 8 planets and Pluto. Exercises consider two models for the planets. Does Pluto behave like a planet?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.windows2universe.org/our_solar_system/planets_table.html}
}
\concept{Correlation}
\concept{Scatterplot}
