% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.police_pay_2016}
\alias{dasl.police_pay_2016}
\title{Police pay 2016}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/police-pay-2016/?sf_paged=31}{Police pay 2016}
}
\description{
Police pay 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.how-to-become-a-police-officer.com/salaries/}  \url{https://ucr.fbi.gov/crime-in-the-u.s/2016/crime-in-the-u.s.-2016/tables/table-3}  \url{http://www.governing.com/gov-data/high-school-graduation-rates-by-state.html}
}
\concept{Multiple Regression}
