% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.polling}
\alias{dasl.polling}
\title{Polling}
\format{14 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/polling/?sf_paged=31}{Polling}
}
\description{
How accurate are pollsters in predicting the outcomes of Congressional elections? The table shows the actual number of Democrat seats in the House of Representatives and the number predicted by the Gallup organization for nonpresidential election years in the 4 decades following World War II.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Polling data from Gallup
}
\concept{Comparing Two Groups}
\concept{Correlation}
\concept{Regression}
