% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.popcorn}
\alias{dasl.popcorn}
\title{Popcorn}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/popcorn/?sf_paged=31}{Popcorn}
}
\description{
Yvon Hopps ran an experiment to determine optimum power and time settings for microwave popcorn. His goal was to find a combination of power and time that would deliver high-quality popcorn with less than 10% of the kernels left unpopped, on average. After experimenting with several bags, he determined that power 9 at 4 minutes was the best combination. To be sure that the method was successful, he popped 8 more bags of popcorn (selected at random) at this setting. The data report the percentages of uncooked popcorn.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
De Veaux class project
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
