% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.population_commute_times}
\alias{dasl.population_commute_times}
\title{Population Commute Times}
\format{5000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/population-commute-times/?sf_paged=31}{Population Commute Times}
}
\description{
Population Commute Times
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
De Veaux personal data (generated from actual data)
}
\concept{Confidence Intervals for Means}
\concept{Data Display}
\concept{Normal Probability Plots}
