% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.population_growth_2010}
\alias{dasl.population_growth_2010}
\title{Population growth 2010}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/population-growth-2010/?sf_paged=31}{Population growth 2010}
}
\description{
the percentage change in population for the 50 states and the District of Columbia from the 2000 census to the 2010 census.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.census.gov/compendia/statab/rankings.html
}
\concept{Boxplots}
\concept{Data Display}
