% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.poverty_and_region_2015}
\alias{dasl.poverty_and_region_2015}
\title{Poverty and Region 2015}
\format{4 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/poverty-and-region-2015/?sf_paged=32}{Poverty and Region 2015}
}
\description{
Poverty and Region 2015
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www2.census.gov/programs-surveys/demo/tables/p60/256/pov_table3.xls}
}
\concept{Data Displays}
