% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.pregnancies}
\alias{dasl.pregnancies}
\title{Pregnancies}
\format{4 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/pregnancies/?sf_paged=32}{Pregnancies}
}
\description{
Pregnancies
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
