% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.prices_and_earnings}
\alias{dasl.prices_and_earnings}
\title{Prices and Earnings}
\format{72 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/prices-and-earnings/?sf_paged=32}{Prices and Earnings}
}
\description{
UBS (one of the largest banks in the world) prepared a report comparing prices, wages, and other economic conditions in cities around the world for it’s international clients. Some of the variables it measured in 73 cities are Cost of Living, Food Costs, Average Hourly Wage, average number of Working Hours per Year, average number of Vacation Days, hours of work (at the average wage) needed to buy an iPhone, minutes of work needed to buy a Big Mac, and Women’s Clothing Cost.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
USB - Prices and Earning: A comparison of purchasing power around the globe/2012 edition www.economist.com/node/14288808?story_id=14288808
}
\concept{Correlation}
\concept{Scatterplot}
