% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.prisons_2014}
\alias{dasl.prisons_2014}
\title{Prisons 2014}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/prisons-2014/?sf_paged=32}{Prisons 2014}
}
\description{
Prisons 2014
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Bureau of Justice Statistics \url{http://www.bjs.gov/}  \url{http://www.bjs.gov/content/pub/pdf/p14.pdf}
}
\concept{Boxplots}
