% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.process_improvements}
\alias{dasl.process_improvements}
\title{Process improvements}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/process-improvements/?sf_paged=32}{Process improvements}
}
\description{
Process improvements
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Analysis of Variance}
