% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.product_testing}
\alias{dasl.product_testing}
\title{Product testing}
\format{22 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/product-testing/?sf_paged=32}{Product testing}
}
\description{
A company is producing and marketing new reading activities for elementary school children that it believes will improve reading comprehension scores. A researcher randomly assigns third graders to an eight-week program in which some will use these activities and others will experience traditional teaching methods. At the end of the experiment, both groups take a reading comprehension exam. Do these results suggest that the new activities are better?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented data
}
\concept{Comparing Two Groups}
\concept{Nonparametric Methods}
