% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.productivity_2016}
\alias{dasl.productivity_2016}
\title{Productivity 2016}
\format{48 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/productivity-2016/?sf_paged=32}{Productivity 2016}
}
\description{
Productivity 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.bls.gov/lpc/tables_by_sector_and_industry.htm,}  Three Digit Data
}
\concept{Comparing Two Groups}
\concept{Regression}
