% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.property_assessment}
\alias{dasl.property_assessment}
\title{Property assessment}
\format{18 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/property-assessment/?sf_paged=32}{Property assessment}
}
\description{
The size in square feet and assessment of 18 homes in Ithaca NY
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
public records
}
\concept{Regression}
