% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.pulse_rates}
\alias{dasl.pulse_rates}
\title{Pulse rates}
\format{52 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/pulse-rates/?sf_paged=32}{Pulse rates}
}
\description{
A medical researcher measured the pulse rates (beats per minute) of a sample of randomly selected adults.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Confidence Intervals for Means}
