% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.push_ups}
\alias{dasl.push_ups}
\title{Push ups}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/push-ups/?sf_paged=32}{Push ups}
}
\description{
Push ups
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Paired Data}
