% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.pva}
\alias{dasl.pva}
\title{PVA}
\format{7600 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/pva/?sf_paged=33}{PVA}
}
\description{
The Paralyzed Veterans of America (PVA) is a Congressionally chartered veterans’ service organization that represents the interests of paralyzed veterans. The agency provides a range of services to veterans who have spinal cord injury or dysfunction. It derives most of its funding from contributions. The data set PVA contains a sample of the data on donors who recently gave money to the organization.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://kdd.ics.uci.edu/databases/kddcup98/kddcup98.html}
}
\concept{Multiple Regression}
