% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.qc_clarity}
\alias{dasl.qc_clarity}
\title{QC Clarity}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/qc-clarity/?sf_paged=33}{QC Clarity}
}
\description{
QC Clarity
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented realistic data
}
\concept{Quality Control}
