% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.qc_control}
\alias{dasl.qc_control}
\title{QC Control}
\format{250 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/qc-control/?sf_paged=33}{QC Control}
}
\description{
QC Control
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented realistic data
}
\concept{Quality Control}
