% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.qc_production}
\alias{dasl.qc_production}
\title{QC Production}
\format{150 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/qc-production/?sf_paged=33}{QC Production}
}
\description{
QC Production
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented realistic data
}
\concept{Quality Control}
