% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.quality_control_holes}
\alias{dasl.quality_control_holes}
\title{Quality control holes}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/quality-control-holes/?sf_paged=33}{Quality control holes}
}
\description{
Quality control holes
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
