% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.quarterback_performance_2017}
\alias{dasl.quarterback_performance_2017}
\title{Quarterback performance 2017}
\format{30 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/quarterback-performance-2017/?sf_paged=33}{Quarterback performance 2017}
}
\description{
Quarterback performance 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://overthecap.com/position/quarterback/2017}  \url{http://www.espn.com/nfl/qbr}
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
