% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.racial_steering}
\alias{dasl.racial_steering}
\title{Racial steering}
\format{212 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/racial-steering/?sf_paged=33}{Racial steering}
}
\description{
Racial steering
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
