% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\name{dasl}
\alias{dasl}
\title{DASL – The Data And Story Library}
\description{
519 data sets:
\tabular{rrr}{
\strong{Name in R}\tab\strong{Data set name}\tab\strong{Description}\cr
\code{dasl.a_rod_2016} \tab \link[=dasl.a_rod_2016]{A-Rod 2016} \tab  \cr
\code{dasl.AR2016} \tab \link[=dasl.a_rod_2016]{A-Rod 2016} \tab  \cr
\code{dasl.aig_daily} \tab \link[=dasl.aig_daily]{AIG daily} \tab  \cr
\code{dasl.aig_monthly} \tab \link[=dasl.aig_monthly]{AIG monthly} \tab  \cr
\code{dasl.accidents} \tab \link[=dasl.accidents]{Accidents} \tab  \cr
\code{dasl.acela} \tab \link[=dasl.acela]{Acela} \tab  \cr
\code{dasl.acid_rain} \tab \link[=dasl.acid_rain]{Acid rain} \tab  \cr
\code{dasl.activating_baking_yeast} \tab \link[=dasl.activating_baking_yeast]{Activating baking yeast} \tab  \cr
\code{dasl.ABY} \tab \link[=dasl.activating_baking_yeast]{Activating baking yeast} \tab  \cr
\code{dasl.adoptions} \tab \link[=dasl.adoptions]{Adoptions} \tab  \cr
\code{dasl.advertising} \tab \link[=dasl.advertising]{Advertising} \tab  \cr
\code{dasl.age_and_party_2016} \tab \link[=dasl.age_and_party_2016]{Age and party 2016} \tab  \cr
\code{dasl.AAP2016} \tab \link[=dasl.age_and_party_2016]{Age and party 2016} \tab  \cr
\code{dasl.airline_bumping_2017} \tab \link[=dasl.airline_bumping_2017]{Airline bumping 2017} \tab  \cr
\code{dasl.AB2017} \tab \link[=dasl.airline_bumping_2017]{Airline bumping 2017} \tab  \cr
\code{dasl.airport_screening} \tab \link[=dasl.airport_screening]{Airport screening} \tab  \cr
\code{dasl.all_the_efficiency} \tab \link[=dasl.all_the_efficiency]{All the efficiency} \tab  \cr
\code{dasl.ATE} \tab \link[=dasl.all_the_efficiency]{All the efficiency} \tab  \cr
\code{dasl.alternative_energy_2016} \tab \link[=dasl.alternative_energy_2016]{Alternative energy 2016} \tab  \cr
\code{dasl.AE2016} \tab \link[=dasl.alternative_energy_2016]{Alternative energy 2016} \tab  \cr
\code{dasl.amazon_books} \tab \link[=dasl.amazon_books]{Amazon books} \tab  \cr
\code{dasl.analgesics} \tab \link[=dasl.analgesics]{Analgesics} \tab  \cr
\code{dasl.antidepressants} \tab \link[=dasl.antidepressants]{Antidepressants} \tab  \cr
\code{dasl.apples_and_gas} \tab \link[=dasl.apples_and_gas]{Apples and gas} \tab  \cr
\code{dasl.AAG} \tab \link[=dasl.apples_and_gas]{Apples and gas} \tab  \cr
\code{dasl.archery} \tab \link[=dasl.archery]{Archery} \tab  \cr
\code{dasl.attendance_2016} \tab \link[=dasl.attendance_2016]{Attendance 2016} \tab  \cr
\code{dasl.auto_noise_filters} \tab \link[=dasl.auto_noise_filters]{Auto noise filters} \tab  \cr
\code{dasl.ANF} \tab \link[=dasl.auto_noise_filters]{Auto noise filters} \tab  \cr
\code{dasl.bls_output} \tab \link[=dasl.bls_output]{BLS output} \tab  \cr
\code{dasl.babysamp_98} \tab \link[=dasl.babysamp_98]{Babysamp 98} \tab  \cr
\code{dasl.ballplayer_births} \tab \link[=dasl.ballplayer_births]{Ballplayer births} \tab  \cr
\code{dasl.barbershop_music} \tab \link[=dasl.barbershop_music]{Barbershop music} \tab  \cr
\code{dasl.baseball_2016} \tab \link[=dasl.baseball_2016]{Baseball 2016} \tab  \cr
\code{dasl.baseball_attendance} \tab \link[=dasl.baseball_attendance]{Baseball attendance} \tab  \cr
\code{dasl.baseball_circumferences} \tab \link[=dasl.baseball_circumferences]{Baseball circumferences} \tab  \cr
\code{dasl.baseball_salaries_2015} \tab \link[=dasl.baseball_salaries_2015]{Baseball salaries 2015} \tab  \cr
\code{dasl.BS2015} \tab \link[=dasl.baseball_salaries_2015]{Baseball salaries 2015} \tab  \cr
\code{dasl.baseball_salaries_2016} \tab \link[=dasl.baseball_salaries_2016]{Baseball salaries 2016} \tab  \cr
\code{dasl.BS2016} \tab \link[=dasl.baseball_salaries_2016]{Baseball salaries 2016} \tab  \cr
\code{dasl.baseball_weights} \tab \link[=dasl.baseball_weights]{Baseball weights} \tab  \cr
\code{dasl.basketball_shots} \tab \link[=dasl.basketball_shots]{Basketball shots} \tab  \cr
\code{dasl.basketballs} \tab \link[=dasl.basketballs]{Basketballs} \tab  \cr
\code{dasl.baterial_soap} \tab \link[=dasl.baterial_soap]{Baterial soap} \tab  \cr
\code{dasl.batteries} \tab \link[=dasl.batteries]{Batteries} \tab  \cr
\code{dasl.being_successful} \tab \link[=dasl.being_successful]{Being successful} \tab  \cr
\code{dasl.belmont_stakes_2015} \tab \link[=dasl.belmont_stakes_2015]{Belmont stakes 2015} \tab  \cr
\code{dasl.bike_safety_2015} \tab \link[=dasl.bike_safety_2015]{Bike safety 2015} \tab  \cr
\code{dasl.bird_species_2013} \tab \link[=dasl.bird_species_2013]{Bird Species 2013} \tab  \cr
\code{dasl.BS2013} \tab \link[=dasl.bird_species_2013]{Bird Species 2013} \tab  \cr
\code{dasl.birth_order} \tab \link[=dasl.birth_order]{Birth order} \tab  \cr
\code{dasl.birthrates_2015} \tab \link[=dasl.birthrates_2015]{Birthrates 2015} \tab  \cr
\code{dasl.blood_pressure} \tab \link[=dasl.blood_pressure]{Blood Pressure} \tab  \cr
\code{dasl.bodyfat} \tab \link[=dasl.bodyfat]{Bodyfat} \tab  \cr
\code{dasl.bond_funds} \tab \link[=dasl.bond_funds]{Bond funds} \tab  \cr
\code{dasl.bookstore_sales} \tab \link[=dasl.bookstore_sales]{Bookstore sales} \tab  \cr
\code{dasl.boomtowns_2015} \tab \link[=dasl.boomtowns_2015]{Boomtowns 2015} \tab  \cr
\code{dasl.boyle} \tab \link[=dasl.boyle]{Boyle} \tab  \cr
\code{dasl.brain_size} \tab \link[=dasl.brain_size]{Brain size} \tab  \cr
\code{dasl.brain_waves} \tab \link[=dasl.brain_waves]{Brain waves} \tab  \cr
\code{dasl.brakes} \tab \link[=dasl.brakes]{Brakes} \tab  \cr
\code{dasl.broadway_shows} \tab \link[=dasl.broadway_shows]{Broadway shows} \tab  \cr
\code{dasl.burger_king_items} \tab \link[=dasl.burger_king_items]{Burger King items} \tab  \cr
\code{dasl.BKI} \tab \link[=dasl.burger_king_items]{Burger King items} \tab  \cr
\code{dasl.burgers} \tab \link[=dasl.burgers]{Burgers} \tab  \cr
\code{dasl.buy_from_a_friend} \tab \link[=dasl.buy_from_a_friend]{Buy from a friend} \tab  \cr
\code{dasl.BFAF} \tab \link[=dasl.buy_from_a_friend]{Buy from a friend} \tab  \cr
\code{dasl.ca_house_prices} \tab \link[=dasl.ca_house_prices]{CA House Prices} \tab  \cr
\code{dasl.CHP} \tab \link[=dasl.ca_house_prices]{CA House Prices} \tab  \cr
\code{dasl.cape10_2017} \tab \link[=dasl.cape10_2017]{CAPE10 2017} \tab  \cr
\code{dasl.ceo_compensation_2014} \tab \link[=dasl.ceo_compensation_2014]{CEO Compensation 2014} \tab  \cr
\code{dasl.CC2014} \tab \link[=dasl.ceo_compensation_2014]{CEO Compensation 2014} \tab  \cr
\code{dasl.ceo_salary_2012} \tab \link[=dasl.ceo_salary_2012]{CEO Salary 2012} \tab  \cr
\code{dasl.CS2012} \tab \link[=dasl.ceo_salary_2012]{CEO Salary 2012} \tab  \cr
\code{dasl.col_2016} \tab \link[=dasl.col_2016]{COL 2016} \tab  \cr
\code{dasl.colall_2016} \tab \link[=dasl.colall_2016]{COLall 2016} \tab  \cr
\code{dasl.cpi_2017} \tab \link[=dasl.cpi_2017]{CPI 2017} \tab  \cr
\code{dasl.cpi_worldwide_2016} \tab \link[=dasl.cpi_worldwide_2016]{CPI Worldwide 2016} \tab  \cr
\code{dasl.CW2016} \tab \link[=dasl.cpi_worldwide_2016]{CPI Worldwide 2016} \tab  \cr
\code{dasl.camp_sites} \tab \link[=dasl.camp_sites]{Camp sites} \tab  \cr
\code{dasl.candy_bars} \tab \link[=dasl.candy_bars]{Candy bars} \tab  \cr
\code{dasl.car_discounts} \tab \link[=dasl.car_discounts]{Car discounts} \tab  \cr
\code{dasl.car_origins} \tab \link[=dasl.car_origins]{Car origins} \tab  \cr
\code{dasl.car_rental} \tab \link[=dasl.car_rental]{Car rental} \tab  \cr
\code{dasl.car_speeds} \tab \link[=dasl.car_speeds]{Car speeds} \tab  \cr
\code{dasl.car_speeds_100} \tab \link[=dasl.car_speeds_100]{Car speeds 100} \tab  \cr
\code{dasl.CS100} \tab \link[=dasl.car_speeds_100]{Car speeds 100} \tab  \cr
\code{dasl.carbon_footprint} \tab \link[=dasl.carbon_footprint]{Carbon footprint} \tab  \cr
\code{dasl.carbon_footprint_2015} \tab \link[=dasl.carbon_footprint_2015]{Carbon footprint 2015} \tab  \cr
\code{dasl.CF2015} \tab \link[=dasl.carbon_footprint_2015]{Carbon footprint 2015} \tab  \cr
\code{dasl.cars} \tab \link[=dasl.cars]{Cars} \tab  \cr
\code{dasl.case_study_i_pv} \tab \link[=dasl.case_study_i_pv]{Case study I-PV} \tab  \cr
\code{dasl.CSIP} \tab \link[=dasl.case_study_i_pv]{Case study I-PV} \tab  \cr
\code{dasl.case_shiller_by_city} \tab \link[=dasl.case_shiller_by_city]{Case-Shiller by city} \tab  \cr
\code{dasl.CSBC} \tab \link[=dasl.case_shiller_by_city]{Case-Shiller by city} \tab  \cr
\code{dasl.cereal_company} \tab \link[=dasl.cereal_company]{Cereal company} \tab  \cr
\code{dasl.cereals} \tab \link[=dasl.cereals]{Cereals} \tab  \cr
\code{dasl.chips} \tab \link[=dasl.chips]{Chips} \tab  \cr
\code{dasl.chips_ahoy} \tab \link[=dasl.chips_ahoy]{Chips Ahoy!} \tab  \cr
\code{dasl.cholesterol_and_smoking} \tab \link[=dasl.cholesterol_and_smoking]{Cholesterol and smoking} \tab  \cr
\code{dasl.CAS} \tab \link[=dasl.cholesterol_and_smoking]{Cholesterol and smoking} \tab  \cr
\code{dasl.chromatography} \tab \link[=dasl.chromatography]{Chromatography} \tab  \cr
\code{dasl.cigarettes} \tab \link[=dasl.cigarettes]{Cigarettes} \tab  \cr
\code{dasl.city_climate} \tab \link[=dasl.city_climate]{City climate} \tab  \cr
\code{dasl.city_temperatures} \tab \link[=dasl.city_temperatures]{City temperatures} \tab  \cr
\code{dasl.climate_change_2016} \tab \link[=dasl.climate_change_2016]{Climate change 2016} \tab  \cr
\code{dasl.CC2016} \tab \link[=dasl.climate_change_2016]{Climate change 2016} \tab  \cr
\code{dasl.cloud_seeding} \tab \link[=dasl.cloud_seeding]{Cloud seeding} \tab  \cr
\code{dasl.coasters_2015} \tab \link[=dasl.coasters_2015]{Coasters 2015} \tab  \cr
\code{dasl.coffee_prices_2017} \tab \link[=dasl.coffee_prices_2017]{Coffee prices 2017} \tab  \cr
\code{dasl.CP2017} \tab \link[=dasl.coffee_prices_2017]{Coffee prices 2017} \tab  \cr
\code{dasl.commercial_properties} \tab \link[=dasl.commercial_properties]{Commercial properties} \tab  \cr
\code{dasl.commute_times_sample100} \tab \link[=dasl.commute_times_sample100]{Commute times sample100} \tab  \cr
\code{dasl.CTS} \tab \link[=dasl.commute_times_sample100]{Commute times sample100} \tab  \cr
\code{dasl.commute_times_sample50} \tab \link[=dasl.commute_times_sample50]{Commute times sample50} \tab  \cr
\code{dasl.companies} \tab \link[=dasl.companies]{Companies} \tab  \cr
\code{dasl.companies_quickstart} \tab \link[=dasl.companies_quickstart]{Companies Quickstart} \tab  \cr
\code{dasl.computer_chip_manufacturing} \tab \link[=dasl.computer_chip_manufacturing]{Computer chip manufacturing} \tab  \cr
\code{dasl.CCM} \tab \link[=dasl.computer_chip_manufacturing]{Computer chip manufacturing} \tab  \cr
\code{dasl.computer_lab_fees} \tab \link[=dasl.computer_lab_fees]{Computer lab fees} \tab  \cr
\code{dasl.CLF} \tab \link[=dasl.computer_lab_fees]{Computer lab fees} \tab  \cr
\code{dasl.congestion_cost} \tab \link[=dasl.congestion_cost]{Congestion cost} \tab  \cr
\code{dasl.consumer_spending} \tab \link[=dasl.consumer_spending]{Consumer spending} \tab  \cr
\code{dasl.consumer_spending_post_holiday} \tab \link[=dasl.consumer_spending_post_holiday]{Consumer spending post holiday} \tab  \cr
\code{dasl.CSPH} \tab \link[=dasl.consumer_spending_post_holiday]{Consumer spending post holiday} \tab  \cr
\code{dasl.containers_revisited} \tab \link[=dasl.containers_revisited]{Containers revisited} \tab  \cr
\code{dasl.contrast_baths} \tab \link[=dasl.contrast_baths]{Contrast baths} \tab  \cr
\code{dasl.cookies} \tab \link[=dasl.cookies]{Cookies} \tab  \cr
\code{dasl.cost_of_living_2017} \tab \link[=dasl.cost_of_living_2017]{Cost of Living 2017} \tab  \cr
\code{dasl.COL2017} \tab \link[=dasl.cost_of_living_2017]{Cost of Living 2017} \tab  \cr
\code{dasl.cost_of_living_2016} \tab \link[=dasl.cost_of_living_2016]{Cost of living 2016} \tab  \cr
\code{dasl.COL2016} \tab \link[=dasl.cost_of_living_2016]{Cost of living 2016} \tab  \cr
\code{dasl.cost_of_living_2018} \tab \link[=dasl.cost_of_living_2018]{Cost of living 2018} \tab  \cr
\code{dasl.COL2018} \tab \link[=dasl.cost_of_living_2018]{Cost of living 2018} \tab  \cr
\code{dasl.couples} \tab \link[=dasl.couples]{Couples} \tab  \cr
\code{dasl.cramming} \tab \link[=dasl.cramming]{Cramming} \tab  \cr
\code{dasl.cranberry_juice} \tab \link[=dasl.cranberry_juice]{Cranberry juice} \tab  \cr
\code{dasl.craters} \tab \link[=dasl.craters]{Craters} \tab  \cr
\code{dasl.crawling} \tab \link[=dasl.crawling]{Crawling} \tab  \cr
\code{dasl.credit_card_bank} \tab \link[=dasl.credit_card_bank]{Credit card bank} \tab  \cr
\code{dasl.CCB} \tab \link[=dasl.credit_card_bank]{Credit card bank} \tab  \cr
\code{dasl.credit_card_bank_info} \tab \link[=dasl.credit_card_bank_info]{Credit card bank info} \tab  \cr
\code{dasl.CCBI} \tab \link[=dasl.credit_card_bank_info]{Credit card bank info} \tab  \cr
\code{dasl.credit_card_charges} \tab \link[=dasl.credit_card_charges]{Credit card charges} \tab  \cr
\code{dasl.CCC} \tab \link[=dasl.credit_card_charges]{Credit card charges} \tab  \cr
\code{dasl.crocodile_lengths} \tab \link[=dasl.crocodile_lengths]{Crocodile lengths} \tab  \cr
\code{dasl.crowdedness} \tab \link[=dasl.crowdedness]{Crowdedness} \tab  \cr
\code{dasl.cuckoos_and_quality_control} \tab \link[=dasl.cuckoos_and_quality_control]{Cuckoos and quality control} \tab  \cr
\code{dasl.CAQC} \tab \link[=dasl.cuckoos_and_quality_control]{Cuckoos and quality control} \tab  \cr
\code{dasl.cups} \tab \link[=dasl.cups]{Cups} \tab  \cr
\code{dasl.cyclists_2015} \tab \link[=dasl.cyclists_2015]{Cyclists 2015} \tab  \cr
\code{dasl.djia_2017} \tab \link[=dasl.djia_2017]{DJIA 2017} \tab  \cr
\code{dasl.dairy_sales} \tab \link[=dasl.dairy_sales]{Dairy sales} \tab  \cr
\code{dasl.dalia} \tab \link[=dasl.dalia]{Dalia} \tab  \cr
\code{dasl.darts} \tab \link[=dasl.darts]{Darts} \tab  \cr
\code{dasl.defect_monitoring} \tab \link[=dasl.defect_monitoring]{Defect monitoring} \tab  \cr
\code{dasl.defect_monitoring_second_product} \tab \link[=dasl.defect_monitoring_second_product]{Defect monitoring second product} \tab  \cr
\code{dasl.DMSP} \tab \link[=dasl.defect_monitoring_second_product]{Defect monitoring second product} \tab  \cr
\code{dasl.demographics} \tab \link[=dasl.demographics]{Demographics} \tab  \cr
\code{dasl.depression_and_the_internet} \tab \link[=dasl.depression_and_the_internet]{Depression and the internet} \tab  \cr
\code{dasl.DATI} \tab \link[=dasl.depression_and_the_internet]{Depression and the internet} \tab  \cr
\code{dasl.designated_hitter_2012} \tab \link[=dasl.designated_hitter_2012]{Designated hitter 2012} \tab  \cr
\code{dasl.DH2012} \tab \link[=dasl.designated_hitter_2012]{Designated hitter 2012} \tab  \cr
\code{dasl.dexterity} \tab \link[=dasl.dexterity]{Dexterity} \tab  \cr
\code{dasl.diamond_quality} \tab \link[=dasl.diamond_quality]{Diamond quality} \tab  \cr
\code{dasl.diamonds} \tab \link[=dasl.diamonds]{Diamonds} \tab  \cr
\code{dasl.diet} \tab \link[=dasl.diet]{Diet} \tab  \cr
\code{dasl.direct_mail} \tab \link[=dasl.direct_mail]{Direct mail} \tab  \cr
\code{dasl.direct_mail_follow_up} \tab \link[=dasl.direct_mail_follow_up]{Direct mail follow up} \tab  \cr
\code{dasl.DMFU} \tab \link[=dasl.direct_mail_follow_up]{Direct mail follow up} \tab  \cr
\code{dasl.dirt_bikes_2014} \tab \link[=dasl.dirt_bikes_2014]{Dirt bikes 2014} \tab  \cr
\code{dasl.DB2014} \tab \link[=dasl.dirt_bikes_2014]{Dirt bikes 2014} \tab  \cr
\code{dasl.disk_drives_2016} \tab \link[=dasl.disk_drives_2016]{Disk drives 2016} \tab  \cr
\code{dasl.DD2016} \tab \link[=dasl.disk_drives_2016]{Disk drives 2016} \tab  \cr
\code{dasl.doctors_and_life_expectancy} \tab \link[=dasl.doctors_and_life_expectancy]{Doctors and life expectancy} \tab  \cr
\code{dasl.DALE} \tab \link[=dasl.doctors_and_life_expectancy]{Doctors and life expectancy} \tab  \cr
\code{dasl.donor_profiles} \tab \link[=dasl.donor_profiles]{Donor profiles} \tab  \cr
\code{dasl.doritos} \tab \link[=dasl.doritos]{Doritos} \tab  \cr
\code{dasl.dow_jones_2015} \tab \link[=dasl.dow_jones_2015]{Dow Jones 2015} \tab  \cr
\code{dasl.DJ2015} \tab \link[=dasl.dow_jones_2015]{Dow Jones 2015} \tab  \cr
\code{dasl.dow_jones_2017} \tab \link[=dasl.dow_jones_2017]{Dow Jones 2017} \tab  \cr
\code{dasl.DJ2017} \tab \link[=dasl.dow_jones_2017]{Dow Jones 2017} \tab  \cr
\code{dasl.down_the_drain} \tab \link[=dasl.down_the_drain]{Down the Drain} \tab  \cr
\code{dasl.DTD} \tab \link[=dasl.down_the_drain]{Down the Drain} \tab  \cr
\code{dasl.downloading} \tab \link[=dasl.downloading]{Downloading} \tab  \cr
\code{dasl.downloads} \tab \link[=dasl.downloads]{Downloads} \tab  \cr
\code{dasl.drivers_licenses_2014} \tab \link[=dasl.drivers_licenses_2014]{Drivers Licenses 2014} \tab  \cr
\code{dasl.DL2014} \tab \link[=dasl.drivers_licenses_2014]{Drivers Licenses 2014} \tab  \cr
\code{dasl.drug_abuse} \tab \link[=dasl.drug_abuse]{Drug abuse} \tab  \cr
\code{dasl.drug_use_2013} \tab \link[=dasl.drug_use_2013]{Drug use 2013} \tab  \cr
\code{dasl.DU2013} \tab \link[=dasl.drug_use_2013]{Drug use 2013} \tab  \cr
\code{dasl.e_commerce} \tab \link[=dasl.e_commerce]{E-commerce} \tab  \cr
\code{dasl.e_mails} \tab \link[=dasl.e_mails]{E-mails} \tab  \cr
\code{dasl.epa_mpg_2013} \tab \link[=dasl.epa_mpg_2013]{EPA mpg 2013} \tab  \cr
\code{dasl.EM2013} \tab \link[=dasl.epa_mpg_2013]{EPA mpg 2013} \tab  \cr
\code{dasl.erp_effectiveness} \tab \link[=dasl.erp_effectiveness]{ERP Effectiveness} \tab  \cr
\code{dasl.eu_imports} \tab \link[=dasl.eu_imports]{EU Imports} \tab  \cr
\code{dasl.earnings} \tab \link[=dasl.earnings]{Earnings} \tab  \cr
\code{dasl.education_and_mortality} \tab \link[=dasl.education_and_mortality]{Education and mortality} \tab  \cr
\code{dasl.EAM} \tab \link[=dasl.education_and_mortality]{Education and mortality} \tab  \cr
\code{dasl.education_by_age} \tab \link[=dasl.education_by_age]{Education by age} \tab  \cr
\code{dasl.EBA} \tab \link[=dasl.education_by_age]{Education by age} \tab  \cr
\code{dasl.eggs} \tab \link[=dasl.eggs]{Eggs} \tab  \cr
\code{dasl.egyptians} \tab \link[=dasl.egyptians]{Egyptians} \tab  \cr
\code{dasl.el_nino_2017} \tab \link[=dasl.el_nino_2017]{El Nino 2017} \tab  \cr
\code{dasl.EN2017} \tab \link[=dasl.el_nino_2017]{El Nino 2017} \tab  \cr
\code{dasl.election_2000} \tab \link[=dasl.election_2000]{Election 2000} \tab  \cr
\code{dasl.energy_use_and_recession} \tab \link[=dasl.energy_use_and_recession]{Energy use and recession} \tab  \cr
\code{dasl.EUAR} \tab \link[=dasl.energy_use_and_recession]{Energy use and recession} \tab  \cr
\code{dasl.euro_dollar} \tab \link[=dasl.euro_dollar]{Euro Dollar} \tab  \cr
\code{dasl.euro_us_gdp_growth_2017} \tab \link[=dasl.euro_us_gdp_growth_2017]{Euro-US GDP growth 2017} \tab  \cr
\code{dasl.EUGG2017} \tab \link[=dasl.euro_us_gdp_growth_2017]{Euro-US GDP growth 2017} \tab  \cr
\code{dasl.exercise_equipment} \tab \link[=dasl.exercise_equipment]{Exercise equipment} \tab  \cr
\code{dasl.expensive_cities} \tab \link[=dasl.expensive_cities]{Expensive cities} \tab  \cr
\code{dasl.eye_and_hair_color} \tab \link[=dasl.eye_and_hair_color]{Eye and Hair color} \tab  \cr
\code{dasl.EAHC} \tab \link[=dasl.eye_and_hair_color]{Eye and Hair color} \tab  \cr
\code{dasl.f_stops} \tab \link[=dasl.f_stops]{F-stops} \tab  \cr
\code{dasl.farmed_salmon} \tab \link[=dasl.farmed_salmon]{Farmed salmon} \tab  \cr
\code{dasl.faulty_pixels} \tab \link[=dasl.faulty_pixels]{Faulty pixels} \tab  \cr
\code{dasl.female_president} \tab \link[=dasl.female_president]{Female president} \tab  \cr
\code{dasl.fertility_and_life_expectancy_2014} \tab \link[=dasl.fertility_and_life_expectancy_2014]{Fertility and life expectancy 2014} \tab  \cr
\code{dasl.FALE2014} \tab \link[=dasl.fertility_and_life_expectancy_2014]{Fertility and life expectancy 2014} \tab  \cr
\code{dasl.fertilizers} \tab \link[=dasl.fertilizers]{Fertilizers} \tab  \cr
\code{dasl.fingers_and_heights} \tab \link[=dasl.fingers_and_heights]{Fingers and Heights} \tab  \cr
\code{dasl.FAH} \tab \link[=dasl.fingers_and_heights]{Fingers and Heights} \tab  \cr
\code{dasl.fire_sale} \tab \link[=dasl.fire_sale]{Fire sale} \tab  \cr
\code{dasl.fish_diet} \tab \link[=dasl.fish_diet]{Fish diet} \tab  \cr
\code{dasl.fishers_irises} \tab \link[=dasl.fishers_irises]{Fisher&#8217;s Irises} \tab  \cr
\code{dasl.fitness_data} \tab \link[=dasl.fitness_data]{Fitness data} \tab  \cr
\code{dasl.flights_2016} \tab \link[=dasl.flights_2016]{Flights 2016} \tab  \cr
\code{dasl.flights_on_time_2016} \tab \link[=dasl.flights_on_time_2016]{Flights on time 2016} \tab  \cr
\code{dasl.FOT2016} \tab \link[=dasl.flights_on_time_2016]{Flights on time 2016} \tab  \cr
\code{dasl.floods_2015} \tab \link[=dasl.floods_2015]{Floods 2015} \tab  \cr
\code{dasl.food_consumption} \tab \link[=dasl.food_consumption]{Food consumption} \tab  \cr
\code{dasl.food_sales} \tab \link[=dasl.food_sales]{Food sales} \tab  \cr
\code{dasl.football_salaries_2017} \tab \link[=dasl.football_salaries_2017]{Football salaries 2017} \tab  \cr
\code{dasl.FS2017} \tab \link[=dasl.football_salaries_2017]{Football salaries 2017} \tab  \cr
\code{dasl.foreclosure_rates_2008} \tab \link[=dasl.foreclosure_rates_2008]{Foreclosure rates 2008} \tab  \cr
\code{dasl.FR2008} \tab \link[=dasl.foreclosure_rates_2008]{Foreclosure rates 2008} \tab  \cr
\code{dasl.foreclosures_2016} \tab \link[=dasl.foreclosures_2016]{Foreclosures 2016} \tab  \cr
\code{dasl.framingham} \tab \link[=dasl.framingham]{Framingham} \tab  \cr
\code{dasl.freshman_15} \tab \link[=dasl.freshman_15]{Freshman 15} \tab  \cr
\code{dasl.friday_the_13th_accidents} \tab \link[=dasl.friday_the_13th_accidents]{Friday the 13th Accidents} \tab  \cr
\code{dasl.FT1A} \tab \link[=dasl.friday_the_13th_accidents]{Friday the 13th Accidents} \tab  \cr
\code{dasl.friday_the_13th_traffic} \tab \link[=dasl.friday_the_13th_traffic]{Friday the 13th traffic} \tab  \cr
\code{dasl.FT1T} \tab \link[=dasl.friday_the_13th_traffic]{Friday the 13th traffic} \tab  \cr
\code{dasl.frisbee_throws} \tab \link[=dasl.frisbee_throws]{Frisbee throws} \tab  \cr
\code{dasl.fritos} \tab \link[=dasl.fritos]{Fritos} \tab  \cr
\code{dasl.fudge} \tab \link[=dasl.fudge]{Fudge} \tab  \cr
\code{dasl.fuel_economy} \tab \link[=dasl.fuel_economy]{Fuel economy} \tab  \cr
\code{dasl.fuel_economy_2016} \tab \link[=dasl.fuel_economy_2016]{Fuel economy 2016} \tab  \cr
\code{dasl.FE2016} \tab \link[=dasl.fuel_economy_2016]{Fuel economy 2016} \tab  \cr
\code{dasl.fuel_economy_and_cylinders} \tab \link[=dasl.fuel_economy_and_cylinders]{Fuel economy and cylinders} \tab  \cr
\code{dasl.FEAC} \tab \link[=dasl.fuel_economy_and_cylinders]{Fuel economy and cylinders} \tab  \cr
\code{dasl.fuel_economy_sample_2016} \tab \link[=dasl.fuel_economy_sample_2016]{Fuel economy sample 2016} \tab  \cr
\code{dasl.FES2016} \tab \link[=dasl.fuel_economy_sample_2016]{Fuel economy sample 2016} \tab  \cr
\code{dasl.fuel_efficiency} \tab \link[=dasl.fuel_efficiency]{Fuel efficiency} \tab  \cr
\code{dasl.full_moon} \tab \link[=dasl.full_moon]{Full moon} \tab  \cr
\code{dasl.gdp_and_djia_2017} \tab \link[=dasl.gdp_and_djia_2017]{GDP and DJIA 2017} \tab  \cr
\code{dasl.GAD2017} \tab \link[=dasl.gdp_and_djia_2017]{GDP and DJIA 2017} \tab  \cr
\code{dasl.gdp_by_state} \tab \link[=dasl.gdp_by_state]{GDP by state} \tab  \cr
\code{dasl.GBS} \tab \link[=dasl.gdp_by_state]{GDP by state} \tab  \cr
\code{dasl.gdp_growth_2017} \tab \link[=dasl.gdp_growth_2017]{GDP growth 2017} \tab  \cr
\code{dasl.GG2017} \tab \link[=dasl.gdp_growth_2017]{GDP growth 2017} \tab  \cr
\code{dasl.gas_prices_2017} \tab \link[=dasl.gas_prices_2017]{Gas Prices 2017} \tab  \cr
\code{dasl.GP2017} \tab \link[=dasl.gas_prices_2017]{Gas Prices 2017} \tab  \cr
\code{dasl.gas_additives} \tab \link[=dasl.gas_additives]{Gas additives} \tab  \cr
\code{dasl.gas_prices_2016} \tab \link[=dasl.gas_prices_2016]{Gas prices 2016} \tab  \cr
\code{dasl.GP2016} \tab \link[=dasl.gas_prices_2016]{Gas prices 2016} \tab  \cr
\code{dasl.gas_prices_2018} \tab \link[=dasl.gas_prices_2018]{Gas prices 2018} \tab  \cr
\code{dasl.GP2018} \tab \link[=dasl.gas_prices_2018]{Gas prices 2018} \tab  \cr
\code{dasl.gas_prices_monthly} \tab \link[=dasl.gas_prices_monthly]{Gas prices monthly} \tab  \cr
\code{dasl.GPM} \tab \link[=dasl.gas_prices_monthly]{Gas prices monthly} \tab  \cr
\code{dasl.gasoline} \tab \link[=dasl.gasoline]{Gasoline} \tab  \cr
\code{dasl.gators} \tab \link[=dasl.gators]{Gators} \tab  \cr
\code{dasl.gemstones} \tab \link[=dasl.gemstones]{Gemstones} \tab  \cr
\code{dasl.gestation} \tab \link[=dasl.gestation]{Gestation} \tab  \cr
\code{dasl.global} \tab \link[=dasl.global]{Global} \tab  \cr
\code{dasl.global500_2014} \tab \link[=dasl.global500_2014]{Global500 2014} \tab  \cr
\code{dasl.golf_courses} \tab \link[=dasl.golf_courses]{Golf courses} \tab  \cr
\code{dasl.golf_drives_2015} \tab \link[=dasl.golf_drives_2015]{Golf drives 2015} \tab  \cr
\code{dasl.GD2015} \tab \link[=dasl.golf_drives_2015]{Golf drives 2015} \tab  \cr
\code{dasl.golfers_2017} \tab \link[=dasl.golfers_2017]{Golfers 2017} \tab  \cr
\code{dasl.google_stock_prices} \tab \link[=dasl.google_stock_prices]{Google stock prices} \tab  \cr
\code{dasl.GSP} \tab \link[=dasl.google_stock_prices]{Google stock prices} \tab  \cr
\code{dasl.grades} \tab \link[=dasl.grades]{Grades} \tab  \cr
\code{dasl.graduate_earnings} \tab \link[=dasl.graduate_earnings]{Graduate Earnings} \tab  \cr
\code{dasl.graphite_production} \tab \link[=dasl.graphite_production]{Graphite production} \tab  \cr
\code{dasl.grocery_prices} \tab \link[=dasl.grocery_prices]{Grocery prices} \tab  \cr
\code{dasl.gross_domestic_product} \tab \link[=dasl.gross_domestic_product]{Gross domestic product} \tab  \cr
\code{dasl.GDP} \tab \link[=dasl.gross_domestic_product]{Gross domestic product} \tab  \cr
\code{dasl.hdi_2015} \tab \link[=dasl.hdi_2015]{HDI 2015} \tab  \cr
\code{dasl.hdi_2016} \tab \link[=dasl.hdi_2016]{HDI 2016} \tab  \cr
\code{dasl.hand_dexterity} \tab \link[=dasl.hand_dexterity]{Hand dexterity} \tab  \cr
\code{dasl.hand_washing} \tab \link[=dasl.hand_washing]{Hand washing} \tab  \cr
\code{dasl.hard_water} \tab \link[=dasl.hard_water]{Hard water} \tab  \cr
\code{dasl.harry_potter_revenue} \tab \link[=dasl.harry_potter_revenue]{Harry Potter revenue} \tab  \cr
\code{dasl.HPR} \tab \link[=dasl.harry_potter_revenue]{Harry Potter revenue} \tab  \cr
\code{dasl.hawaii_tourism} \tab \link[=dasl.hawaii_tourism]{Hawaii tourism} \tab  \cr
\code{dasl.health_expenditures} \tab \link[=dasl.health_expenditures]{Health expenditures} \tab  \cr
\code{dasl.hearing} \tab \link[=dasl.hearing]{Hearing} \tab  \cr
\code{dasl.hearing_4_lists} \tab \link[=dasl.hearing_4_lists]{Hearing 4 lists} \tab  \cr
\code{dasl.H4L} \tab \link[=dasl.hearing_4_lists]{Hearing 4 lists} \tab  \cr
\code{dasl.heart_attack_charges} \tab \link[=dasl.heart_attack_charges]{Heart attack charges} \tab  \cr
\code{dasl.HAC} \tab \link[=dasl.heart_attack_charges]{Heart attack charges} \tab  \cr
\code{dasl.heart_attack_stays} \tab \link[=dasl.heart_attack_stays]{Heart attack stays} \tab  \cr
\code{dasl.HAS} \tab \link[=dasl.heart_attack_stays]{Heart attack stays} \tab  \cr
\code{dasl.heights_and_weights} \tab \link[=dasl.heights_and_weights]{Heights and weights} \tab  \cr
\code{dasl.HAW} \tab \link[=dasl.heights_and_weights]{Heights and weights} \tab  \cr
\code{dasl.historical_oil_prices_2016} \tab \link[=dasl.historical_oil_prices_2016]{Historical Oil Prices 2016} \tab  \cr
\code{dasl.HOP2016} \tab \link[=dasl.historical_oil_prices_2016]{Historical Oil Prices 2016} \tab  \cr
\code{dasl.holiday_shopping} \tab \link[=dasl.holiday_shopping]{Holiday shopping} \tab  \cr
\code{dasl.holiday_spending} \tab \link[=dasl.holiday_spending]{Holiday spending} \tab  \cr
\code{dasl.home_price_index_2017} \tab \link[=dasl.home_price_index_2017]{Home Price Index 2017} \tab  \cr
\code{dasl.HPI2017} \tab \link[=dasl.home_price_index_2017]{Home Price Index 2017} \tab  \cr
\code{dasl.home_depot_sales} \tab \link[=dasl.home_depot_sales]{Home depot sales} \tab  \cr
\code{dasl.HDS} \tab \link[=dasl.home_depot_sales]{Home depot sales} \tab  \cr
\code{dasl.hopkins_forest} \tab \link[=dasl.hopkins_forest]{Hopkins Forest} \tab  \cr
\code{dasl.hormones} \tab \link[=dasl.hormones]{Hormones} \tab  \cr
\code{dasl.hot_dogs} \tab \link[=dasl.hot_dogs]{Hot dogs} \tab  \cr
\code{dasl.hotel_maids} \tab \link[=dasl.hotel_maids]{Hotel maids} \tab  \cr
\code{dasl.housing_prices} \tab \link[=dasl.housing_prices]{Housing prices} \tab  \cr
\code{dasl.housing_prices_ge19} \tab \link[=dasl.housing_prices_ge19]{Housing prices GE19} \tab  \cr
\code{dasl.HPG} \tab \link[=dasl.housing_prices_ge19]{Housing prices GE19} \tab  \cr
\code{dasl.how_old_is_that_tree} \tab \link[=dasl.how_old_is_that_tree]{How old is that Tree} \tab  \cr
\code{dasl.HOITT} \tab \link[=dasl.how_old_is_that_tree]{How old is that Tree} \tab  \cr
\code{dasl.hurricane_frequencies} \tab \link[=dasl.hurricane_frequencies]{Hurricane frequencies} \tab  \cr
\code{dasl.hurricane_history} \tab \link[=dasl.hurricane_history]{Hurricane history} \tab  \cr
\code{dasl.hurricanes_2015} \tab \link[=dasl.hurricanes_2015]{Hurricanes 2015} \tab  \cr
\code{dasl.iq_brain} \tab \link[=dasl.iq_brain]{IQ Brain} \tab  \cr
\code{dasl.igf} \tab \link[=dasl.igf]{Igf} \tab  \cr
\code{dasl.igf13} \tab \link[=dasl.igf13]{Igf13} \tab  \cr
\code{dasl.illiad_injuries} \tab \link[=dasl.illiad_injuries]{Illiad Injuries} \tab  \cr
\code{dasl.illiad_weapons} \tab \link[=dasl.illiad_weapons]{Illiad weapons} \tab  \cr
\code{dasl.income_and_internet} \tab \link[=dasl.income_and_internet]{Income and Internet} \tab  \cr
\code{dasl.IAI} \tab \link[=dasl.income_and_internet]{Income and Internet} \tab  \cr
\code{dasl.income_and_housing} \tab \link[=dasl.income_and_housing]{Income and housing} \tab  \cr
\code{dasl.IAH} \tab \link[=dasl.income_and_housing]{Income and housing} \tab  \cr
\code{dasl.income_and_pets} \tab \link[=dasl.income_and_pets]{Income and pets} \tab  \cr
\code{dasl.IAP} \tab \link[=dasl.income_and_pets]{Income and pets} \tab  \cr
\code{dasl.income_vs_hours_2013} \tab \link[=dasl.income_vs_hours_2013]{Income vs Hours 2013} \tab  \cr
\code{dasl.IVH2013} \tab \link[=dasl.income_vs_hours_2013]{Income vs Hours 2013} \tab  \cr
\code{dasl.industrial_experiment} \tab \link[=dasl.industrial_experiment]{Industrial experiment} \tab  \cr
\code{dasl.indy_500_2016} \tab \link[=dasl.indy_500_2016]{Indy 500 2016} \tab  \cr
\code{dasl.I5002016} \tab \link[=dasl.indy_500_2016]{Indy 500 2016} \tab  \cr
\code{dasl.indy_500_2017} \tab \link[=dasl.indy_500_2017]{Indy 500 2017} \tab  \cr
\code{dasl.I5002017} \tab \link[=dasl.indy_500_2017]{Indy 500 2017} \tab  \cr
\code{dasl.indy_500_2018} \tab \link[=dasl.indy_500_2018]{Indy 500 2018} \tab  \cr
\code{dasl.I5002018} \tab \link[=dasl.indy_500_2018]{Indy 500 2018} \tab  \cr
\code{dasl.inflation_2016} \tab \link[=dasl.inflation_2016]{Inflation 2016} \tab  \cr
\code{dasl.insurance_life_expectancy} \tab \link[=dasl.insurance_life_expectancy]{Insurance life expectancy} \tab  \cr
\code{dasl.ILE} \tab \link[=dasl.insurance_life_expectancy]{Insurance life expectancy} \tab  \cr
\code{dasl.insurance_policies} \tab \link[=dasl.insurance_policies]{Insurance policies} \tab  \cr
\code{dasl.insurance_profits} \tab \link[=dasl.insurance_profits]{Insurance profits} \tab  \cr
\code{dasl.insurance_profits_sales_rep} \tab \link[=dasl.insurance_profits_sales_rep]{Insurance profits sales rep} \tab  \cr
\code{dasl.IPSR} \tab \link[=dasl.insurance_profits_sales_rep]{Insurance profits sales rep} \tab  \cr
\code{dasl.interest_rates_2009} \tab \link[=dasl.interest_rates_2009]{Interest rates 2009} \tab  \cr
\code{dasl.IR2009} \tab \link[=dasl.interest_rates_2009]{Interest rates 2009} \tab  \cr
\code{dasl.interest_rates_and_mortgages} \tab \link[=dasl.interest_rates_and_mortgages]{Interest rates and mortgages} \tab  \cr
\code{dasl.IRAM} \tab \link[=dasl.interest_rates_and_mortgages]{Interest rates and mortgages} \tab  \cr
\code{dasl.interest_rates_and_mortgages_2015} \tab \link[=dasl.interest_rates_and_mortgages_2015]{Interest rates and mortgages 2015} \tab  \cr
\code{dasl.IRAM2015} \tab \link[=dasl.interest_rates_and_mortgages_2015]{Interest rates and mortgages 2015} \tab  \cr
\code{dasl.internet_users_2014} \tab \link[=dasl.internet_users_2014]{Internet users 2014} \tab  \cr
\code{dasl.IU2014} \tab \link[=dasl.internet_users_2014]{Internet users 2014} \tab  \cr
\code{dasl.jcrew} \tab \link[=dasl.jcrew]{JCrew} \tab  \cr
\code{dasl.job_growth} \tab \link[=dasl.job_growth]{Job Growth} \tab  \cr
\code{dasl.job_satisfaction} \tab \link[=dasl.job_satisfaction]{Job satisfaction} \tab  \cr
\code{dasl.kentucky_derby_2016} \tab \link[=dasl.kentucky_derby_2016]{Kentucky Derby 2016} \tab  \cr
\code{dasl.KD2016} \tab \link[=dasl.kentucky_derby_2016]{Kentucky Derby 2016} \tab  \cr
\code{dasl.kentucky_derby_2017} \tab \link[=dasl.kentucky_derby_2017]{Kentucky Derby 2017} \tab  \cr
\code{dasl.KD2017} \tab \link[=dasl.kentucky_derby_2017]{Kentucky Derby 2017} \tab  \cr
\code{dasl.kentucky_derby_2018} \tab \link[=dasl.kentucky_derby_2018]{Kentucky Derby 2018} \tab  \cr
\code{dasl.KD2018} \tab \link[=dasl.kentucky_derby_2018]{Kentucky Derby 2018} \tab  \cr
\code{dasl.komtek_technologies} \tab \link[=dasl.komtek_technologies]{Komtek Technologies} \tab  \cr
\code{dasl.la_rainfall} \tab \link[=dasl.la_rainfall]{LA rainfall} \tab  \cr
\code{dasl.late_arrivals_2016} \tab \link[=dasl.late_arrivals_2016]{Late arrivals 2016} \tab  \cr
\code{dasl.LA2016} \tab \link[=dasl.late_arrivals_2016]{Late arrivals 2016} \tab  \cr
\code{dasl.legionnaires_disease} \tab \link[=dasl.legionnaires_disease]{Legionnaires disease} \tab  \cr
\code{dasl.libertys_nose} \tab \link[=dasl.libertys_nose]{Libertys nose} \tab  \cr
\code{dasl.life_expectancy} \tab \link[=dasl.life_expectancy]{Life Expectancy} \tab  \cr
\code{dasl.life_expectancy_us} \tab \link[=dasl.life_expectancy_us]{Life expectancy US} \tab  \cr
\code{dasl.LEU} \tab \link[=dasl.life_expectancy_us]{Life expectancy US} \tab  \cr
\code{dasl.little_league_product_testing} \tab \link[=dasl.little_league_product_testing]{Little league product testing} \tab  \cr
\code{dasl.LLPT} \tab \link[=dasl.little_league_product_testing]{Little league product testing} \tab  \cr
\code{dasl.load_factors_2016} \tab \link[=dasl.load_factors_2016]{Load factors 2016} \tab  \cr
\code{dasl.LF2016} \tab \link[=dasl.load_factors_2016]{Load factors 2016} \tab  \cr
\code{dasl.load_factors_2017} \tab \link[=dasl.load_factors_2017]{Load factors 2017} \tab  \cr
\code{dasl.LF2017} \tab \link[=dasl.load_factors_2017]{Load factors 2017} \tab  \cr
\code{dasl.lobsters_2016} \tab \link[=dasl.lobsters_2016]{Lobsters 2016} \tab  \cr
\code{dasl.lottery_numbers} \tab \link[=dasl.lottery_numbers]{Lottery numbers} \tab  \cr
\code{dasl.loyalty_program} \tab \link[=dasl.loyalty_program]{Loyalty program} \tab  \cr
\code{dasl.lunchtime} \tab \link[=dasl.lunchtime]{Lunchtime} \tab  \cr
\code{dasl.mpg_2016} \tab \link[=dasl.mpg_2016]{MPG 2016} \tab  \cr
\code{dasl.magnet_schools} \tab \link[=dasl.magnet_schools]{Magnet schools} \tab  \cr
\code{dasl.mail} \tab \link[=dasl.mail]{Mail} \tab  \cr
\code{dasl.major_hurricanes_2013} \tab \link[=dasl.major_hurricanes_2013]{Major hurricanes 2013} \tab  \cr
\code{dasl.MH2013} \tab \link[=dasl.major_hurricanes_2013]{Major hurricanes 2013} \tab  \cr
\code{dasl.male_unemployment_2017} \tab \link[=dasl.male_unemployment_2017]{Male unemployment 2017} \tab  \cr
\code{dasl.MU2017} \tab \link[=dasl.male_unemployment_2017]{Male unemployment 2017} \tab  \cr
\code{dasl.manatees_2015} \tab \link[=dasl.manatees_2015]{Manatees 2015} \tab  \cr
\code{dasl.marijuana_2015} \tab \link[=dasl.marijuana_2015]{Marijuana 2015} \tab  \cr
\code{dasl.market_value} \tab \link[=dasl.market_value]{Market value} \tab  \cr
\code{dasl.marketing_managers_salaries} \tab \link[=dasl.marketing_managers_salaries]{Marketing managers salaries} \tab  \cr
\code{dasl.MMS} \tab \link[=dasl.marketing_managers_salaries]{Marketing managers salaries} \tab  \cr
\code{dasl.marriage_age_2015} \tab \link[=dasl.marriage_age_2015]{Marriage age 2015} \tab  \cr
\code{dasl.MA2015} \tab \link[=dasl.marriage_age_2015]{Marriage age 2015} \tab  \cr
\code{dasl.marriage_age_2016} \tab \link[=dasl.marriage_age_2016]{Marriage age 2016} \tab  \cr
\code{dasl.MA2016} \tab \link[=dasl.marriage_age_2016]{Marriage age 2016} \tab  \cr
\code{dasl.marriage_age_2017} \tab \link[=dasl.marriage_age_2017]{Marriage age 2017} \tab  \cr
\code{dasl.MA2017} \tab \link[=dasl.marriage_age_2017]{Marriage age 2017} \tab  \cr
\code{dasl.math_scores_2013} \tab \link[=dasl.math_scores_2013]{Math scores 2013} \tab  \cr
\code{dasl.MS2013} \tab \link[=dasl.math_scores_2013]{Math scores 2013} \tab  \cr
\code{dasl.maze} \tab \link[=dasl.maze]{Maze} \tab  \cr
\code{dasl.mazes_and_smells} \tab \link[=dasl.mazes_and_smells]{Mazes and smells} \tab  \cr
\code{dasl.MAS} \tab \link[=dasl.mazes_and_smells]{Mazes and smells} \tab  \cr
\code{dasl.memory} \tab \link[=dasl.memory]{Memory} \tab  \cr
\code{dasl.mens_weekly_earnings_2013} \tab \link[=dasl.mens_weekly_earnings_2013]{Men&#8217;s weekly earnings 2013} \tab  \cr
\code{dasl.MWE2013} \tab \link[=dasl.mens_weekly_earnings_2013]{Men&#8217;s weekly earnings 2013} \tab  \cr
\code{dasl.mens_weights} \tab \link[=dasl.mens_weights]{Mens Weights} \tab  \cr
\code{dasl.michelson} \tab \link[=dasl.michelson]{Michelson} \tab  \cr
\code{dasl.midterms} \tab \link[=dasl.midterms]{Midterms} \tab  \cr
\code{dasl.minimum_wage_workers} \tab \link[=dasl.minimum_wage_workers]{Minimum Wage Workers} \tab  \cr
\code{dasl.MWW} \tab \link[=dasl.minimum_wage_workers]{Minimum Wage Workers} \tab  \cr
\code{dasl.montana} \tab \link[=dasl.montana]{Montana} \tab  \cr
\code{dasl.montana_revisited} \tab \link[=dasl.montana_revisited]{Montana revisited} \tab  \cr
\code{dasl.mothers_and_fathers_1965_2011} \tab \link[=dasl.mothers_and_fathers_1965_2011]{Mothers and fathers 1965-2011} \tab  \cr
\code{dasl.MAF19652011} \tab \link[=dasl.mothers_and_fathers_1965_2011]{Mothers and fathers 1965-2011} \tab  \cr
\code{dasl.mothers_and_fathers_aspirations} \tab \link[=dasl.mothers_and_fathers_aspirations]{Mothers and fathers aspirations} \tab  \cr
\code{dasl.MAFA} \tab \link[=dasl.mothers_and_fathers_aspirations]{Mothers and fathers aspirations} \tab  \cr
\code{dasl.mouth_volume} \tab \link[=dasl.mouth_volume]{Mouth volume} \tab  \cr
\code{dasl.movie_admissions_2016} \tab \link[=dasl.movie_admissions_2016]{Movie Admissions 2016} \tab  \cr
\code{dasl.movie_budgets} \tab \link[=dasl.movie_budgets]{Movie budgets} \tab  \cr
\code{dasl.movie_lengths_2010} \tab \link[=dasl.movie_lengths_2010]{Movie lengths 2010} \tab  \cr
\code{dasl.ML2010} \tab \link[=dasl.movie_lengths_2010]{Movie lengths 2010} \tab  \cr
\code{dasl.movie_profits} \tab \link[=dasl.movie_profits]{Movie profits} \tab  \cr
\code{dasl.mozart} \tab \link[=dasl.mozart]{Mozart} \tab  \cr
\code{dasl.music_library} \tab \link[=dasl.music_library]{Music library} \tab  \cr
\code{dasl.mutual_fund_flows} \tab \link[=dasl.mutual_fund_flows]{Mutual fund flows} \tab  \cr
\code{dasl.MFF} \tab \link[=dasl.mutual_fund_flows]{Mutual fund flows} \tab  \cr
\code{dasl.mutual_fund_returns_2017} \tab \link[=dasl.mutual_fund_returns_2017]{Mutual fund returns 2017} \tab  \cr
\code{dasl.MFR2017} \tab \link[=dasl.mutual_fund_returns_2017]{Mutual fund returns 2017} \tab  \cr
\code{dasl.mutual_funds_2016} \tab \link[=dasl.mutual_funds_2016]{Mutual funds 2016} \tab  \cr
\code{dasl.MF2016} \tab \link[=dasl.mutual_funds_2016]{Mutual funds 2016} \tab  \cr
\code{dasl.nhanes} \tab \link[=dasl.nhanes]{NHANES} \tab  \cr
\code{dasl.ny_marathon_2016} \tab \link[=dasl.ny_marathon_2016]{NY Marathon 2016} \tab  \cr
\code{dasl.NM2016} \tab \link[=dasl.ny_marathon_2016]{NY Marathon 2016} \tab  \cr
\code{dasl.nypd} \tab \link[=dasl.nypd]{NYPD} \tab  \cr
\code{dasl.nail_polish} \tab \link[=dasl.nail_polish]{Nail polish} \tab  \cr
\code{dasl.nambe} \tab \link[=dasl.nambe]{Nambe} \tab  \cr
\code{dasl.nambe_2018} \tab \link[=dasl.nambe_2018]{Nambe 2018} \tab  \cr
\code{dasl.nashville} \tab \link[=dasl.nashville]{Nashville} \tab  \cr
\code{dasl.neck_size} \tab \link[=dasl.neck_size]{Neck size} \tab  \cr
\code{dasl.nenana_2017} \tab \link[=dasl.nenana_2017]{Nenana 2017} \tab  \cr
\code{dasl.nest_egg_index} \tab \link[=dasl.nest_egg_index]{Nest Egg Index} \tab  \cr
\code{dasl.NEI} \tab \link[=dasl.nest_egg_index]{Nest Egg Index} \tab  \cr
\code{dasl.new_york_bridges_2016} \tab \link[=dasl.new_york_bridges_2016]{New York bridges 2016} \tab  \cr
\code{dasl.NYB2016} \tab \link[=dasl.new_york_bridges_2016]{New York bridges 2016} \tab  \cr
\code{dasl.newspapers} \tab \link[=dasl.newspapers]{Newspapers} \tab  \cr
\code{dasl.nightmares} \tab \link[=dasl.nightmares]{Nightmares} \tab  \cr
\code{dasl.nissan} \tab \link[=dasl.nissan]{Nissan} \tab  \cr
\code{dasl.normal_temperature} \tab \link[=dasl.normal_temperature]{Normal temperature} \tab  \cr
\code{dasl.nuclear_plants} \tab \link[=dasl.nuclear_plants]{Nuclear plants} \tab  \cr
\code{dasl.oecd_gdp} \tab \link[=dasl.oecd_gdp]{OECD GDP} \tab  \cr
\code{dasl.oecd_gdp_growth} \tab \link[=dasl.oecd_gdp_growth]{OECD GDP Growth} \tab  \cr
\code{dasl.OGG} \tab \link[=dasl.oecd_gdp_growth]{OECD GDP Growth} \tab  \cr
\code{dasl.oecd_unemployment} \tab \link[=dasl.oecd_unemployment]{OECD Unemployment} \tab  \cr
\code{dasl.oecd_economic_regulations} \tab \link[=dasl.oecd_economic_regulations]{OECD economic regulations} \tab  \cr
\code{dasl.OER} \tab \link[=dasl.oecd_economic_regulations]{OECD economic regulations} \tab  \cr
\code{dasl.okcupid_catsdogs} \tab \link[=dasl.okcupid_catsdogs]{OKCupid CatsDogs} \tab  \cr
\code{dasl.oakland_passengers_2016} \tab \link[=dasl.oakland_passengers_2016]{Oakland passengers 2016} \tab  \cr
\code{dasl.OP2016} \tab \link[=dasl.oakland_passengers_2016]{Oakland passengers 2016} \tab  \cr
\code{dasl.obesity_and_exercise} \tab \link[=dasl.obesity_and_exercise]{Obesity and exercise} \tab  \cr
\code{dasl.OAE} \tab \link[=dasl.obesity_and_exercise]{Obesity and exercise} \tab  \cr
\code{dasl.oil_prices_2016} \tab \link[=dasl.oil_prices_2016]{Oil prices 2016} \tab  \cr
\code{dasl.oil_prices_monthly} \tab \link[=dasl.oil_prices_monthly]{Oil prices monthly} \tab  \cr
\code{dasl.OPM} \tab \link[=dasl.oil_prices_monthly]{Oil prices monthly} \tab  \cr
\code{dasl.old_faithful} \tab \link[=dasl.old_faithful]{Old Faithful} \tab  \cr
\code{dasl.old_pitchers} \tab \link[=dasl.old_pitchers]{Old pitchers} \tab  \cr
\code{dasl.olympic_jumps_2016} \tab \link[=dasl.olympic_jumps_2016]{Olympic jumps 2016} \tab  \cr
\code{dasl.OJ2016} \tab \link[=dasl.olympic_jumps_2016]{Olympic jumps 2016} \tab  \cr
\code{dasl.online_shopping} \tab \link[=dasl.online_shopping]{Online Shopping} \tab  \cr
\code{dasl.online_insurance} \tab \link[=dasl.online_insurance]{Online insurance} \tab  \cr
\code{dasl.oranges} \tab \link[=dasl.oranges]{Oranges} \tab  \cr
\code{dasl.ozone} \tab \link[=dasl.ozone]{Ozone} \tab  \cr
\code{dasl.pva} \tab \link[=dasl.pva]{PVA} \tab  \cr
\code{dasl.paper_airplanes} \tab \link[=dasl.paper_airplanes]{Paper airplanes} \tab  \cr
\code{dasl.paralyzed_veterans} \tab \link[=dasl.paralyzed_veterans]{Paralyzed veterans} \tab  \cr
\code{dasl.passive_smoke} \tab \link[=dasl.passive_smoke]{Passive smoke} \tab  \cr
\code{dasl.pendulum} \tab \link[=dasl.pendulum]{Pendulum} \tab  \cr
\code{dasl.penguins} \tab \link[=dasl.penguins]{Penguins} \tab  \cr
\code{dasl.personal_appearance} \tab \link[=dasl.personal_appearance]{Personal appearance} \tab  \cr
\code{dasl.pi} \tab \link[=dasl.pi]{Pi} \tab  \cr
\code{dasl.pima_indians} \tab \link[=dasl.pima_indians]{Pima indians} \tab  \cr
\code{dasl.pizza_prices} \tab \link[=dasl.pizza_prices]{Pizza prices} \tab  \cr
\code{dasl.pizza_ratings} \tab \link[=dasl.pizza_ratings]{Pizza ratings} \tab  \cr
\code{dasl.planets} \tab \link[=dasl.planets]{Planets} \tab  \cr
\code{dasl.planets_more_data} \tab \link[=dasl.planets_more_data]{Planets more data} \tab  \cr
\code{dasl.PMD} \tab \link[=dasl.planets_more_data]{Planets more data} \tab  \cr
\code{dasl.police_pay_2016} \tab \link[=dasl.police_pay_2016]{Police pay 2016} \tab  \cr
\code{dasl.PP2016} \tab \link[=dasl.police_pay_2016]{Police pay 2016} \tab  \cr
\code{dasl.polling} \tab \link[=dasl.polling]{Polling} \tab  \cr
\code{dasl.popcorn} \tab \link[=dasl.popcorn]{Popcorn} \tab  \cr
\code{dasl.population_commute_times} \tab \link[=dasl.population_commute_times]{Population Commute Times} \tab  \cr
\code{dasl.PCT} \tab \link[=dasl.population_commute_times]{Population Commute Times} \tab  \cr
\code{dasl.population_growth_2010} \tab \link[=dasl.population_growth_2010]{Population growth 2010} \tab  \cr
\code{dasl.PG2010} \tab \link[=dasl.population_growth_2010]{Population growth 2010} \tab  \cr
\code{dasl.poverty_and_region_2015} \tab \link[=dasl.poverty_and_region_2015]{Poverty and Region 2015} \tab  \cr
\code{dasl.PAR2015} \tab \link[=dasl.poverty_and_region_2015]{Poverty and Region 2015} \tab  \cr
\code{dasl.pregnancies} \tab \link[=dasl.pregnancies]{Pregnancies} \tab  \cr
\code{dasl.prices_and_earnings} \tab \link[=dasl.prices_and_earnings]{Prices and Earnings} \tab  \cr
\code{dasl.PAE} \tab \link[=dasl.prices_and_earnings]{Prices and Earnings} \tab  \cr
\code{dasl.prisons_2014} \tab \link[=dasl.prisons_2014]{Prisons 2014} \tab  \cr
\code{dasl.process_improvements} \tab \link[=dasl.process_improvements]{Process improvements} \tab  \cr
\code{dasl.product_placement} \tab \link[=dasl.product_placement]{Product placement} \tab  \cr
\code{dasl.product_testing} \tab \link[=dasl.product_testing]{Product testing} \tab  \cr
\code{dasl.production_problems} \tab \link[=dasl.production_problems]{Production problems} \tab  \cr
\code{dasl.productivity_2016} \tab \link[=dasl.productivity_2016]{Productivity 2016} \tab  \cr
\code{dasl.property_assessment} \tab \link[=dasl.property_assessment]{Property assessment} \tab  \cr
\code{dasl.pulse_rates} \tab \link[=dasl.pulse_rates]{Pulse rates} \tab  \cr
\code{dasl.push_ups} \tab \link[=dasl.push_ups]{Push ups} \tab  \cr
\code{dasl.qc_clarity} \tab \link[=dasl.qc_clarity]{QC Clarity} \tab  \cr
\code{dasl.qc_control} \tab \link[=dasl.qc_control]{QC Control} \tab  \cr
\code{dasl.qc_production} \tab \link[=dasl.qc_production]{QC Production} \tab  \cr
\code{dasl.quality_control_holes} \tab \link[=dasl.quality_control_holes]{Quality control holes} \tab  \cr
\code{dasl.QCH} \tab \link[=dasl.quality_control_holes]{Quality control holes} \tab  \cr
\code{dasl.quarterback_performance_2017} \tab \link[=dasl.quarterback_performance_2017]{Quarterback performance 2017} \tab  \cr
\code{dasl.QP2017} \tab \link[=dasl.quarterback_performance_2017]{Quarterback performance 2017} \tab  \cr
\code{dasl.racial_steering} \tab \link[=dasl.racial_steering]{Racial steering} \tab  \cr
\code{dasl.reading} \tab \link[=dasl.reading]{Reading} \tab  \cr
\code{dasl.real_estate} \tab \link[=dasl.real_estate]{Real Estate} \tab  \cr
\code{dasl.real_estate_sample_1200} \tab \link[=dasl.real_estate_sample_1200]{Real estate sample 1200} \tab  \cr
\code{dasl.RES1200} \tab \link[=dasl.real_estate_sample_1200]{Real estate sample 1200} \tab  \cr
\code{dasl.receivers_2015} \tab \link[=dasl.receivers_2015]{Receivers 2015} \tab  \cr
\code{dasl.regular_gas_2017} \tab \link[=dasl.regular_gas_2017]{Regular gas 2017} \tab  \cr
\code{dasl.RG2017} \tab \link[=dasl.regular_gas_2017]{Regular gas 2017} \tab  \cr
\code{dasl.retail_trade_index} \tab \link[=dasl.retail_trade_index]{Retail trade index} \tab  \cr
\code{dasl.RTI} \tab \link[=dasl.retail_trade_index]{Retail trade index} \tab  \cr
\code{dasl.rip_currents_2013} \tab \link[=dasl.rip_currents_2013]{Rip currents 2013} \tab  \cr
\code{dasl.RC2013} \tab \link[=dasl.rip_currents_2013]{Rip currents 2013} \tab  \cr
\code{dasl.rock_concert_deaths} \tab \link[=dasl.rock_concert_deaths]{Rock concert deaths} \tab  \cr
\code{dasl.RCD} \tab \link[=dasl.rock_concert_deaths]{Rock concert deaths} \tab  \cr
\code{dasl.ruffles} \tab \link[=dasl.ruffles]{Ruffles} \tab  \cr
\code{dasl.rulers} \tab \link[=dasl.rulers]{Rulers} \tab  \cr
\code{dasl.run_times} \tab \link[=dasl.run_times]{Run times} \tab  \cr
\code{dasl.running_heats_london} \tab \link[=dasl.running_heats_london]{Running heats London} \tab  \cr
\code{dasl.RHL} \tab \link[=dasl.running_heats_london]{Running heats London} \tab  \cr
\code{dasl.sat_scores} \tab \link[=dasl.sat_scores]{SAT scores} \tab  \cr
\code{dasl.sales_and_profits} \tab \link[=dasl.sales_and_profits]{Sales and profits} \tab  \cr
\code{dasl.SAP} \tab \link[=dasl.sales_and_profits]{Sales and profits} \tab  \cr
\code{dasl.salmon} \tab \link[=dasl.salmon]{Salmon} \tab  \cr
\code{dasl.saratoga_house_prices} \tab \link[=dasl.saratoga_house_prices]{Saratoga house prices} \tab  \cr
\code{dasl.SHP} \tab \link[=dasl.saratoga_house_prices]{Saratoga house prices} \tab  \cr
\code{dasl.saratoga_houses} \tab \link[=dasl.saratoga_houses]{Saratoga houses} \tab  \cr
\code{dasl.school_system} \tab \link[=dasl.school_system]{School system} \tab  \cr
\code{dasl.scottish_hill_races} \tab \link[=dasl.scottish_hill_races]{Scottish Hill Races} \tab  \cr
\code{dasl.SHR} \tab \link[=dasl.scottish_hill_races]{Scottish Hill Races} \tab  \cr
\code{dasl.sea_ice} \tab \link[=dasl.sea_ice]{Sea ice} \tab  \cr
\code{dasl.seasonal_spending} \tab \link[=dasl.seasonal_spending]{Seasonal spending} \tab  \cr
\code{dasl.seat_belts_2015} \tab \link[=dasl.seat_belts_2015]{Seat belts 2015} \tab  \cr
\code{dasl.SB2015} \tab \link[=dasl.seat_belts_2015]{Seat belts 2015} \tab  \cr
\code{dasl.sex_sells} \tab \link[=dasl.sex_sells]{Sex sells} \tab  \cr
\code{dasl.shirt_sizes} \tab \link[=dasl.shirt_sizes]{Shirt sizes} \tab  \cr
\code{dasl.shirt_sizes_full} \tab \link[=dasl.shirt_sizes_full]{Shirt sizes full} \tab  \cr
\code{dasl.SSF} \tab \link[=dasl.shirt_sizes_full]{Shirt sizes full} \tab  \cr
\code{dasl.shoe_sizes} \tab \link[=dasl.shoe_sizes]{Shoe Sizes} \tab  \cr
\code{dasl.shoot_to_score_2016} \tab \link[=dasl.shoot_to_score_2016]{Shoot to Score 2016} \tab  \cr
\code{dasl.STS2016} \tab \link[=dasl.shoot_to_score_2016]{Shoot to Score 2016} \tab  \cr
\code{dasl.singers_by_parts} \tab \link[=dasl.singers_by_parts]{Singers by parts} \tab  \cr
\code{dasl.SBP} \tab \link[=dasl.singers_by_parts]{Singers by parts} \tab  \cr
\code{dasl.ski_wax} \tab \link[=dasl.ski_wax]{Ski wax} \tab  \cr
\code{dasl.slalom_times_2014} \tab \link[=dasl.slalom_times_2014]{Slalom times 2014} \tab  \cr
\code{dasl.ST2014} \tab \link[=dasl.slalom_times_2014]{Slalom times 2014} \tab  \cr
\code{dasl.slalom_times_2018} \tab \link[=dasl.slalom_times_2018]{Slalom times 2018} \tab  \cr
\code{dasl.ST2018} \tab \link[=dasl.slalom_times_2018]{Slalom times 2018} \tab  \cr
\code{dasl.sleep} \tab \link[=dasl.sleep]{Sleep} \tab  \cr
\code{dasl.slower_is_cheaper} \tab \link[=dasl.slower_is_cheaper]{Slower is cheaper} \tab  \cr
\code{dasl.SIC} \tab \link[=dasl.slower_is_cheaper]{Slower is cheaper} \tab  \cr
\code{dasl.smoking_2014} \tab \link[=dasl.smoking_2014]{Smoking 2014} \tab  \cr
\code{dasl.smoking_and_pregnancy_2011} \tab \link[=dasl.smoking_and_pregnancy_2011]{Smoking and Pregnancy 2011} \tab  \cr
\code{dasl.SAP2011} \tab \link[=dasl.smoking_and_pregnancy_2011]{Smoking and Pregnancy 2011} \tab  \cr
\code{dasl.social_networking} \tab \link[=dasl.social_networking]{Social networking} \tab  \cr
\code{dasl.speed_and_density} \tab \link[=dasl.speed_and_density]{Speed and density} \tab  \cr
\code{dasl.SAD} \tab \link[=dasl.speed_and_density]{Speed and density} \tab  \cr
\code{dasl.sprouts} \tab \link[=dasl.sprouts]{Sprouts} \tab  \cr
\code{dasl.stereograms} \tab \link[=dasl.stereograms]{Stereograms} \tab  \cr
\code{dasl.stopping_distance} \tab \link[=dasl.stopping_distance]{Stopping distance} \tab  \cr
\code{dasl.stopping_distance_60} \tab \link[=dasl.stopping_distance_60]{Stopping distance 60} \tab  \cr
\code{dasl.SD60} \tab \link[=dasl.stopping_distance_60]{Stopping distance 60} \tab  \cr
\code{dasl.storks} \tab \link[=dasl.storks]{Storks} \tab  \cr
\code{dasl.streams} \tab \link[=dasl.streams]{Streams} \tab  \cr
\code{dasl.strikes} \tab \link[=dasl.strikes]{Strikes} \tab  \cr
\code{dasl.student_survey} \tab \link[=dasl.student_survey]{Student survey} \tab  \cr
\code{dasl.student_testing} \tab \link[=dasl.student_testing]{Student testing} \tab  \cr
\code{dasl.sugar_in_cereal} \tab \link[=dasl.sugar_in_cereal]{Sugar in cereal} \tab  \cr
\code{dasl.summer_olympics_2016} \tab \link[=dasl.summer_olympics_2016]{Summer Olympics 2016} \tab  \cr
\code{dasl.SO2016} \tab \link[=dasl.summer_olympics_2016]{Summer Olympics 2016} \tab  \cr
\code{dasl.summer_school} \tab \link[=dasl.summer_school]{Summer school} \tab  \cr
\code{dasl.super_bowl_2016} \tab \link[=dasl.super_bowl_2016]{Super Bowl 2016} \tab  \cr
\code{dasl.SB2016} \tab \link[=dasl.super_bowl_2016]{Super Bowl 2016} \tab  \cr
\code{dasl.swim_and_run} \tab \link[=dasl.swim_and_run]{Swim and Run} \tab  \cr
\code{dasl.SAR} \tab \link[=dasl.swim_and_run]{Swim and Run} \tab  \cr
\code{dasl.swim_the_lake_2016} \tab \link[=dasl.swim_the_lake_2016]{Swim the lake 2016} \tab  \cr
\code{dasl.STL2016} \tab \link[=dasl.swim_the_lake_2016]{Swim the lake 2016} \tab  \cr
\code{dasl.swimming_heats_2016} \tab \link[=dasl.swimming_heats_2016]{Swimming heats 2016} \tab  \cr
\code{dasl.SH2016} \tab \link[=dasl.swimming_heats_2016]{Swimming heats 2016} \tab  \cr
\code{dasl.swimming_heats_london} \tab \link[=dasl.swimming_heats_london]{Swimming heats London} \tab  \cr
\code{dasl.SHL} \tab \link[=dasl.swimming_heats_london]{Swimming heats London} \tab  \cr
\code{dasl.tbill_rates_2016} \tab \link[=dasl.tbill_rates_2016]{TBill rates 2016} \tab  \cr
\code{dasl.TR2016} \tab \link[=dasl.tbill_rates_2016]{TBill rates 2016} \tab  \cr
\code{dasl.tv_watching} \tab \link[=dasl.tv_watching]{TV watching} \tab  \cr
\code{dasl.tablet_computers_2014} \tab \link[=dasl.tablet_computers_2014]{Tablet computers 2014} \tab  \cr
\code{dasl.TC2014} \tab \link[=dasl.tablet_computers_2014]{Tablet computers 2014} \tab  \cr
\code{dasl.tableware} \tab \link[=dasl.tableware]{Tableware} \tab  \cr
\code{dasl.tattoos} \tab \link[=dasl.tattoos]{Tattoos} \tab  \cr
\code{dasl.tellers} \tab \link[=dasl.tellers]{Tellers} \tab  \cr
\code{dasl.tendon_transfers} \tab \link[=dasl.tendon_transfers]{Tendon transfers} \tab  \cr
\code{dasl.tennis_balls} \tab \link[=dasl.tennis_balls]{Tennis balls} \tab  \cr
\code{dasl.thrills_2013} \tab \link[=dasl.thrills_2013]{Thrills 2013} \tab  \cr
\code{dasl.tiffany_2017} \tab \link[=dasl.tiffany_2017]{Tiffany 2017} \tab  \cr
\code{dasl.time_on_market} \tab \link[=dasl.time_on_market]{Time on market} \tab  \cr
\code{dasl.TOM} \tab \link[=dasl.time_on_market]{Time on market} \tab  \cr
\code{dasl.titanic} \tab \link[=dasl.titanic]{Titanic} \tab  \cr
\code{dasl.tobacco_and_alcohol} \tab \link[=dasl.tobacco_and_alcohol]{Tobacco and alcohol} \tab  \cr
\code{dasl.TAA} \tab \link[=dasl.tobacco_and_alcohol]{Tobacco and alcohol} \tab  \cr
\code{dasl.tompkins_county_bridges_2016} \tab \link[=dasl.tompkins_county_bridges_2016]{Tompkins county bridges 2016} \tab  \cr
\code{dasl.TCB2016} \tab \link[=dasl.tompkins_county_bridges_2016]{Tompkins county bridges 2016} \tab  \cr
\code{dasl.tomtoms} \tab \link[=dasl.tomtoms]{Tomtoms} \tab  \cr
\code{dasl.tornadoes_2015} \tab \link[=dasl.tornadoes_2015]{Tornadoes 2015} \tab  \cr
\code{dasl.tour_de_france_2016} \tab \link[=dasl.tour_de_france_2016]{Tour de France 2016} \tab  \cr
\code{dasl.TDF2016} \tab \link[=dasl.tour_de_france_2016]{Tour de France 2016} \tab  \cr
\code{dasl.tour_de_france_2017} \tab \link[=dasl.tour_de_france_2017]{Tour de France 2017} \tab  \cr
\code{dasl.TDF2017} \tab \link[=dasl.tour_de_france_2017]{Tour de France 2017} \tab  \cr
\code{dasl.toy_makers} \tab \link[=dasl.toy_makers]{Toy makers} \tab  \cr
\code{dasl.toyota_stock_prices_2013} \tab \link[=dasl.toyota_stock_prices_2013]{Toyota stock prices 2013} \tab  \cr
\code{dasl.TSP2013} \tab \link[=dasl.toyota_stock_prices_2013]{Toyota stock prices 2013} \tab  \cr
\code{dasl.tracking_hurricanes_2015} \tab \link[=dasl.tracking_hurricanes_2015]{Tracking hurricanes 2015} \tab  \cr
\code{dasl.TH2015} \tab \link[=dasl.tracking_hurricanes_2015]{Tracking hurricanes 2015} \tab  \cr
\code{dasl.tracking_hurricanes_2016} \tab \link[=dasl.tracking_hurricanes_2016]{Tracking hurricanes 2016} \tab  \cr
\code{dasl.TH2016} \tab \link[=dasl.tracking_hurricanes_2016]{Tracking hurricanes 2016} \tab  \cr
\code{dasl.traffic_fatalities_2013} \tab \link[=dasl.traffic_fatalities_2013]{Traffic fatalities 2013} \tab  \cr
\code{dasl.TF2013} \tab \link[=dasl.traffic_fatalities_2013]{Traffic fatalities 2013} \tab  \cr
\code{dasl.transportation_energy_use} \tab \link[=dasl.transportation_energy_use]{Transportation Energy use} \tab  \cr
\code{dasl.TEU} \tab \link[=dasl.transportation_energy_use]{Transportation Energy use} \tab  \cr
\code{dasl.tree_growth} \tab \link[=dasl.tree_growth]{Tree growth} \tab  \cr
\code{dasl.trophy_sizes} \tab \link[=dasl.trophy_sizes]{Trophy sizes} \tab  \cr
\code{dasl.trucks} \tab \link[=dasl.trucks]{Trucks} \tab  \cr
\code{dasl.tsunamis_2016} \tab \link[=dasl.tsunamis_2016]{Tsunamis 2016} \tab  \cr
\code{dasl.tsunamis_2018} \tab \link[=dasl.tsunamis_2018]{Tsunamis 2018} \tab  \cr
\code{dasl.tuition_2016} \tab \link[=dasl.tuition_2016]{Tuition 2016} \tab  \cr
\code{dasl.tuition_all_schools_2016} \tab \link[=dasl.tuition_all_schools_2016]{Tuition All Schools 2016} \tab  \cr
\code{dasl.TAS2016} \tab \link[=dasl.tuition_all_schools_2016]{Tuition All Schools 2016} \tab  \cr
\code{dasl.twin_births} \tab \link[=dasl.twin_births]{Twin Births} \tab  \cr
\code{dasl.twins_by_year_2014} \tab \link[=dasl.twins_by_year_2014]{Twins by Year 2014} \tab  \cr
\code{dasl.TBY2014} \tab \link[=dasl.twins_by_year_2014]{Twins by Year 2014} \tab  \cr
\code{dasl.usgdp_2016} \tab \link[=dasl.usgdp_2016]{USGDP 2016} \tab  \cr
\code{dasl.unemployment_2017} \tab \link[=dasl.unemployment_2017]{Unemployment 2017} \tab  \cr
\code{dasl.used_bmw_m5_2017} \tab \link[=dasl.used_bmw_m5_2017]{Used BMW M5 2017} \tab  \cr
\code{dasl.UBM2017} \tab \link[=dasl.used_bmw_m5_2017]{Used BMW M5 2017} \tab  \cr
\code{dasl.used_civics_2017} \tab \link[=dasl.used_civics_2017]{Used Civics 2017} \tab  \cr
\code{dasl.UC2017} \tab \link[=dasl.used_civics_2017]{Used Civics 2017} \tab  \cr
\code{dasl.used_cars_2014} \tab \link[=dasl.used_cars_2014]{Used cars 2014} \tab  \cr
\code{dasl.UC2014} \tab \link[=dasl.used_cars_2014]{Used cars 2014} \tab  \cr
\code{dasl.vehicle_weights} \tab \link[=dasl.vehicle_weights]{Vehicle weights} \tab  \cr
\code{dasl.vineyards} \tab \link[=dasl.vineyards]{Vineyards} \tab  \cr
\code{dasl.walmart_revenue} \tab \link[=dasl.walmart_revenue]{Walmart revenue} \tab  \cr
\code{dasl.washing} \tab \link[=dasl.washing]{Washing} \tab  \cr
\code{dasl.watch_the_super_bowl} \tab \link[=dasl.watch_the_super_bowl]{Watch the Super bowl} \tab  \cr
\code{dasl.WTSB} \tab \link[=dasl.watch_the_super_bowl]{Watch the Super bowl} \tab  \cr
\code{dasl.water_hardness} \tab \link[=dasl.water_hardness]{Water hardness} \tab  \cr
\code{dasl.wealth_redistribution} \tab \link[=dasl.wealth_redistribution]{Wealth Redistribution} \tab  \cr
\code{dasl.weather_forecasts} \tab \link[=dasl.weather_forecasts]{Weather forecasts} \tab  \cr
\code{dasl.weightlifting_2016} \tab \link[=dasl.weightlifting_2016]{Weightlifting 2016} \tab  \cr
\code{dasl.wheelchair_marathon_2016} \tab \link[=dasl.wheelchair_marathon_2016]{Wheelchair Marathon 2016} \tab  \cr
\code{dasl.WM2016} \tab \link[=dasl.wheelchair_marathon_2016]{Wheelchair Marathon 2016} \tab  \cr
\code{dasl.whole_foods_2016} \tab \link[=dasl.whole_foods_2016]{Whole Foods 2016} \tab  \cr
\code{dasl.WF2016} \tab \link[=dasl.whole_foods_2016]{Whole Foods 2016} \tab  \cr
\code{dasl.wild_horses} \tab \link[=dasl.wild_horses]{Wild horses} \tab  \cr
\code{dasl.wildfires_2015} \tab \link[=dasl.wildfires_2015]{Wildfires 2015} \tab  \cr
\code{dasl.williams_football} \tab \link[=dasl.williams_football]{Williams football} \tab  \cr
\code{dasl.williams_vs_texas} \tab \link[=dasl.williams_vs_texas]{Williams vs Texas} \tab  \cr
\code{dasl.WVT} \tab \link[=dasl.williams_vs_texas]{Williams vs Texas} \tab  \cr
\code{dasl.wind_power} \tab \link[=dasl.wind_power]{Wind power} \tab  \cr
\code{dasl.wind_speed} \tab \link[=dasl.wind_speed]{Wind speed} \tab  \cr
\code{dasl.wine_production} \tab \link[=dasl.wine_production]{Wine production} \tab  \cr
\code{dasl.winter_olympics_2010_speed_skating} \tab \link[=dasl.winter_olympics_2010_speed_skating]{Winter Olympics 2010 speed skating} \tab  \cr
\code{dasl.WO2010SS} \tab \link[=dasl.winter_olympics_2010_speed_skating]{Winter Olympics 2010 speed skating} \tab  \cr
\code{dasl.winter_olympics_2014} \tab \link[=dasl.winter_olympics_2014]{Winter Olympics 2014} \tab  \cr
\code{dasl.WO2014} \tab \link[=dasl.winter_olympics_2014]{Winter Olympics 2014} \tab  \cr
\code{dasl.winter_olympics_2018_downhill} \tab \link[=dasl.winter_olympics_2018_downhill]{Winter olympics 2018 downhill} \tab  \cr
\code{dasl.WO2018D} \tab \link[=dasl.winter_olympics_2018_downhill]{Winter olympics 2018 downhill} \tab  \cr
\code{dasl.wisconsin_act_2015} \tab \link[=dasl.wisconsin_act_2015]{Wisconsin ACT 2015} \tab  \cr
\code{dasl.WA2015} \tab \link[=dasl.wisconsin_act_2015]{Wisconsin ACT 2015} \tab  \cr
\code{dasl.wisconsin_act_math} \tab \link[=dasl.wisconsin_act_math]{Wisconsin ACT math} \tab  \cr
\code{dasl.WAM} \tab \link[=dasl.wisconsin_act_math]{Wisconsin ACT math} \tab  \cr
\code{dasl.womens_weekly_earnings_2013} \tab \link[=dasl.womens_weekly_earnings_2013]{Women&#8217;s weekly earnings 2013} \tab  \cr
\code{dasl.WWE2013} \tab \link[=dasl.womens_weekly_earnings_2013]{Women&#8217;s weekly earnings 2013} \tab  \cr
\code{dasl.womens_heptathlon_2016} \tab \link[=dasl.womens_heptathlon_2016]{Womens Heptathlon 2016} \tab  \cr
\code{dasl.WH2016} \tab \link[=dasl.womens_heptathlon_2016]{Womens Heptathlon 2016} \tab  \cr
\code{dasl.womens_short_track_2018} \tab \link[=dasl.womens_short_track_2018]{Womens short track 2018} \tab  \cr
\code{dasl.WST2018} \tab \link[=dasl.womens_short_track_2018]{Womens short track 2018} \tab  \cr
\code{dasl.womens_speed_skating_2006} \tab \link[=dasl.womens_speed_skating_2006]{Womens speed skating 2006} \tab  \cr
\code{dasl.WSS2006} \tab \link[=dasl.womens_speed_skating_2006]{Womens speed skating 2006} \tab  \cr
\code{dasl.work_week} \tab \link[=dasl.work_week]{Work week} \tab  \cr
\code{dasl.working_parents} \tab \link[=dasl.working_parents]{Working parents} \tab  \cr
\code{dasl.world_data} \tab \link[=dasl.world_data]{World data} \tab  \cr
\code{dasl.years_to_live_2016} \tab \link[=dasl.years_to_live_2016]{Years to live 2016} \tab  \cr
\code{dasl.YTL2016} \tab \link[=dasl.years_to_live_2016]{Years to live 2016} \tab  \cr
\code{dasl.yeast} \tab \link[=dasl.yeast]{Yeast} \tab  \cr
\code{dasl.yogurt} \tab \link[=dasl.yogurt]{Yogurt} \tab  \cr
\code{dasl.yogurt_flavors} \tab \link[=dasl.yogurt_flavors]{Yogurt flavors} \tab  \cr
\code{dasl.youth_unemployment_2016} \tab \link[=dasl.youth_unemployment_2016]{Youth Unemployment 2016} \tab  \cr
\code{dasl.YU2016} \tab \link[=dasl.youth_unemployment_2016]{Youth Unemployment 2016} \tab  \cr
\code{dasl.zodiac} \tab \link[=dasl.zodiac]{Zodiac} \tab  \cr
\code{dasl.ipod_failures} \tab \link[=dasl.ipod_failures]{iPod failures} \tab  \cr
}
}
