% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.reading}
\alias{dasl.reading}
\title{Reading}
\format{22 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/reading/?sf_paged=33}{Reading}
}
\description{
Reading
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Test for Difference between Two Groups}
