% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.real_estate}
\alias{dasl.real_estate}
\title{Real Estate}
\format{894 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/real-estate/?sf_paged=33}{Real Estate}
}
\description{
As a class project, students in a large Statistics class collected publicly available  information on recent home sales in their hometowns. There are 894 properties. These are not a random sample, but they may be representative of home sales during a short period of time, nationwide. Among the variables available is an indication of whether the home was in an urban, suburban, or rural setting.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Class project
}
\concept{Multiple Regression}
