% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.real_estate_sample_1200}
\alias{dasl.real_estate_sample_1200}
\title{Real estate sample 1200}
\format{1200 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/real-estate-sample-1200/?sf_paged=33}{Real estate sample 1200}
}
\description{
Real estate sample 1200
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
random sample of 1200 houses taken from full Saratoga Housing Data (De Veaux)
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Nonparametric Methods}
