% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.receivers_2015}
\alias{dasl.receivers_2015}
\title{Receivers 2015}
\format{483 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/receivers-2015/?sf_paged=33}{Receivers 2015}
}
\description{
NFL data from the 2015 football season reported the number of yards gained by each of the league’s 488 receivers
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.espn.com/nfl/statistics/player/_/stat/receiving/sort/receivingYards/year/2015/qualified/false}
}
\concept{Normal Probability Plots}
