% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.regular_gas_2017}
\alias{dasl.regular_gas_2017}
\title{Regular gas 2017}
\format{132 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/regular-gas-2017/?sf_paged=33}{Regular gas 2017}
}
\description{
Regular gas 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.bls.gov/cgi-bin/surveymost}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
