% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.retail_trade_index}
\alias{dasl.retail_trade_index}
\title{Retail trade index}
\format{43 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/retail-trade-index/?sf_paged=34}{Retail trade index}
}
\description{
Retail trade index
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
United Nations Statistics Division
}
\concept{Regression}
