% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.rip_currents_2013}
\alias{dasl.rip_currents_2013}
\title{Rip currents 2013}
\format{19 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/rip-currents-2013/?sf_paged=34}{Rip currents 2013}
}
\description{
Rip currents 2013
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.nws.noaa.gov/om/hazstats/resources/weather_fatalities.pdf
}
