% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.rock_concert_deaths}
\alias{dasl.rock_concert_deaths}
\title{Rock concert deaths}
\format{164 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/rock-concert-deaths/?sf_paged=34}{Rock concert deaths}
}
\description{
Crowd Management Strategies monitors accidents at rock concerts. In their database, they list the names and other variables of victims whose deaths were attributed to “crowd crush” at rock concerts. The data give the victims’ ages for data from a one-year period:
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Crowd Management Strategies www.crowdsafe.com
}
\concept{Boxplots}
