% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ruffles}
\alias{dasl.ruffles}
\title{Ruffles}
\format{6 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ruffles/?sf_paged=34}{Ruffles}
}
\description{
Students purchased 6 bags of Lay’s Ruffles potato chips marked with a net weight of 28.3 grams. They carefully weighed the contents of each bag, recording the weights (in grams).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
