% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.rulers}
\alias{dasl.rulers}
\title{Rulers}
\format{24 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/rulers/?sf_paged=34}{Rulers}
}
\description{
A company that traditionally made rulers and yardsticks is setting up a manufacturing process to make metersticks. Obviously, accuracy is important for their new product. The company runs a 24-hour production process. Using a calibration set of data they found: X = 1.000 R = 0.0016 They took a sample of 3 metersticks each hour and recorded the results.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented
}
\concept{Quality Control}
