% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.run_times}
\alias{dasl.run_times}
\title{Run times}
\format{236 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/run-times/?sf_paged=34}{Run times}
}
\description{
Times (in minutes) for one runner to run 4 miles on various courses during a 10-year period.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
DeVeaux personal communication
}
\concept{Data Display}
