% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.running_heats_london}
\alias{dasl.running_heats_london}
\title{Running heats London}
\format{14 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/running-heats-london/?sf_paged=34}{Running heats London}
}
\description{
Running heats London
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Test for Difference between Two Groups}
