% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.sales_and_profits}
\alias{dasl.sales_and_profits}
\title{Sales and profits}
\format{79 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sales-and-profits/?sf_paged=34}{Sales and profits}
}
\description{
Sales and profits
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
A sample from Fortune 500 companies
}
\concept{Regression}
