% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.salmon}
\alias{dasl.salmon}
\title{Salmon}
\format{153 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/salmon/?sf_paged=34}{Salmon}
}
\description{
Salmon
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
EPA (?)
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
