% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.saratoga_house_prices}
\alias{dasl.saratoga_house_prices}
\title{Saratoga house prices}
\format{1063 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/saratoga-house-prices/?sf_paged=34}{Saratoga house prices}
}
\description{
Prices of homes in Saratoga NY along with facts about them. Good basis for multiple regressions to predict the Price of the house. But several predictors are collinear.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
public records
}
\concept{Comparing Two Groups}
\concept{Confidence Intervals for Proportions}
\concept{Regression}
