% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.saratoga_houses}
\alias{dasl.saratoga_houses}
\title{Saratoga houses}
\format{1728 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/saratoga-houses/?sf_paged=34}{Saratoga houses}
}
\description{
Prices of homes in Saratoga NY along with facts about them. Good basis for multiple regressions to predict the Price of the house. But several predictors are collinear.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
public records Saratoga County. Also in R library mosaic
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
