% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.sat_scores}
\alias{dasl.sat_scores}
\title{SAT scores}
\format{162 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sat-scores/?sf_paged=34}{SAT scores}
}
\description{
Scores on SAT tests for 162 students at the same school. (The identity of the school is not provided for privacy.) How are Math and Verbal scores related? Would a regression model be appropriate? Is there a difference in male and female scores? How would that difference be modeled?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Real data—source unspecified to protect privacy
}
\concept{Regression}
