% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.school_system}
\alias{dasl.school_system}
\title{School system}
\format{120 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/school-system/?sf_paged=35}{School system}
}
\description{
A school district superintendent wants to test a new method of teaching arithmetic in the fourth grade at his 15 schools. He plans to select 8 students from each school to take part in the experiment, but to make sure they are roughly of the same ability, he first gives a test to all 120 students. The data hold the scores of the test by school.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.scottishhillracing.co.uk}
}
\concept{Analysis of Variance}
