% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.scottish_hill_races}
\alias{dasl.scottish_hill_races}
\title{Scottish Hill Races}
\format{94 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/scottish-hill-races/?sf_paged=35}{Scottish Hill Races}
}
\description{
Hill races are races that climb generally steep hills, held throughout Scotland throughout the year. The file holds records for men and women in these races the last time those were posted in an accessible table along with facts about the races. In particular, we know the length(km) and total climb(m). These are two independent predictors of the record times. Sex of the runner can be an additional indicator variable.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.scottishhillracing.co.uk}
}
\concept{Multiple Regression}
