% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.sea_ice}
\alias{dasl.sea_ice}
\title{Sea ice}
\format{38 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sea-ice/?sf_paged=35}{Sea ice}
}
\description{
Climate scientists have been observing the extent of sea ice using satellite observations. Many have expressed concern because, since 1980, the extent of sea ice has declined precipitously—possibly due to global climate change. But a multiple regression of Extent on temp and year gives a coefficient for temp that is essentially zero.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Ice extent from \url{http://nsidc.org}  Temperature from \url{http://climate.nasa.gov/system/internal_resources/details/original/647_Global_Temperature_Data_File.txt}
}
\concept{Regression}
