% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.seasonal_spending}
\alias{dasl.seasonal_spending}
\title{Seasonal spending}
\format{99 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/seasonal-spending/?sf_paged=35}{Seasonal spending}
}
\description{
Seasonal spending
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Regression}
