% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.seat_belts_2015}
\alias{dasl.seat_belts_2015}
\title{Seat belts 2015}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/seat-belts-2015/?sf_paged=35}{Seat belts 2015}
}
\description{
The National Highway Traffic Safety Administration reports seat belt use and fatalities in car accidents by state. How do fatalities relate to seat belt use?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
National Highway Traffic Safety Administration. 2016. Seat belt use in 2015 — use rates in the states and territories. Report no. DOT HS-812-274. Washington, DC: U.S. Department of Transportation. found at \url{http://www.iihs.org/iihs/topics/t/general-statistics/fatalityfacts/state-by-state-overview#Rural-versus-urban}
}
