% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.sex_sells}
\alias{dasl.sex_sells}
\title{Sex sells}
\format{39 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sex-sells/?sf_paged=35}{Sex sells}
}
\description{
A group of Statistics students cut ads out of magazines. They were careful to find two ads for each of 10 similar items, one with a sexual image and one without. They arranged the ads in random order and had 39 subjects look at them for one minute. Then they asked the subjects to list as many of the products as they could remember. Their data are shown in the table. Is there evidence that the sexual images mattered?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Real study - unknown source
}
\concept{Blocking}
\concept{Paired Data}
