% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.shirt_sizes}
\alias{dasl.shirt_sizes}
\title{Shirt sizes}
\format{250 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/shirt-sizes/?sf_paged=35}{Shirt sizes}
}
\description{
Shirt sizes
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
BYU Human Performance Research Center \url{http://www.byu.edu/chhp/intro.html#lrc}  Director: Mark Ricard  116A RB, (801) 378-8958
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Multiple Regression}
\concept{Normal Probability Plots}
\concept{Regression}
