% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.shirt_sizes_full}
\alias{dasl.shirt_sizes_full}
\title{Shirt sizes full}
\format{250 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/shirt-sizes-full/?sf_paged=35}{Shirt sizes full}
}
\description{
Shirt sizes full
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Based on data from BYU Human Performance Research Center \url{http://www.byu.edu/chhp/intro.html#lrc}  Director: Mark Ricard  116A RB, (801) 378-8958 Shirt size = Ceiling( (2 * Neck) / 2.54 + 0.5 ) / 2
}
\concept{Data Mining}
