% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.shoe_sizes}
\alias{dasl.shoe_sizes}
\title{Shoe Sizes}
\format{268 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/shoe-sizes/?sf_paged=35}{Shoe Sizes}
}
\description{
A last is a form, traditionally made of wood, in the shape of the human foot. Lasts of various sizes are used by shoemakers to make shoes. In the United States, shoe sizes are defined differently for men and women: U.S. men’s shoe size = (last size in inches * 3) – 24 U.S. women’s shoe size = (last size in inches * 3) – 22.5 But in Europe, they are both: Euro size = last size in cm * 3/2 The data give the European shoe sizes of 269 college students (converted from their reported U.S. shoe sizes.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Real study - unknown source
}
\concept{Confidence Intervals for Means}
\concept{Data Display}
