% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.shoot_to_score_2016}
\alias{dasl.shoot_to_score_2016}
\title{Shoot to Score 2016}
\format{97 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/shoot-to-score-2016/?sf_paged=35}{Shoot to Score 2016}
}
\description{
A college hockey coach collected data from the 2016–2017 National Hockey League season. He hopes to convince his players that the number of shots taken has an effect on the number of goals scored. The data includes both offensive and defensive players.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.nhl.com/stats/player?aggregate=0&gameType=2&report=skatersummary&pos=S&reportType=season&seasonFrom=20162017&seasonTo=20162017&filter=gamesPlayed,gte,1&sort=gamesPlayed}  referenced January 16, 2917
}
\concept{Regression}
