% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.singers_by_parts}
\alias{dasl.singers_by_parts}
\title{Singers by parts}
\format{130 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/singers-by-parts/?sf_paged=35}{Singers by parts}
}
\description{
The dataset gives the heights (in inches) of 130 members of a choir and the part they sing. Note that Sopranos and Altos are typically women and Tenors and Basses are typically men.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Real study - unknown source
}
