% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.ski_wax}
\alias{dasl.ski_wax}
\title{Ski wax}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/ski-wax/?sf_paged=35}{Ski wax}
}
\description{
Bjork Larsen was trying to decide whether to use a new racing wax for cross-country skis. He decided that the wax would be worth the price if he could average less than 55 seconds on a course he knew well, so he planned to study the wax by racing on the course 8 times. The data report his race times.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
De Veaux class project
}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
