% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.slalom_times_2014}
\alias{dasl.slalom_times_2014}
\title{Slalom times 2014}
\format{73 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/slalom-times-2014/?sf_paged=36}{Slalom times 2014}
}
\description{
The Men’s Giant Slalom skiing event consists of two runs whose times are added together for a final score. The data give the giant slalom times in the 2014 Winter Olympics at Sochi.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.sports-reference.com/olympics/winter/2014/ASK/mens-giant-slalom.html}
}
\concept{Boxplots}
