% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.slalom_times_2018}
\alias{dasl.slalom_times_2018}
\title{Slalom times 2018}
\format{75 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/slalom-times-2018/?sf_paged=36}{Slalom times 2018}
}
\description{
The Men’s Giant Slalom skiing event consists of two runs whose times are added together for a final score. The data give the giant slalom times in the 2018 Winter Olympics at PyeongChang.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/Alpine_skiing_at_the_2018_Winter_Olympics_} –_Men's_giant_slalom
}
\concept{Boxplots}
