% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.sleep}
\alias{dasl.sleep}
\title{Sleep}
\format{25 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sleep/?sf_paged=36}{Sleep}
}
\description{
The Sleep Foundation (www.sleepfoundation.org)  says that adults should get at least 7 hours of sleep each night. A survey of students at a small school in the northeast U.S.  asked, among other things, “How much did you sleep last night?” The data are the responses.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
student survey
}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Regression}
