% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.slower_is_cheaper}
\alias{dasl.slower_is_cheaper}
\title{Slower is cheaper}
\format{9 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/slower-is-cheaper/?sf_paged=36}{Slower is cheaper}
}
\description{
Researchers studying how a car’s fuel efficiency (in Miles Per Gallon) varies with its Speed drove a compact car 200 miles at various speeds on a test track. Their data are shown in the table.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
