% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.smoking_2014}
\alias{dasl.smoking_2014}
\title{Smoking 2014}
\format{21 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/smoking-2014/?sf_paged=36}{Smoking 2014}
}
\description{
The Centers for Disease Control and Prevention track cigarette smoking in the United States. How has the percentage of people who smoke changed since the danger became clear during the last half of the 20th century? The data give percentages of smokers among men 18–24 years of age, as estimated by surveys, from 1965 through 2014.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.cdc.gov/nchs}
}
\concept{Regression}
