% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.smoking_and_pregnancy_2011}
\alias{dasl.smoking_and_pregnancy_2011}
\title{Smoking and Pregnancy 2011}
\format{22 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/smoking-and-pregnancy-2011/?sf_paged=36}{Smoking and Pregnancy 2011}
}
\description{
There has been a steady decline the the percentage of pregnant mothers who smoke. These data document the trend. The run only until 2011, which appears to be the latest date for which the CDC has data.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Possibly from the Pregnancy Risk Assessment Monitoring System, www.cdc.gov/prams/
}
