% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.social_networking}
\alias{dasl.social_networking}
\title{Social networking}
\format{6 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/social-networking/?sf_paged=36}{Social networking}
}
\description{
Pew Research conducted a survey about social networking in several countries. They asked whether respondents had access to and used social networking. Responses were “yes” (use social networking), “no”, and “not available”.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Possibly from the Pregnancy Risk Assessment Monitoring System, www.cdc.gov/prams/
}
\concept{Chi Square Tests}
