% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.speed_and_density}
\alias{dasl.speed_and_density}
\title{Speed and density}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sapped-and-density/?sf_paged=36}{Speed and density}
}
\description{
The data report the density (cars per mile) and average speed of traffic on city highways. The data were collected at the same location at 10 different times randomly selected within a span of 3 months.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Regression}
