% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.sprouts}
\alias{dasl.sprouts}
\title{Sprouts}
\format{36 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sprouts/?sf_paged=36}{Sprouts}
}
\description{
An experiment on mung beans was performed to investigate the environmental effects of salinity and water temperature on sprouting. Forty beans were randomly allocated to each of 36 petri dishes that were subject to one of four levels of Salinity (0, 4, 8, and 12 ppm) and one of three Temperatures (32°, 34°, or 36° C). After 48 hours, the biomass of the sprouts in gm was measured. The percent of beans germinating is also recorded.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Student experiment
}
\concept{Multifactor Analysis of Variance}
