% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.stereograms}
\alias{dasl.stereograms}
\title{Stereograms}
\format{78 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/stereograms/?sf_paged=36}{Stereograms}
}
\description{
Stereograms appear to be composed entirely of random dots. However, they contain separate images that a viewer can “fuse” into a three-dimensional (3D) image by staring at the dots while defocusing the eyes. An experiment was performed to determine whether knowledge of the embedded image affected the time required for subjects to fuse the images. One group of subjects (group NV) received no information or just verbal information about the shape of the embedded object. A second group (group VV) received both verbal information and visual information (specifically, a drawing of the object). The experimenters measured how many seconds it took for the subject to report that he or she saw the 3D image.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Real study - unknown source
}
\concept{Boxplots}
\concept{Test for Difference between Two Groups}
