% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.stopping_distance_60}
\alias{dasl.stopping_distance_60}
\title{Stopping distance 60}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/stopping-distance-60/?sf_paged=36}{Stopping distance 60}
}
\description{
A tire manufacturer tested the braking performance of one of its tire models on a test track. The company tried the tires on 10 different cars, recording the stopping distance for each car on both wet and dry pavement from 60 miles per hour. The test was run on both dry and wet pavement. (The actual braking distance takes into account the driver’s reaction time, which typically adds nearly 300 feet at 60 mph!)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Confidence Intervals for Means}
\concept{Paired Data}
