% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.storks}
\alias{dasl.storks}
\title{Storks}
\format{7 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/storks/?sf_paged=37}{Storks}
}
\description{
The number of storks in Oldenburg, Germany, plotted against the population of the town for 7 years in the 1930s. Do storks bring babies?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Statistics for Experimenters, by Box, Hunter and Hunter. Originally from Ornithologische Monatsberichte, 44, no. 2)
}
\concept{Correlation}
\concept{Scatterplot}
