% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.streams}
\alias{dasl.streams}
\title{Streams}
\format{172 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/streams/?sf_paged=37}{Streams}
}
\description{
As part of the course work, a class at an upstate NY college collects data on streams each year. Students record a number of biological, chemical, and physical variables, including the stream name, the substrate of the stream (limestone (L), shale (S), or mixed (M)), the pH, the temperature (C), and the BCI, a measure of biological diversity.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
\concept{Test for Difference between Two Groups}
