% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.student_survey}
\alias{dasl.student_survey}
\title{Student survey}
\format{299 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/student-survey/?sf_paged=37}{Student survey}
}
\description{
The dataset contains data from a class survey
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
De Veaux personal data
}
\concept{Confidence Intervals for Means}
\concept{Data Display}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Tables}
