% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.student_testing}
\alias{dasl.student_testing}
\title{Student testing}
\format{20 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/student-testing/?sf_paged=37}{Student testing}
}
\description{
Researchers randomly assigned subjects to take one of two tests (form A or form B) either electronically or with pencil and paper. Subjects then took the other test using the other method. The two forms had been designed to be equivalent in difficulty, but nevertheless, that equivalence was checked as part of the experiment. Our concern is whether subjects did equally well with each testing method.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented
}
\concept{Nonparametric Methods}
\concept{Quality Control}
