% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.sugar_in_cereal}
\alias{dasl.sugar_in_cereal}
\title{Sugar in cereal}
\format{28 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/sugar-in-cereal/?sf_paged=37}{Sugar in cereal}
}
\description{
Sugar is a major ingredient in many breakfast cereals. The data gives the sugar content as a percentage of weight for 49 brands of cereal. Data were collected from nutrition labels in a supermarket.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Boxplots}
\concept{Test for Difference between Two Groups}
