% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.summer_olympics_2016}
\alias{dasl.summer_olympics_2016}
\title{Summer Olympics 2016}
\format{59 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/summer-olympics-2016/?sf_paged=37}{Summer Olympics 2016}
}
\description{
Fifty nine countries won gold medals in the 2016 Summer Olympics. The dataset lists them, along with the total number of gold medals each won. It can be a challenge to find a good display for data like these.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.google.com/search?client=safari&rls=en&q=olympic+medals+by+country+2016&spell=1&sa=X&ved=0ahUKEwjAwPXqgNvOAhWFHB4KHcxYDF8QvwUIGygA&biw=1370&bih=902}
}
\concept{Data Display}
