% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.summer_school}
\alias{dasl.summer_school}
\title{Summer school}
\format{6 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/summer-school/?sf_paged=37}{Summer school}
}
\description{
Summer school
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Paired Data}
\concept{Test for Difference between Two Groups}
