% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.super_bowl_2016}
\alias{dasl.super_bowl_2016}
\title{Super Bowl 2016}
\format{50 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/super-bowl-2016/?sf_paged=37}{Super Bowl 2016}
}
\description{
Super Bowl 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Wikipedia?
}
\concept{Data Display}
