% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.swim_and_run}
\alias{dasl.swim_and_run}
\title{Swim and Run}
\format{2127 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/swim-and-run/?sf_paged=37}{Swim and Run}
}
\description{
Swim and Run
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Personal Data
}
