% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.swim_the_lake_2016}
\alias{dasl.swim_the_lake_2016}
\title{Swim the lake 2016}
\format{62 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/swim-the-lake-2016/?sf_paged=37}{Swim the lake 2016}
}
\description{
People swim across Lake Ontario from Niagara on the Lake to Toronto–a distance of 52 km (32,3 miles). Because the lake is fresh water, this swim is considered more difficult than ocean swims of similar length because salt water provides more boyancy than fresh water. (For comparison, the English Channel is 21 miles across and, despite strong currents, generally takes less time to cross.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.soloswims.com/swims.htm}
}
\concept{Regression}
\concept{Test for Difference between Two Groups}
