% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.swimming_heats_2016}
\alias{dasl.swimming_heats_2016}
\title{Swimming heats 2016}
\format{15 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/swimming-heats-2016/?sf_paged=37}{Swimming heats 2016}
}
\description{
Unlike track events, swimming heats are not determined at random. Instead, swimmers are seeded so that better swimmers are placed in later heats. Here are the times (in seconds) for the women’s 400-m freestyle for two heats in the 2016 Olympics.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.olympic.org/rio-2016/swimming/200m-freestyle-women}
}
\concept{Comparing Two Groups}
