% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.swimming_heats_london}
\alias{dasl.swimming_heats_london}
\title{Swimming heats London}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/swimming-heats-london/?sf_paged=38}{Swimming heats London}
}
\description{
Swimming heats London
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Test for Difference between Two Groups}
