% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tablet_computers_2014}
\alias{dasl.tablet_computers_2014}
\title{Tablet computers 2014}
\format{44 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tablet-computers-2014/?sf_paged=38}{Tablet computers 2014}
}
\description{
Cnet.com tests tablet computers and continuously updates its list. As of January 2014, the list included the battery life (in hours) and luminous intensity (i.e., screen brightness, in cd/m^2).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.cnet.com/news/cnet-tablet-battery-life-results/}
}
\concept{Regression}
