% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tableware}
\alias{dasl.tableware}
\title{Tableware}
\format{59 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tableware/?sf_paged=38}{Tableware}
}
\description{
Data modified from original data from Nambe Mills.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Modified from real data obtained from the manufacturer
}
