% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tattoos}
\alias{dasl.tattoos}
\title{Tattoos}
\format{626 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tattoos/?sf_paged=38}{Tattoos}
}
\description{
The data give counts of 626 individuals categorized according to their “tattoo status” and their “hepatitis status.” Is there a relationship?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
University of Texas Southwestern Medical Center
}
\concept{Chi Square Tests}
\concept{Data Displays}
\concept{Tables}
