% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tbill_rates_2016}
\alias{dasl.tbill_rates_2016}
\title{TBill rates 2016}
\format{82 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tbill-rates-2016/?sf_paged=38}{TBill rates 2016}
}
\description{
The data give the federal rate on 3-month Treasury bills from 1950 to 1980 and Years Since 1950.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://fred.stlouisfed.org/series/TB3MS/downloaddata}
}
\concept{Regression}
