% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tellers}
\alias{dasl.tellers}
\title{Tellers}
\format{140 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tellers/?sf_paged=38}{Tellers}
}
\description{
A bank is studying the time that it takes 6 of its tellers to serve an average customer. Customers line up in the queue and then go to the next available teller. Is there a difference? Can we pick out the best or worst performing teller?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented example
}
\concept{Analysis of Variance}
