% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tendon_transfers}
\alias{dasl.tendon_transfers}
\title{Tendon transfers}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tendon-transfers/?sf_paged=38}{Tendon transfers}
}
\description{
People with spinal cord injuries may lose function in some, but not all, of their muscles. The ability to push oneself up is particularly important for shifting position when seated and for transferring into and out of wheelchairs. Surgeons compared two operations to restore the ability to push up in children.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
The Journal of Hand Surgery / Vol. 28A No. 6 November 2003 "Prospective Evaluation of Biceps to Triceps and Deltoid to Triceps for Elbow Extension in Tetraplegia,"Mary Jane Mulcahey, MS, Cheryl Lutz, MS, Scott H. Kozin, MD, Randal R. Betz, MD, Philadelphia, PA
}
\concept{Boxplots}
