% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tennis_balls}
\alias{dasl.tennis_balls}
\title{Tennis balls}
\format{18 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tennis-balls/?sf_paged=38}{Tennis balls}
}
\description{
Tennis balls
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Student experiment from De Veaux
}
\concept{Multifactor Analysis of Variance}
