% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.thrills_2013}
\alias{dasl.thrills_2013}
\title{Thrills 2013}
\format{14 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/thrills-2013/?sf_paged=38}{Thrills 2013}
}
\description{
Since 1994, the Best Roller Coaster Poll (www. ushsho.com/bestrollercoasterpoll.htm) has been ranking the world’s best roller coasters. In 2013, Bizarro dropped to 4th after earning the top steel coaster rank for six straight years. Data on the top 14 steel coasters from this poll are given.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.ushsho.com/bestrollercoasterpoll.htm;}   Wikipedia; rcdb.com
}
\concept{Correlation}
\concept{Scatterplot}
