% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tiffany_2017}
\alias{dasl.tiffany_2017}
\title{Tiffany 2017}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tiffany-2017/?sf_paged=38}{Tiffany 2017}
}
\description{
Tiffany was founded in 1837, when Charles Lewis Tiffany opened his first store in downtown Manhattan. Tiffany retails and distributes a selection of Tiffany & Co. brand jewelry at a range of prices. Today, more than 150 Tiffany & Co. stores sell to customers in U.S. and international markets. The dataset holds quarterly sales data from 2005 through the middle of 2017. The data are suitable for time series modeling.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://investor.tiffany.com/results.cfm}
}
\concept{Time Series Analysis}
