% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.time_on_market}
\alias{dasl.time_on_market}
\title{Time on market}
\format{1115 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/time-on-market/?sf_paged=38}{Time on market}
}
\description{
Time on market
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented example from Saratoga data
}
\concept{Multiple Regression}
