% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.titanic}
\alias{dasl.titanic}
\title{Titanic}
\format{2208 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/titanic/?sf_paged=38}{Titanic}
}
\description{
A complete and corrected dataset listing all those aboard HMS Titanic and including related facts about them including Age, how much they paid for their ticket, which boat they survived in (if they survived), and their job if they were crew members. Stories, biographies and pictures can be found on the site: www.encyclopedia-titanica.org/.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.encyclopedia-titanica.org}
}
\concept{Chi Square Tests}
\concept{Data Display}
\concept{Data Displays}
\concept{Tables}
