% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tobacco_and_alcohol}
\alias{dasl.tobacco_and_alcohol}
\title{Tobacco and alcohol}
\format{11 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tobacco-and-alcohol/?sf_paged=39}{Tobacco and alcohol}
}
\description{
Are people who use tobacco products more likely to consume alcohol? Here are data on household spending (in pounds) taken by the British government on 11 regions in Great Britain. Do tobacco and alcohol spending appear to be related? What questions do you have about these data? What conclusions can you draw?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
