% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tomtoms}
\alias{dasl.tomtoms}
\title{Tomtoms}
\format{4000 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tomtoms/?sf_paged=39}{Tomtoms}
}
\description{
Tomtoms
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
invented
}
\concept{Analysis of Variance}
