% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tornadoes_2015}
\alias{dasl.tornadoes_2015}
\title{Tornadoes 2015}
\format{21 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tornadoes-2015/?sf_paged=39}{Tornadoes 2015}
}
\description{
Tornadoes 2015
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.nws.noaa.gov/om/hazstats/resources/weather_fatalities.pdf
}
\concept{Data Display}
