% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tour_de_france_2016}
\alias{dasl.tour_de_france_2016}
\title{Tour de France 2016}
\format{102 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tour-de-france-2016/?sf_paged=39}{Tour de France 2016}
}
\description{
Tour de France 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://bikeraceinfo.com/tdf/tdfstats.html}
}
\concept{Regression}
