% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tour_de_france_2017}
\alias{dasl.tour_de_france_2017}
\title{Tour de France 2017}
\format{104 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tour-de-france-2017/?sf_paged=39}{Tour de France 2017}
}
\description{
The Tour de France is the most famous bicycle race in the world. It has been run every year since 1903, except for a few during wars. The data report facts about the winners including age, time, distance, and average speed. Lance Armstrong’s 7 consecutive victories been disqualified due to the use of performance-enhancing drugs, but his statistics are still included here.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://bikeraceinfo.com/tdf/tdfstats.html}
}
\concept{Regression}
