% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.toy_makers}
\alias{dasl.toy_makers}
\title{Toy makers}
\format{11 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/toy-makers/?sf_paged=39}{Toy makers}
}
\description{
Toy makers
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.toyassociation.org/ta/research/data/annual/toys/research-and-data/data/annual-us-sales-data.aspx}
}
\concept{Data Displays}
