% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.toyota_stock_prices_2013}
\alias{dasl.toyota_stock_prices_2013}
\title{Toyota stock prices 2013}
\format{5081 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/toyota-stock-prices-2013/?sf_paged=39}{Toyota stock prices 2013}
}
\description{
Daily closing stock prices for Toyota Motor Manufacturing from April 1, 2008, through June 21, 2013
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Yahoo! Finance
}
\concept{Time Series Analysis}
