% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tracking_hurricanes_2015}
\alias{dasl.tracking_hurricanes_2015}
\title{Tracking hurricanes 2015}
\format{46 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tracking-hurricanes-2015/?sf_paged=39}{Tracking hurricanes 2015}
}
\description{
Tracking hurricanes 2015
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
1970-present_OFCL_ATL_annual_trk_errors_noTDs.pdf
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
