% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tracking_hurricanes_2016}
\alias{dasl.tracking_hurricanes_2016}
\title{Tracking hurricanes 2016}
\format{47 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tracking-hurricanes-2016/?sf_paged=39}{Tracking hurricanes 2016}
}
\description{
The National Hurricane Center (NHC) of the National Oceanic and Atmospheric Administration (NOAA) tries to predict the path each hurricane will take. But hurricanes tend to wander around aimlessly and are pushed by fronts and other weather phenomena in their area, so they are notoriously difficult to predict. Even relatively small changes in a hurricane’s track can make big differences in the damage it causes. The data give the mean error in nautical miles of the NHC’s 72-hour predictions of Atlantic hurricanes for 1970-2017. NOAA refers to these errors as the Forecast error or the Prediction error and reports annual results.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
1970-present_OFCL_ATL_annual_trk_errors_noTDs.pdf
}
\concept{Regression}
\concept{Scatterplot}
