% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.traffic_fatalities_2013}
\alias{dasl.traffic_fatalities_2013}
\title{Traffic fatalities 2013}
\format{27 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/traffic-fatalities-2013/?sf_paged=39}{Traffic fatalities 2013}
}
\description{
Traffic fatalities in a variety of vehicles and for a variety of situations for the years from 1975 to 2013. These are multiple time series, but can also be related to each other.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
NHTSA
}
\concept{Data Display}
