% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.transportation_energy_use}
\alias{dasl.transportation_energy_use}
\title{Transportation Energy use}
\format{51 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/transportation-energy-use/?sf_paged=39}{Transportation Energy use}
}
\description{
The U.S. Energy Information Administration (EIA) collects data on the total energy used per capita in transportation for each state and the District of Columbia. The data show the per capita consumption in the year 2015 in millions of BTU per person.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.eia.gov/state/seds/data.php?incfile=/state/seds/sep_sum/html/rank_use_capita.html&sid=US}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
