% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.trophy_sizes}
\alias{dasl.trophy_sizes}
\title{Trophy sizes}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/trophy-sizes/?sf_paged=40}{Trophy sizes}
}
\description{
Trophy sizes
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Nonparametric Methods}
