% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.trucks}
\alias{dasl.trucks}
\title{Trucks}
\format{228 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/trucks/?sf_paged=40}{Trucks}
}
\description{
U.S. Department of Transportation reports records of border crossings into each state on the U.S. border. Here are the border crossings by trucks for Alaska, recorded each month from 1999 through 2017.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.transportation.gov/Research-and-Statistics/Border-Crossing-Entry-Data/keg4-3bc2/data}
}
\concept{Time Series Analysis}
