% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tsunamis_2016}
\alias{dasl.tsunamis_2016}
\title{Tsunamis 2016}
\format{1087 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tsunamis-2016/?sf_paged=40}{Tsunamis 2016}
}
\description{
Tsunamis 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.ngdc.noaa.gov/hazard/tsu_db.shtml}
}
