% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tsunamis_2018}
\alias{dasl.tsunamis_2018}
\title{Tsunamis 2018}
\format{1116 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tsunamis-2018/?sf_paged=40}{Tsunamis 2018}
}
\description{
Tsunamis 2018
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.ngdc.noaa.gov/hazard/tsu_db.shtml}  Extracted Event Validity 3 and 4 Cause Codes 1-5 Event Validity: 4 = definite tsunami 3 = probable tsunami 2 = questionable tsunami 1 = very doubtful tsunami 0 = event that only caused a seiche or disturbance in an inland river -1 = erroneous entry Cause Code: Valid values: 0 to 11 The source of the tsunami: 0 = Unknown 1 = Earthquake 2 = Questionable Earthquake 3 = Earthquake and Landslide 4 = Volcano and Earthquake 5 = Volcano, Earthquake, and Landslide 6 = Volcano 7 = Volcano and Landslide 8 = Landslide 9 = Meteorological 10 = Explosion 11 = Astronomical Tide
}
\concept{Data Display}
