% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tuition_2016}
\alias{dasl.tuition_2016}
\title{Tuition 2016}
\format{19 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tuition-2016/?sf_paged=40}{Tuition 2016}
}
\description{
Tuition 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.collegeboard.org
}
\concept{Blocking}
\concept{Paired Data}
