% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tuition_all_schools_2016}
\alias{dasl.tuition_all_schools_2016}
\title{Tuition All Schools 2016}
\format{7703 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tuition-all-schools-2016/?sf_paged=40}{Tuition All Schools 2016}
}
\description{
Tuition All Schools 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://collegescorecard.ed.gov/data/}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
