% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.tv_watching}
\alias{dasl.tv_watching}
\title{TV watching}
\format{197 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/tv-watching/?sf_paged=40}{TV watching}
}
\description{
A study at a liberal arts college attempted to find out whether men and women watch the same amount of TV, on average and whether it mattered if students were varsity athletes or not. Student researchers asked 200 randomly selected students questions about their backgrounds and about their television-viewing habits and received 197 legitimate responses. The researchers found that men watch, on average, about 2.5 hours per week more TV than women, and that varsity athletes watch about 3.5 hours per week more than those who are not varsity athletes. But is this the whole story? To investigate further, they divided the students into four groups: male athletes (MA), male non-athletes (MNA), female athletes (FA), and female non-athletes (FNA).
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Analysis of Variance}
\concept{Multifactor Analysis of Variance}
