% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.twin_births}
\alias{dasl.twin_births}
\title{Twin Births}
\format{278 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/twin-births/?sf_paged=40}{Twin Births}
}
\description{
Twin Births
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
JAMA 284 [2000]:335–341)
}
\concept{Tables}
