% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.twins_by_year_2014}
\alias{dasl.twins_by_year_2014}
\title{Twins by Year 2014}
\format{35 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/twins-by-year-2014/?sf_paged=40}{Twins by Year 2014}
}
\description{
In January 2012, the New York Times published a story called “Twin Births in the U.S., Like Never Before,” in which they reported a 76 percent increase in the rate of twin births from 1980 to 2009. The dataset gives the number of twin births each year (per 1000 live births). Can you confirm the Times report? The dataset also includes the atmospheric CO2 levels (ppm) for those years to offer an alternative predictor in case there appears to be an argument for causation.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.cdc.gov/nchs/births.htm
}
