% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.unemployment_2017}
\alias{dasl.unemployment_2017}
\title{Unemployment 2017}
\format{176 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/unemployment-2017/?sf_paged=40}{Unemployment 2017}
}
\description{
US Unemployment rate from 1/1/2003 to 8/1/17.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.bls.gov/timeseries/LNS14000000}
}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
