% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.used_bmw_m5_2017}
\alias{dasl.used_bmw_m5_2017}
\title{Used BMW M5 2017}
\format{36 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/used-bmw-m5-2017/?sf_paged=40}{Used BMW M5 2017}
}
\description{
Used BMW M5 2017
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Kelly's Blue Book: \url{https://www.kbb.com/cars-for-sale/}  accessed on 31 Aug 2017 using zip code 94305 200 mile radius BMW M5
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
