% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.used_cars_2014}
\alias{dasl.used_cars_2014}
\title{Used cars 2014}
\format{13 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/used-cars-2014/?sf_paged=41}{Used cars 2014}
}
\description{
How does the age of a used car influence its price? This is a small enough data set to find a model with a calculator.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Carmax.com search for Toyota Corollas for sale within 250 radius of Redlands, CA in 2014
}
\concept{Regression}
