% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.used_civics_2017}
\alias{dasl.used_civics_2017}
\title{Used Civics 2017}
\format{55 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/used-civics-2017/?sf_paged=41}{Used Civics 2017}
}
\description{
The web site www.autotrader.com lists cars for sale. On January 22 2017, it listed 55 used Honda Civics for sale by owner. From those listings, we extracted the asking price ($), the mileage, and the model year (from which we computed the age of the car at the time the data were collected Questions include how to best predict the price from mileage and age and whether any of the cars is a particularly good buy. One care is a particularly old (1989) car that has relatively low mileage for such an old car. The seller claims it hasn’t been driven for several years.  It looks like Price might benefit from re-expression by logs.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.autotrader.com accessed on 22 Jan 2017 with a search for Honda Civics for sale by owner.
}
\concept{Correlation}
\concept{Regression}
\concept{Scatterplot}
