% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.usgdp_2016}
\alias{dasl.usgdp_2016}
\title{USGDP 2016}
\format{67 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/usgdp-2016/?sf_paged=41}{USGDP 2016}
}
\description{
The data give the gross domestic product (GDP) of the United States in trillions of 2009 dollars and time.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.multpl.com/us-gdp-inflation-adjusted/table}
}
\concept{Regression}
