% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.vehicle_weights}
\alias{dasl.vehicle_weights}
\title{Vehicle weights}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/vehicle-weights/?sf_paged=41}{Vehicle weights}
}
\description{
The Minnesota Department of Transportation hoped that they could measure the weights of big trucks without actually stopping the vehicles by using a newly developed “weight-in-motion” scale. To see if the new device was accurate, they conducted a calibration test. They weighed several stopped trucks (Static Weight) and assumed that this weight was correct. Then they weighed the trucks again while they were moving to see how well the new scale could estimate the actual weight.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
The Minnesota Department of Transportation
}
\concept{Blocking}
\concept{Correlation}
\concept{Paired Data}
\concept{Scatterplot}
