% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.vineyards}
\alias{dasl.vineyards}
\title{Vineyards}
\format{36 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/vineyards/?sf_paged=41}{Vineyards}
}
\description{
Vineyards
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
local information
}
\concept{Boxplots}
\concept{Displaying Quantitative Data}
\concept{Summarizing Quantitative Data}
