% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.walmart_revenue}
\alias{dasl.walmart_revenue}
\title{Walmart revenue}
\format{39 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/walmart-revenue/?sf_paged=41}{Walmart revenue}
}
\description{
Walmart revenue
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
various
}
\concept{Multiple Regression}
\concept{Time Series Analysis}
