% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.washing}
\alias{dasl.washing}
\title{Washing}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/washing/?sf_paged=41}{Washing}
}
\description{
Washing
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Fabricated
}
\concept{Multifactor Analysis of Variance}
