% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.watch_the_super_bowl}
\alias{dasl.watch_the_super_bowl}
\title{Watch the Super bowl}
\format{1008 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/watch-the-super-bowl/?sf_paged=41}{Watch the Super bowl}
}
\description{
The Gallup poll asked 1008 Americans age 18 and over whether they planned to watch the upcoming Super Bowl. The pollster also asked those who planned to watch whether they were looking forward more to seeing the football game or the commercials.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
gallup.com
}
\concept{Tables}
