% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.water_hardness}
\alias{dasl.water_hardness}
\title{Water hardness}
\format{61 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/water-hardness/?sf_paged=41}{Water hardness}
}
\description{
Water hardness
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Regression}
