% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wealth_redistribution}
\alias{dasl.wealth_redistribution}
\title{Wealth Redistribution}
\format{3045 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wealth-redistribution/?sf_paged=41}{Wealth Redistribution}
}
\description{
Gallup Poll of 1015 U.S. adults on April 9 – 12, 2015. Respondents were classified as high income (over $75,000), middle income ($30k–$75k), or low income (less than $30k). Those polled were asked for their views on  redistributing U.S. wealth by heavily taxing the rich. Counts are reconstructed from percentages published by Gallup.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.gallup.com/poll/182987/americans-continue-say-wealth-distribution-unfair.aspx}
}
