% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.weather_forecasts}
\alias{dasl.weather_forecasts}
\title{Weather forecasts}
\format{365 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/weather-forecasts/?sf_paged=41}{Weather forecasts}
}
\description{
Weather forecasts
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
realistic data
}
\concept{Tables}
