% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.weightlifting_2016}
\alias{dasl.weightlifting_2016}
\title{Weightlifting 2016}
\format{8 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/weightlifting-2016/?sf_paged=41}{Weightlifting 2016}
}
\description{
The world men’s weightlifting records are categorized by weight class of the competitors. How does the weight class relate to the record?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_Olympic_records_in_weightlifting}
}
\concept{Regression}
