% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wheelchair_marathon_2016}
\alias{dasl.wheelchair_marathon_2016}
\title{Wheelchair Marathon 2016}
\format{40 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wheelchair-marathon-2016/?sf_paged=42}{Wheelchair Marathon 2016}
}
\description{
The Boston Marathon hashad a wheelchair division since 1977. Who do you think is typically faster, the men’s marathon winner on foot or the women’s wheelchair marathon winner? Because the conditions differ from year to year, and speeds have improved over the years, it seems best to treat these as paired measurements. Here are summary statistics for the pairwise differences in finishing time (in minutes):
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/List_of_winners_of_the_Boston_Marathon}
}
\concept{Blocking}
\concept{Paired Data}
