% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.whole_foods_2016}
\alias{dasl.whole_foods_2016}
\title{Whole Foods 2016}
\format{88 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/whole-foods-2016/?sf_paged=42}{Whole Foods 2016}
}
\description{
Quarterly sales of Whole Foods Markets from 1995 through 2016. Whole Foods was purchased by Amazon in 2017, so 2016 is the final complete year prior to the merger. The data show a strong seasonal component even though food sales should not be seasonal.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Whole Foods Markets annual reports
}
\concept{Time Series Analysis}
