% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wild_horses}
\alias{dasl.wild_horses}
\title{Wild horses}
\format{38 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wild-horses/?sf_paged=42}{Wild horses}
}
\description{
Large herds of wild horses can become a problem on some federal lands in the West. Researchers hoping to improve the management of these herds collected data to see if they could predict the number of foals that would be born based on the size of the current herd.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
