% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wildfires_2015}
\alias{dasl.wildfires_2015}
\title{Wildfires 2015}
\format{31 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wildfires-2015/?sf_paged=42}{Wildfires 2015}
}
\description{
The National Interagency Fire Center reports statistics about wildfires. They report data from 1960, but the years 1960-1984 are so different from subsequent years that they can’t be analyzed together. These data are for 1985-2015. Is there a pattern over time? What is the relationship between the number of fires and the acres affected? Are fires getting larger or smaller on average?
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.nifc.gov/fireInfo/fireInfo_stats_totalFires.html}
}
\concept{Regression}
