% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.williams_football}
\alias{dasl.williams_football}
\title{Williams football}
\format{10 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/williams-football/?sf_paged=42}{Williams football}
}
\description{
The Sears Cup was established in 1993 to honor institutions that maintain a broad-based athletic program, achieving success in many sports, both men’s and women’s. In the years following its Division III inception in 1995, the cup was won by Williams College 15 of 17 years. Why did the football team win so much? Was it because they were heavier than their opponents? The data gives the average team weights for selected years from 1973 to 1993.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Taken from Williams College Football programs
}
