% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.williams_vs_texas}
\alias{dasl.williams_vs_texas}
\title{Williams vs Texas}
\format{6 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/williams-vs-texas/?sf_paged=42}{Williams vs Texas}
}
\description{
Williams vs Texas
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Larsen and Marx Mathematical Statistics
}
