% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wind_power}
\alias{dasl.wind_power}
\title{Wind power}
\format{1114 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wind-power/?sf_paged=42}{Wind power}
}
\description{
Should you generate electricity with your own personal wind turbine? That depends on whether you have enough wind on your site. To produce enough energy, your site should have an annual average wind speed above 8 miles per hour, according to the Wind Energy Association. One candidate site was monitored for a year, with wind speeds recorded every 6 hours. A total of 1114 readings of wind speed averaged 8.019 mph with a standard deviation of 3.813 mph. The data are provided.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
extracted from Hopkins Forest data set (\url{http://web.williams.edu/weather/archives.php)}
}
\concept{Confidence Intervals for Means}
