% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wind_speed}
\alias{dasl.wind_speed}
\title{Wind speed}
\format{1114 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wind-speed/?sf_paged=42}{Wind speed}
}
\description{
Wind speed
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
unknown
}
\concept{Blocking}
\concept{Comparing Two Groups}
\concept{Confidence Intervals for Means}
\concept{Paired Data}
\concept{Regression}
