% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wine_production}
\alias{dasl.wine_production}
\title{Wine production}
\format{36 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wine-production/?sf_paged=42}{Wine production}
}
\description{
Wine production
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Collected by Dave Bock
}
\concept{Analysis of Variance}
\concept{Experiment Design}
