% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.winter_olympics_2010_speed_skating}
\alias{dasl.winter_olympics_2010_speed_skating}
\title{Winter Olympics 2010 speed skating}
\format{36 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/winter-olympics-2010-speed-skating/?sf_paged=42}{Winter Olympics 2010 speed skating}
}
\description{
The times from the first race of the women’s 2 X 500-m speed skating times at the 2010 Winter Olympics in Vancouver, B.C. are given.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.olympic.org/vancouver-2010/speed-skating/2x500m-women}
}
