% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.winter_olympics_2014}
\alias{dasl.winter_olympics_2014}
\title{Winter Olympics 2014}
\format{34 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/winter-olympics-2014/?sf_paged=42}{Winter Olympics 2014}
}
\description{
Winter Olympics 2014
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://www.olympic.org/sochi-2014/alpine-skiing/slalom-men}
}
\concept{Boxplots}
\concept{Normal Probability Plots}
