% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.winter_olympics_2018_downhill}
\alias{dasl.winter_olympics_2018_downhill}
\title{Winter olympics 2018 downhill}
\format{53 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/winter-olympics-2018-downhill/?sf_paged=42}{Winter olympics 2018 downhill}
}
\description{
53 men completed the men’s alpine downhill. The gold medal winner finished in 100.25 seconds. Here are the times (in seconds) for all competitors.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/Alpine_skiing_at_the_2018_Winter_Olympics_} –_Men%27s_downhill
}
\concept{Normal Probability Plots}
