% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wisconsin_act_2015}
\alias{dasl.wisconsin_act_2015}
\title{Wisconsin ACT 2015}
\format{458 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wisconsin-act-2015/?sf_paged=43}{Wisconsin ACT 2015}
}
\description{
The data give the mean ACT composite scores for all 450 Wisconsin public schools in 2015 along with the type of school and number of students.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://archive.jsonline.com/watchdog/dataondemand/2015-Wisconsin-ACT-scores-by-district-and-school-323020921.html#!/averagescore.desc.19/}
}
\concept{Normal Probability Plots}
