% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.wisconsin_act_math}
\alias{dasl.wisconsin_act_math}
\title{Wisconsin ACT math}
\format{457 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/wisconsin-act-math/?sf_paged=43}{Wisconsin ACT math}
}
\description{
Wisconsin ACT math
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
made up
}
