% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.womens_heptathlon_2016}
\alias{dasl.womens_heptathlon_2016}
\title{Womens Heptathlon 2016}
\format{31 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/womens-heptathlon-2016/?sf_paged=43}{Womens Heptathlon 2016}
}
\description{
The women’s heptathlon in the Olympics consists of seven track-and-field events: the 200 m and 800 m runs, 100 m high hurdles, shot put, javelin, high jump, and long jump. Each contestant is awarded points for each event based on her performance. So, which performance deserves more points? It’s not clear how to compare them. They aren’t measured in the same units, or even in the same direction (longer jumps are better but shorter times are better.)
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://results.nbcolympics.com/track-field/event/women-heptathlon/phase=atw700c00/index.html}
}
\concept{Normal Probability Plots}
\concept{Regression}
