% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.womens_short_track_2018}
\alias{dasl.womens_short_track_2018}
\title{Womens short track 2018}
\format{29 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/womens-short-track-2018/?sf_paged=43}{Womens short track 2018}
}
\description{
The Women’s 500 metres in short track speed skating at the 2018 Winter Olympics took place from 10 to 13 February 2018 at the Gangneung Ice Arena in Gangneung, South Korea.The defending champion from 2014, Li Jianrou, had retired, but the 2014 silver medalist Arianna Fontana competed and eventually won the event. Elise Christie and Choi Min-jeong, who set two Olympic records each in heats and quarterfinals, were considered heavy favorites. They both qualified to the final, but Christie crashed out and finished fourth, and Choi was disqualified for interference after initially finishing second.[3] Fontana became the Olympic champion, with Yara van Kerkhof, who was advanced to the final, winning the silver, and Kim Boutin bronze.[4] Fontana’s gold medal was the first gold medal in women’s short track speed skating won by a European athlete since 1988 (when Monique Velzeboer and the Italian relay team took gold). The data give the times of the qualifying heats.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://en.wikipedia.org/wiki/Short_track_speed_skating_at_the_2018_Winter_Olympics_} –_Women's_500_metres
}
