% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.womens_speed_skating_2006}
\alias{dasl.womens_speed_skating_2006}
\title{Womens speed skating 2006}
\format{35 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/womens-speed-skating-2006/?sf_paged=43}{Womens speed skating 2006}
}
\description{
The women’s 1500 metres speed skating competition for the 2006 Winter Olympics was held in Turin, Italy, on 22 February
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://library.la84.org/6oic/OfficialReports/2006/Results/SpeedSkating.pdf}
}
\concept{Blocking}
\concept{Paired Data}
