% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.womens_weekly_earnings_2013}
\alias{dasl.womens_weekly_earnings_2013}
\title{Women&#8217;s weekly earnings 2013}
\format{41 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/womens-weekly-earnings-2013/?sf_paged=43}{Women&#8217;s weekly earnings 2013}
}
\description{
Quarterly median weekly earnings for U.S. women 25 years of age or older. Data are provided from the first quarter of 2003 through the first quarter of 2013.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
www.bls.gov U.S. Bureau of Labor Statistics
}
\concept{Time Series Analysis}
