% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.work_week}
\alias{dasl.work_week}
\title{Work week}
\format{11 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/work-week/?sf_paged=43}{Work week}
}
\description{
Do flexible work schedules reduce the demand for resources? The Lake County, Illinois, Health Department experimented with a flexible four-day workweek. For a year, the department recorded the mileage driven by 11 field workers on an ordinary five-day workweek. Then it changed to a flexible four-day workweek and recorded mileage for another year.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Charles S. Catlin, “Four-day Work Week Improves Environment,” Journal of Environmental Health, Denver, 59:7, 1997.
}
\concept{Blocking}
\concept{Nonparametric Methods}
\concept{Paired Data}
