% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.working_parents}
\alias{dasl.working_parents}
\title{Working parents}
\format{5 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/working-parents/?sf_paged=43}{Working parents}
}
\description{
Working parents
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
