% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.world_data}
\alias{dasl.world_data}
\title{World data}
\format{240 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/world-data/?sf_paged=43}{World data}
}
\description{
World data
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
The world factbook \url{https://www.cia.gov/library/publications/the-world-factbook/}
}
\concept{Multiple Regression}
