% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.years_to_live_2016}
\alias{dasl.years_to_live_2016}
\title{Years to live 2016}
\format{22 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/years-to-live-2016/?sf_paged=43}{Years to live 2016}
}
\description{
Insurance companies and other organizations use actuarial tables to estimate the remaining lifespans of their customers. The data file gives estimated life expectancy and additional years of life for black males in the United States, according to a 2016 National Vital Statistics Report, A regression model to predict Life expectancy from Age appears to fit well, but consider the residuals.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{http://www.cdc.gov/nchs/data/nvsr/nvsr65/nvsr65_04.pdf}
}
\concept{Regression}
