% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.yeast}
\alias{dasl.yeast}
\title{Yeast}
\format{16 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/yeast/?sf_paged=43}{Yeast}
}
\description{
Yeast
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
From student experiment (De Veaux)
}
\concept{Analysis of Variance}
