% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.yogurt}
\alias{dasl.yogurt}
\title{Yogurt}
\format{11 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/yogurt/?sf_paged=43}{Yogurt}
}
\description{
Consumer Reports tested 11 brands of vanilla yogurt and found these numbers of calories per serving.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Consumer Reports Volume unknown
}
\concept{Blocking}
\concept{Confidence Intervals for Means}
\concept{Hypothesis Tests}
\concept{Hypothesis Tests for Means}
\concept{Paired Data}
