% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.yogurt_flavors}
\alias{dasl.yogurt_flavors}
\title{Yogurt flavors}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/yogurt-flavors/?sf_paged=44}{Yogurt flavors}
}
\description{
Yogurt flavors
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Consumer Reports Volume unknown
}
\concept{Blocking}
\concept{Paired Data}
