% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.youth_unemployment_2016}
\alias{dasl.youth_unemployment_2016}
\title{Youth Unemployment 2016}
\format{33 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/youth-unemployment-2016/?sf_paged=44}{Youth Unemployment 2016}
}
\description{
Youth Unemployment 2016
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
\url{https://data.oecd.org/unemp/youth-unemployment-rate.htm#indicator-chart}
}
\concept{Regression}
