% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasl.R
\docType{data}
\name{dasl.zodiac}
\alias{dasl.zodiac}
\title{Zodiac}
\format{12 observations}
\source{
DASL – The Data And Story Library: \href{https://dasl.datadescription.com/datafile/zodiac/?sf_paged=44}{Zodiac}
}
\description{
Fortune magazine collected the zodiac signs of 256 heads of the largest 400 companies. The data shows the number of births for each sign.
}
\details{
\url{https://github.com/sigbertklinke/wwwdata/tree/master/wwwdata/dasl}
}
\references{
Unknown
}
\concept{Chi Square Tests}
