% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{data}
\usage{
data(..., list = character(), package = NULL, lib.loc = NULL,
  verbose = getOption("verbose"), envir = .GlobalEnv,
  overwrite = TRUE, wwwdatalist = getOption("wwwdata"))
}
\arguments{
\item{...}{literal character strings or names.}

\item{list}{a character vector.}

\item{package}{
    a character vector giving the package(s) to look
    in for data sets, or \code{NULL}.

    By default, all packages in the search path are used, then
    the \file{data} subdirectory (if present) of the current working
    directory.
  }

\item{lib.loc}{a character vector of directory names of \R libraries,
    or \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.}

\item{verbose}{a logical.  If \code{TRUE}, additional diagnostics are
    printed.}

\item{envir}{the \link{environment} where the data should be loaded.}

\item{overwrite}{logical: should existing objects of the same name in
    \env{envir} be replaced?}

\item{wwwdatalist}{list of data sets and URLs (default: \code{getOption('wwwdata')})}
}
\value{
a data set
}
\description{
Loads a specified a data set either locally of from the internet (GitHub).
If the data set name is not in the list of internet data set names then the standard 
\code{\link[utils]{data}} will be called instead.
}
\details{
If the data is loaded from internet then it will be locally stored in \code{~/.mmstat}. 
Later calls will first check that directory before trying to load from internet.
}
\note{
If loading a data set from internet fails then the result is undefined. 
You may get an error, a warning and/or a data set which contains garbage.
}
\examples{
\dontrun{
data(ansur2.male)
str(ansur2.male)
}
}
